--- python/hidet/ir/hidet_parser.py
@@ -1,3249 +1,0 @@
-# The file was automatically generated by Lark v1.1.7
-__version__ = "1.1.7"
-
-#
-#
-#   Lark Stand-alone Generator Tool
-# ----------------------------------
-# Generates a stand-alone LALR(1) parser
-#
-# Git:    https://github.com/erezsh/lark
-# Author: Erez Shinan (erezshin@gmail.com)
-#
-#
-#    >>> LICENSE
-#
-#    This tool and its generated code use a separate license from Lark,
-#    and are subject to the terms of the Mozilla Public License, v. 2.0.
-#    If a copy of the MPL was not distributed with this
-#    file, You can obtain one at https://mozilla.org/MPL/2.0/.
-#
-#    If you wish to purchase a commercial license for this tool and its
-#    generated code, you may contact me via email or otherwise.
-#
-#    If MPL2 is incompatible with your free or open-source project,
-#    contact me and we'll work it out.
-#
-#
-
-from abc import ABC, abstractmethod
-from collections.abc import Sequence
-from types import ModuleType
-from typing import (
-    TypeVar, Generic, Type, Tuple, List, Dict, Iterator, Collection, Callable, Optional, FrozenSet, Any,
-    Union, Iterable, IO, TYPE_CHECKING, overload,
-    Pattern as REPattern, ClassVar, Set, Mapping
-)
-
-
-class LarkError(Exception):
-    pass
-
-
-class ConfigurationError(LarkError, ValueError):
-    pass
-
-
-def assert_config(value, options: Collection, msg='Got %r, expected one of %s'):
-    if value not in options:
-        raise ConfigurationError(msg % (value, options))
-
-
-class GrammarError(LarkError):
-    pass
-
-
-class ParseError(LarkError):
-    pass
-
-
-class LexError(LarkError):
-    pass
-
-T = TypeVar('T')
-
-class UnexpectedInput(LarkError):
-    #--
-    line: int
-    column: int
-    pos_in_stream = None
-    state: Any
-    _terminals_by_name = None
-
-    def get_context(self, text: str, span: int=40) -> str:
-        #--
-        assert self.pos_in_stream is not None, self
-        pos = self.pos_in_stream
-        start = max(pos - span, 0)
-        end = pos + span
-        if not isinstance(text, bytes):
-            before = text[start:pos].rsplit('\n', 1)[-1]
-            after = text[pos:end].split('\n', 1)[0]
-            return before + after + '\n' + ' ' * len(before.expandtabs()) + '^\n'
-        else:
-            before = text[start:pos].rsplit(b'\n', 1)[-1]
-            after = text[pos:end].split(b'\n', 1)[0]
-            return (before + after + b'\n' + b' ' * len(before.expandtabs()) + b'^\n').decode("ascii", "backslashreplace")
-
-    def match_examples(self, parse_fn: 'Callable[[str], Tree]',
-                             examples: Union[Mapping[T, Iterable[str]], Iterable[Tuple[T, Iterable[str]]]],
-                             token_type_match_fallback: bool=False,
-                             use_accepts: bool=True
-                         ) -> Optional[T]:
-        #--
-        assert self.state is not None, "Not supported for this exception"
-
-        if isinstance(examples, Mapping):
-            examples = examples.items()
-
-        candidate = (None, False)
-        for i, (label, example) in enumerate(examples):
-            assert not isinstance(example, str), "Expecting a list"
-
-            for j, malformed in enumerate(example):
-                try:
-                    parse_fn(malformed)
-                except UnexpectedInput as ut:
-                    if ut.state == self.state:
-                        if (
-                            use_accepts
-                            and isinstance(self, UnexpectedToken)
-                            and isinstance(ut, UnexpectedToken)
-                            and ut.accepts != self.accepts
-                        ):
-                            logger.debug("Different accepts with same state[%d]: %s != %s at example [%s][%s]" %
-                                         (self.state, self.accepts, ut.accepts, i, j))
-                            continue
-                        if (
-                            isinstance(self, (UnexpectedToken, UnexpectedEOF))
-                            and isinstance(ut, (UnexpectedToken, UnexpectedEOF))
-                        ):
-                            if ut.token == self.token:  ##
-
-                                logger.debug("Exact Match at example [%s][%s]" % (i, j))
-                                return label
-
-                            if token_type_match_fallback:
-                                ##
-
-                                if (ut.token.type == self.token.type) and not candidate[-1]:
-                                    logger.debug("Token Type Fallback at example [%s][%s]" % (i, j))
-                                    candidate = label, True
-
-                        if candidate[0] is None:
-                            logger.debug("Same State match at example [%s][%s]" % (i, j))
-                            candidate = label, False
-
-        return candidate[0]
-
-    def _format_expected(self, expected):
-        if self._terminals_by_name:
-            d = self._terminals_by_name
-            expected = [d[t_name].user_repr() if t_name in d else t_name for t_name in expected]
-        return "Expected one of: \n\t* %s\n" % '\n\t* '.join(expected)
-
-
-class UnexpectedEOF(ParseError, UnexpectedInput):
-    #--
-    expected: 'List[Token]'
-
-    def __init__(self, expected, state=None, terminals_by_name=None):
-        super(UnexpectedEOF, self).__init__()
-
-        self.expected = expected
-        self.state = state
-        from .lexer import Token
-        self.token = Token("<EOF>", "")  ##
-
-        self.pos_in_stream = -1
-        self.line = -1
-        self.column = -1
-        self._terminals_by_name = terminals_by_name
-
-
-    def __str__(self):
-        message = "Unexpected end-of-input. "
-        message += self._format_expected(self.expected)
-        return message
-
-
-class UnexpectedCharacters(LexError, UnexpectedInput):
-    #--
-
-    allowed: Set[str]
-    considered_tokens: Set[Any]
-
-    def __init__(self, seq, lex_pos, line, column, allowed=None, considered_tokens=None, state=None, token_history=None,
-                 terminals_by_name=None, considered_rules=None):
-        super(UnexpectedCharacters, self).__init__()
-
-        ##
-
-        self.line = line
-        self.column = column
-        self.pos_in_stream = lex_pos
-        self.state = state
-        self._terminals_by_name = terminals_by_name
-
-        self.allowed = allowed
-        self.considered_tokens = considered_tokens
-        self.considered_rules = considered_rules
-        self.token_history = token_history
-
-        if isinstance(seq, bytes):
-            self.char = seq[lex_pos:lex_pos + 1].decode("ascii", "backslashreplace")
-        else:
-            self.char = seq[lex_pos]
-        self._context = self.get_context(seq)
-
-
-    def __str__(self):
-        message = "No terminal matches '%s' in the current parser context, at line %d col %d" % (self.char, self.line, self.column)
-        message += '\n\n' + self._context
-        if self.allowed:
-            message += self._format_expected(self.allowed)
-        if self.token_history:
-            message += '\nPrevious tokens: %s\n' % ', '.join(repr(t) for t in self.token_history)
-        return message
-
-
-class UnexpectedToken(ParseError, UnexpectedInput):
-    #--
-
-    expected: Set[str]
-    considered_rules: Set[str]
-    interactive_parser: 'InteractiveParser'
-
-    def __init__(self, token, expected, considered_rules=None, state=None, interactive_parser=None, terminals_by_name=None, token_history=None):
-        super(UnexpectedToken, self).__init__()
-
-        ##
-
-        self.line = getattr(token, 'line', '?')
-        self.column = getattr(token, 'column', '?')
-        self.pos_in_stream = getattr(token, 'start_pos', None)
-        self.state = state
-
-        self.token = token
-        self.expected = expected  ##
-
-        self._accepts = NO_VALUE
-        self.considered_rules = considered_rules
-        self.interactive_parser = interactive_parser
-        self._terminals_by_name = terminals_by_name
-        self.token_history = token_history
-
-
-    @property
-    def accepts(self) -> Set[str]:
-        if self._accepts is NO_VALUE:
-            self._accepts = self.interactive_parser and self.interactive_parser.accepts()
-        return self._accepts
-
-    def __str__(self):
-        message = ("Unexpected token %r at line %s, column %s.\n%s"
-                   % (self.token, self.line, self.column, self._format_expected(self.accepts or self.expected)))
-        if self.token_history:
-            message += "Previous tokens: %r\n" % self.token_history
-
-        return message
-
-
-
-class VisitError(LarkError):
-    #--
-
-    obj: 'Union[Tree, Token]'
-    orig_exc: Exception
-
-    def __init__(self, rule, obj, orig_exc):
-        message = 'Error trying to process rule "%s":\n\n%s' % (rule, orig_exc)
-        super(VisitError, self).__init__(message)
-
-        self.rule = rule
-        self.obj = obj
-        self.orig_exc = orig_exc
-
-
-class MissingVariableError(LarkError):
-    pass
-
-
-import sys, re
-import logging
-
-logger: logging.Logger = logging.getLogger("lark")
-logger.addHandler(logging.StreamHandler())
-##
-
-##
-
-logger.setLevel(logging.CRITICAL)
-
-
-NO_VALUE = object()
-
-T = TypeVar("T")
-
-
-def classify(seq: Iterable, key: Optional[Callable] = None, value: Optional[Callable] = None) -> Dict:
-    d: Dict[Any, Any] = {}
-    for item in seq:
-        k = key(item) if (key is not None) else item
-        v = value(item) if (value is not None) else item
-        try:
-            d[k].append(v)
-        except KeyError:
-            d[k] = [v]
-    return d
-
-
-def _deserialize(data: Any, namespace: Dict[str, Any], memo: Dict) -> Any:
-    if isinstance(data, dict):
-        if '__type__' in data:  ##
-
-            class_ = namespace[data['__type__']]
-            return class_.deserialize(data, memo)
-        elif '@' in data:
-            return memo[data['@']]
-        return {key:_deserialize(value, namespace, memo) for key, value in data.items()}
-    elif isinstance(data, list):
-        return [_deserialize(value, namespace, memo) for value in data]
-    return data
-
-
-_T = TypeVar("_T", bound="Serialize")
-
-class Serialize:
-    #--
-
-    def memo_serialize(self, types_to_memoize: List) -> Any:
-        memo = SerializeMemoizer(types_to_memoize)
-        return self.serialize(memo), memo.serialize()
-
-    def serialize(self, memo = None) -> Dict[str, Any]:
-        if memo and memo.in_types(self):
-            return {'@': memo.memoized.get(self)}
-
-        fields = getattr(self, '__serialize_fields__')
-        res = {f: _serialize(getattr(self, f), memo) for f in fields}
-        res['__type__'] = type(self).__name__
-        if hasattr(self, '_serialize'):
-            self._serialize(res, memo)  ##
-
-        return res
-
-    @classmethod
-    def deserialize(cls: Type[_T], data: Dict[str, Any], memo: Dict[int, Any]) -> _T:
-        namespace = getattr(cls, '__serialize_namespace__', [])
-        namespace = {c.__name__:c for c in namespace}
-
-        fields = getattr(cls, '__serialize_fields__')
-
-        if '@' in data:
-            return memo[data['@']]
-
-        inst = cls.__new__(cls)
-        for f in fields:
-            try:
-                setattr(inst, f, _deserialize(data[f], namespace, memo))
-            except KeyError as e:
-                raise KeyError("Cannot find key for class", cls, e)
-
-        if hasattr(inst, '_deserialize'):
-            inst._deserialize()  ##
-
-
-        return inst
-
-
-class SerializeMemoizer(Serialize):
-    #--
-
-    __serialize_fields__ = 'memoized',
-
-    def __init__(self, types_to_memoize: List) -> None:
-        self.types_to_memoize = tuple(types_to_memoize)
-        self.memoized = Enumerator()
-
-    def in_types(self, value: Serialize) -> bool:
-        return isinstance(value, self.types_to_memoize)
-
-    def serialize(self) -> Dict[int, Any]:  ##
-
-        return _serialize(self.memoized.reversed(), None)
-
-    @classmethod
-    def deserialize(cls, data: Dict[int, Any], namespace: Dict[str, Any], memo: Dict[Any, Any]) -> Dict[int, Any]:  ##
-
-        return _deserialize(data, namespace, memo)
-
-
-try:
-    import regex
-    _has_regex = True
-except ImportError:
-    _has_regex = False
-
-if sys.version_info >= (3, 11):
-    import re._parser as sre_parse
-    import re._constants as sre_constants
-else:
-    import sre_parse
-    import sre_constants
-
-categ_pattern = re.compile(r'\\p{[A-Za-z_]+}')
-
-def get_regexp_width(expr: str) -> Union[Tuple[int, int], List[int]]:
-    if _has_regex:
-        ##
-
-        ##
-
-        ##
-
-        regexp_final = re.sub(categ_pattern, 'A', expr)
-    else:
-        if re.search(categ_pattern, expr):
-            raise ImportError('`regex` module must be installed in order to use Unicode categories.', expr)
-        regexp_final = expr
-    try:
-        ##
-
-        return [int(x) for x in sre_parse.parse(regexp_final).getwidth()]   ##
-
-    except sre_constants.error:
-        if not _has_regex:
-            raise ValueError(expr)
-        else:
-            ##
-
-            ##
-
-            c = regex.compile(regexp_final)
-            if c.match('') is None:
-                ##
-
-                return 1, int(sre_constants.MAXREPEAT)
-            else:
-                return 0, int(sre_constants.MAXREPEAT)
-
-
-from collections import OrderedDict
-
-class Meta:
-
-    empty: bool
-    line: int
-    column: int
-    start_pos: int
-    end_line: int
-    end_column: int
-    end_pos: int
-    orig_expansion: 'List[TerminalDef]'
-    match_tree: bool
-
-    def __init__(self):
-        self.empty = True
-
-
-_Leaf_T = TypeVar("_Leaf_T")
-Branch = Union[_Leaf_T, 'Tree[_Leaf_T]']
-
-
-class Tree(Generic[_Leaf_T]):
-    #--
-
-    data: str
-    children: 'List[Branch[_Leaf_T]]'
-
-    def __init__(self, data: str, children: 'List[Branch[_Leaf_T]]', meta: Optional[Meta]=None) -> None:
-        self.data = data
-        self.children = children
-        self._meta = meta
-
-    @property
-    def meta(self) -> Meta:
-        if self._meta is None:
-            self._meta = Meta()
-        return self._meta
-
-    def __repr__(self):
-        return 'Tree(%r, %r)' % (self.data, self.children)
-
-    def _pretty_label(self):
-        return self.data
-
-    def _pretty(self, level, indent_str):
-        yield f'{indent_str*level}{self._pretty_label()}'
-        if len(self.children) == 1 and not isinstance(self.children[0], Tree):
-            yield f'\t{self.children[0]}\n'
-        else:
-            yield '\n'
-            for n in self.children:
-                if isinstance(n, Tree):
-                    yield from n._pretty(level+1, indent_str)
-                else:
-                    yield f'{indent_str*(level+1)}{n}\n'
-
-    def pretty(self, indent_str: str='  ') -> str:
-        #--
-        return ''.join(self._pretty(0, indent_str))
-
-    def __rich__(self, parent:Optional['rich.tree.Tree']=None) -> 'rich.tree.Tree':
-        #--
-        return self._rich(parent)
-
-    def _rich(self, parent):
-        if parent:
-            tree = parent.add(f'[bold]{self.data}[/bold]')
-        else:
-            import rich.tree
-            tree = rich.tree.Tree(self.data)
-
-        for c in self.children:
-            if isinstance(c, Tree):
-                c._rich(tree)
-            else:
-                tree.add(f'[green]{c}[/green]')
-
-        return tree
-
-    def __eq__(self, other):
-        try:
-            return self.data == other.data and self.children == other.children
-        except AttributeError:
-            return False
-
-    def __ne__(self, other):
-        return not (self == other)
-
-    def __hash__(self) -> int:
-        return hash((self.data, tuple(self.children)))
-
-    def iter_subtrees(self) -> 'Iterator[Tree[_Leaf_T]]':
-        #--
-        queue = [self]
-        subtrees = OrderedDict()
-        for subtree in queue:
-            subtrees[id(subtree)] = subtree
-            ##
-
-            queue += [c for c in reversed(subtree.children)  ##
-
-                      if isinstance(c, Tree) and id(c) not in subtrees]
-
-        del queue
-        return reversed(list(subtrees.values()))
-
-    def iter_subtrees_topdown(self):
-        #--
-        stack = [self]
-        stack_append = stack.append
-        stack_pop = stack.pop
-        while stack:
-            node = stack_pop()
-            if not isinstance(node, Tree):
-                continue
-            yield node
-            for child in reversed(node.children):
-                stack_append(child)
-
-    def find_pred(self, pred: 'Callable[[Tree[_Leaf_T]], bool]') -> 'Iterator[Tree[_Leaf_T]]':
-        #--
-        return filter(pred, self.iter_subtrees())
-
-    def find_data(self, data: str) -> 'Iterator[Tree[_Leaf_T]]':
-        #--
-        return self.find_pred(lambda t: t.data == data)
-
-
-from functools import wraps, update_wrapper
-from inspect import getmembers, getmro
-
-_Return_T = TypeVar('_Return_T')
-_Return_V = TypeVar('_Return_V')
-_Leaf_T = TypeVar('_Leaf_T')
-_Leaf_U = TypeVar('_Leaf_U')
-_R = TypeVar('_R')
-_FUNC = Callable[..., _Return_T]
-_DECORATED = Union[_FUNC, type]
-
-class _DiscardType:
-    #--
-
-    def __repr__(self):
-        return "lark.visitors.Discard"
-
-Discard = _DiscardType()
-
-##
-
-
-class _Decoratable:
-    #--
-
-    @classmethod
-    def _apply_v_args(cls, visit_wrapper):
-        mro = getmro(cls)
-        assert mro[0] is cls
-        libmembers = {name for _cls in mro[1:] for name, _ in getmembers(_cls)}
-        for name, value in getmembers(cls):
-
-            ##
-
-            if name.startswith('_') or (name in libmembers and name not in cls.__dict__):
-                continue
-            if not callable(value):
-                continue
-
-            ##
-
-            if isinstance(cls.__dict__[name], _VArgsWrapper):
-                continue
-
-            setattr(cls, name, _VArgsWrapper(cls.__dict__[name], visit_wrapper))
-        return cls
-
-    def __class_getitem__(cls, _):
-        return cls
-
-
-class Transformer(_Decoratable, ABC, Generic[_Leaf_T, _Return_T]):
-    #--
-    __visit_tokens__ = True   ##
-
-
-    def __init__(self,  visit_tokens: bool=True) -> None:
-        self.__visit_tokens__ = visit_tokens
-
-    def _call_userfunc(self, tree, new_children=None):
-        ##
-
-        children = new_children if new_children is not None else tree.children
-        try:
-            f = getattr(self, tree.data)
-        except AttributeError:
-            return self.__default__(tree.data, children, tree.meta)
-        else:
-            try:
-                wrapper = getattr(f, 'visit_wrapper', None)
-                if wrapper is not None:
-                    return f.visit_wrapper(f, tree.data, children, tree.meta)
-                else:
-                    return f(children)
-            except GrammarError:
-                raise
-            except Exception as e:
-                raise VisitError(tree.data, tree, e)
-
-    def _call_userfunc_token(self, token):
-        try:
-            f = getattr(self, token.type)
-        except AttributeError:
-            return self.__default_token__(token)
-        else:
-            try:
-                return f(token)
-            except GrammarError:
-                raise
-            except Exception as e:
-                raise VisitError(token.type, token, e)
-
-    def _transform_children(self, children):
-        for c in children:
-            if isinstance(c, Tree):
-                res = self._transform_tree(c)
-            elif self.__visit_tokens__ and isinstance(c, Token):
-                res = self._call_userfunc_token(c)
-            else:
-                res = c
-
-            if res is not Discard:
-                yield res
-
-    def _transform_tree(self, tree):
-        children = list(self._transform_children(tree.children))
-        return self._call_userfunc(tree, children)
-
-    def transform(self, tree: Tree[_Leaf_T]) -> _Return_T:
-        #--
-        return self._transform_tree(tree)
-
-    def __mul__(
-            self: 'Transformer[_Leaf_T, Tree[_Leaf_U]]',
-            other: 'Union[Transformer[_Leaf_U, _Return_V], TransformerChain[_Leaf_U, _Return_V,]]'
-    ) -> 'TransformerChain[_Leaf_T, _Return_V]':
-        #--
-        return TransformerChain(self, other)
-
-    def __default__(self, data, children, meta):
-        #--
-        return Tree(data, children, meta)
-
-    def __default_token__(self, token):
-        #--
-        return token
-
-
-def merge_transformers(base_transformer=None, **transformers_to_merge):
-    #--
-    if base_transformer is None:
-        base_transformer = Transformer()
-    for prefix, transformer in transformers_to_merge.items():
-        for method_name in dir(transformer):
-            method = getattr(transformer, method_name)
-            if not callable(method):
-                continue
-            if method_name.startswith("_") or method_name == "transform":
-                continue
-            prefixed_method = prefix + "__" + method_name
-            if hasattr(base_transformer, prefixed_method):
-                raise AttributeError("Cannot merge: method '%s' appears more than once" % prefixed_method)
-
-            setattr(base_transformer, prefixed_method, method)
-
-    return base_transformer
-
-
-class InlineTransformer(Transformer):   ##
-
-    def _call_userfunc(self, tree, new_children=None):
-        ##
-
-        children = new_children if new_children is not None else tree.children
-        try:
-            f = getattr(self, tree.data)
-        except AttributeError:
-            return self.__default__(tree.data, children, tree.meta)
-        else:
-            return f(*children)
-
-
-class TransformerChain(Generic[_Leaf_T, _Return_T]):
-
-    transformers: 'Tuple[Union[Transformer, TransformerChain], ...]'
-
-    def __init__(self, *transformers: 'Union[Transformer, TransformerChain]') -> None:
-        self.transformers = transformers
-
-    def transform(self, tree: Tree[_Leaf_T]) -> _Return_T:
-        for t in self.transformers:
-            tree = t.transform(tree)
-        return cast(_Return_T, tree)
-
-    def __mul__(
-            self: 'TransformerChain[_Leaf_T, Tree[_Leaf_U]]',
-            other: 'Union[Transformer[_Leaf_U, _Return_V], TransformerChain[_Leaf_U, _Return_V]]'
-    ) -> 'TransformerChain[_Leaf_T, _Return_V]':
-        return TransformerChain(*self.transformers + (other,))
-
-
-class Transformer_InPlace(Transformer):
-    #--
-    def _transform_tree(self, tree):           ##
-
-        return self._call_userfunc(tree)
-
-    def transform(self, tree: Tree[_Leaf_T]) -> _Return_T:
-        for subtree in tree.iter_subtrees():
-            subtree.children = list(self._transform_children(subtree.children))
-
-        return self._transform_tree(tree)
-
-
-class Transformer_NonRecursive(Transformer):
-    #--
-
-    def transform(self, tree: Tree[_Leaf_T]) -> _Return_T:
-        ##
-
-        rev_postfix = []
-        q: List[Branch[_Leaf_T]] = [tree]
-        while q:
-            t = q.pop()
-            rev_postfix.append(t)
-            if isinstance(t, Tree):
-                q += t.children
-
-        ##
-
-        stack: List = []
-        for x in reversed(rev_postfix):
-            if isinstance(x, Tree):
-                size = len(x.children)
-                if size:
-                    args = stack[-size:]
-                    del stack[-size:]
-                else:
-                    args = []
-
-                res = self._call_userfunc(x, args)
-                if res is not Discard:
-                    stack.append(res)
-
-            elif self.__visit_tokens__ and isinstance(x, Token):
-                res = self._call_userfunc_token(x)
-                if res is not Discard:
-                    stack.append(res)
-            else:
-                stack.append(x)
-
-        result, = stack  ##
-
-        ##
-
-        ##
-
-        ##
-
-        return cast(_Return_T, result)
-
-
-class Transformer_InPlaceRecursive(Transformer):
-    #--
-    def _transform_tree(self, tree):
-        tree.children = list(self._transform_children(tree.children))
-        return self._call_userfunc(tree)
-
-
-##
-
-
-class VisitorBase:
-    def _call_userfunc(self, tree):
-        return getattr(self, tree.data, self.__default__)(tree)
-
-    def __default__(self, tree):
-        #--
-        return tree
-
-    def __class_getitem__(cls, _):
-        return cls
-
-
-class Visitor(VisitorBase, ABC, Generic[_Leaf_T]):
-    #--
-
-    def visit(self, tree: Tree[_Leaf_T]) -> Tree[_Leaf_T]:
-        #--
-        for subtree in tree.iter_subtrees():
-            self._call_userfunc(subtree)
-        return tree
-
-    def visit_topdown(self, tree: Tree[_Leaf_T]) -> Tree[_Leaf_T]:
-        #--
-        for subtree in tree.iter_subtrees_topdown():
-            self._call_userfunc(subtree)
-        return tree
-
-
-class Visitor_Recursive(VisitorBase, Generic[_Leaf_T]):
-    #--
-
-    def visit(self, tree: Tree[_Leaf_T]) -> Tree[_Leaf_T]:
-        #--
-        for child in tree.children:
-            if isinstance(child, Tree):
-                self.visit(child)
-
-        self._call_userfunc(tree)
-        return tree
-
-    def visit_topdown(self,tree: Tree[_Leaf_T]) -> Tree[_Leaf_T]:
-        #--
-        self._call_userfunc(tree)
-
-        for child in tree.children:
-            if isinstance(child, Tree):
-                self.visit_topdown(child)
-
-        return tree
-
-
-class Interpreter(_Decoratable, ABC, Generic[_Leaf_T, _Return_T]):
-    #--
-
-    def visit(self, tree: Tree[_Leaf_T]) -> _Return_T:
-        ##
-
-        ##
-
-        ##
-
-        return self._visit_tree(tree)
-
-    def _visit_tree(self, tree: Tree[_Leaf_T]):
-        f = getattr(self, tree.data)
-        wrapper = getattr(f, 'visit_wrapper', None)
-        if wrapper is not None:
-            return f.visit_wrapper(f, tree.data, tree.children, tree.meta)
-        else:
-            return f(tree)
-
-    def visit_children(self, tree: Tree[_Leaf_T]) -> List:
-        return [self._visit_tree(child) if isinstance(child, Tree) else child
-                for child in tree.children]
-
-    def __getattr__(self, name):
-        return self.__default__
-
-    def __default__(self, tree):
-        return self.visit_children(tree)
-
-
-_InterMethod = Callable[[Type[Interpreter], _Return_T], _R]
-
-def visit_children_decor(func: _InterMethod) -> _InterMethod:
-    #--
-    @wraps(func)
-    def inner(cls, tree):
-        values = cls.visit_children(tree)
-        return func(cls, values)
-    return inner
-
-##
-
-
-def _apply_v_args(obj, visit_wrapper):
-    try:
-        _apply = obj._apply_v_args
-    except AttributeError:
-        return _VArgsWrapper(obj, visit_wrapper)
-    else:
-        return _apply(visit_wrapper)
-
-
-class _VArgsWrapper:
-    #--
-    base_func: Callable
-
-    def __init__(self, func: Callable, visit_wrapper: Callable[[Callable, str, list, Any], Any]):
-        if isinstance(func, _VArgsWrapper):
-            func = func.base_func
-        ##
-
-        self.base_func = func  ##
-
-        self.visit_wrapper = visit_wrapper
-        update_wrapper(self, func)
-
-    def __call__(self, *args, **kwargs):
-        return self.base_func(*args, **kwargs)
-
-    def __get__(self, instance, owner=None):
-        try:
-            ##
-
-            ##
-
-            g = type(self.base_func).__get__
-        except AttributeError:
-            return self
-        else:
-            return _VArgsWrapper(g(self.base_func, instance, owner), self.visit_wrapper)
-
-    def __set_name__(self, owner, name):
-        try:
-            f = type(self.base_func).__set_name__
-        except AttributeError:
-            return
-        else:
-            f(self.base_func, owner, name)
-
-
-def _vargs_inline(f, _data, children, _meta):
-    return f(*children)
-def _vargs_meta_inline(f, _data, children, meta):
-    return f(meta, *children)
-def _vargs_meta(f, _data, children, meta):
-    return f(meta, children)
-def _vargs_tree(f, data, children, meta):
-    return f(Tree(data, children, meta))
-
-
-def v_args(inline: bool = False, meta: bool = False, tree: bool = False, wrapper: Optional[Callable] = None) -> Callable[[_DECORATED], _DECORATED]:
-    #--
-    if tree and (meta or inline):
-        raise ValueError("Visitor functions cannot combine 'tree' with 'meta' or 'inline'.")
-
-    func = None
-    if meta:
-        if inline:
-            func = _vargs_meta_inline
-        else:
-            func = _vargs_meta
-    elif inline:
-        func = _vargs_inline
-    elif tree:
-        func = _vargs_tree
-
-    if wrapper is not None:
-        if func is not None:
-            raise ValueError("Cannot use 'wrapper' along with 'tree', 'meta' or 'inline'.")
-        func = wrapper
-
-    def _visitor_args_dec(obj):
-        return _apply_v_args(obj, func)
-    return _visitor_args_dec
-
-
-
-TOKEN_DEFAULT_PRIORITY = 0
-
-
-class Symbol(Serialize):
-    __slots__ = ('name',)
-
-    name: str
-    is_term: ClassVar[bool] = NotImplemented
-
-    def __init__(self, name: str) -> None:
-        self.name = name
-
-    def __eq__(self, other):
-        assert isinstance(other, Symbol), other
-        return self.is_term == other.is_term and self.name == other.name
-
-    def __ne__(self, other):
-        return not (self == other)
-
-    def __hash__(self):
-        return hash(self.name)
-
-    def __repr__(self):
-        return '%s(%r)' % (type(self).__name__, self.name)
-
-    fullrepr = property(__repr__)
-
-    def renamed(self, f):
-        return type(self)(f(self.name))
-
-
-class Terminal(Symbol):
-    __serialize_fields__ = 'name', 'filter_out'
-
-    is_term: ClassVar[bool] = True
-
-    def __init__(self, name, filter_out=False):
-        self.name = name
-        self.filter_out = filter_out
-
-    @property
-    def fullrepr(self):
-        return '%s(%r, %r)' % (type(self).__name__, self.name, self.filter_out)
-
-    def renamed(self, f):
-        return type(self)(f(self.name), self.filter_out)
-
-
-class NonTerminal(Symbol):
-    __serialize_fields__ = 'name',
-
-    is_term: ClassVar[bool] = False
-
-
-class RuleOptions(Serialize):
-    __serialize_fields__ = 'keep_all_tokens', 'expand1', 'priority', 'template_source', 'empty_indices'
-
-    keep_all_tokens: bool
-    expand1: bool
-    priority: Optional[int]
-    template_source: Optional[str]
-    empty_indices: Tuple[bool, ...]
-
-    def __init__(self, keep_all_tokens: bool=False, expand1: bool=False, priority: Optional[int]=None, template_source: Optional[str]=None, empty_indices: Tuple[bool, ...]=()) -> None:
-        self.keep_all_tokens = keep_all_tokens
-        self.expand1 = expand1
-        self.priority = priority
-        self.template_source = template_source
-        self.empty_indices = empty_indices
-
-    def __repr__(self):
-        return 'RuleOptions(%r, %r, %r, %r)' % (
-            self.keep_all_tokens,
-            self.expand1,
-            self.priority,
-            self.template_source
-        )
-
-
-class Rule(Serialize):
-    #--
-    __slots__ = ('origin', 'expansion', 'alias', 'options', 'order', '_hash')
-
-    __serialize_fields__ = 'origin', 'expansion', 'order', 'alias', 'options'
-    __serialize_namespace__ = Terminal, NonTerminal, RuleOptions
-
-    def __init__(self, origin, expansion, order=0, alias=None, options=None):
-        self.origin = origin
-        self.expansion = expansion
-        self.alias = alias
-        self.order = order
-        self.options = options or RuleOptions()
-        self._hash = hash((self.origin, tuple(self.expansion)))
-
-    def _deserialize(self):
-        self._hash = hash((self.origin, tuple(self.expansion)))
-
-    def __str__(self):
-        return '<%s : %s>' % (self.origin.name, ' '.join(x.name for x in self.expansion))
-
-    def __repr__(self):
-        return 'Rule(%r, %r, %r, %r)' % (self.origin, self.expansion, self.alias, self.options)
-
-    def __hash__(self):
-        return self._hash
-
-    def __eq__(self, other):
-        if not isinstance(other, Rule):
-            return False
-        return self.origin == other.origin and self.expansion == other.expansion
-
-
-
-from copy import copy
-
-try:  ##
-
-    has_interegular = bool(interegular)
-except NameError:
-    has_interegular = False
-
-class Pattern(Serialize, ABC):
-
-    value: str
-    flags: Collection[str]
-    raw: Optional[str]
-    type: ClassVar[str]
-
-    def __init__(self, value: str, flags: Collection[str] = (), raw: Optional[str] = None) -> None:
-        self.value = value
-        self.flags = frozenset(flags)
-        self.raw = raw
-
-    def __repr__(self):
-        return repr(self.to_regexp())
-
-    ##
-
-    def __hash__(self):
-        return hash((type(self), self.value, self.flags))
-
-    def __eq__(self, other):
-        return type(self) == type(other) and self.value == other.value and self.flags == other.flags
-
-    @abstractmethod
-    def to_regexp(self) -> str:
-        raise NotImplementedError()
-
-    @property
-    @abstractmethod
-    def min_width(self) -> int:
-        raise NotImplementedError()
-
-    @property
-    @abstractmethod
-    def max_width(self) -> int:
-        raise NotImplementedError()
-
-    def _get_flags(self, value):
-        for f in self.flags:
-            value = ('(?%s:%s)' % (f, value))
-        return value
-
-
-class PatternStr(Pattern):
-    __serialize_fields__ = 'value', 'flags', 'raw'
-
-    type: ClassVar[str] = "str"
-
-    def to_regexp(self) -> str:
-        return self._get_flags(re.escape(self.value))
-
-    @property
-    def min_width(self) -> int:
-        return len(self.value)
-
-    @property
-    def max_width(self) -> int:
-        return len(self.value)
-
-
-class PatternRE(Pattern):
-    __serialize_fields__ = 'value', 'flags', 'raw', '_width'
-
-    type: ClassVar[str] = "re"
-
-    def to_regexp(self) -> str:
-        return self._get_flags(self.value)
-
-    _width = None
-    def _get_width(self):
-        if self._width is None:
-            self._width = get_regexp_width(self.to_regexp())
-        return self._width
-
-    @property
-    def min_width(self) -> int:
-        return self._get_width()[0]
-
-    @property
-    def max_width(self) -> int:
-        return self._get_width()[1]
-
-
-class TerminalDef(Serialize):
-    __serialize_fields__ = 'name', 'pattern', 'priority'
-    __serialize_namespace__ = PatternStr, PatternRE
-
-    name: str
-    pattern: Pattern
-    priority: int
-
-    def __init__(self, name: str, pattern: Pattern, priority: int = TOKEN_DEFAULT_PRIORITY) -> None:
-        assert isinstance(pattern, Pattern), pattern
-        self.name = name
-        self.pattern = pattern
-        self.priority = priority
-
-    def __repr__(self):
-        return '%s(%r, %r)' % (type(self).__name__, self.name, self.pattern)
-
-    def user_repr(self) -> str:
-        if self.name.startswith('__'):  ##
-
-            return self.pattern.raw or self.name
-        else:
-            return self.name
-
-_T = TypeVar('_T', bound="Token")
-
-class Token(str):
-    #--
-    __slots__ = ('type', 'start_pos', 'value', 'line', 'column', 'end_line', 'end_column', 'end_pos')
-
-    __match_args__ = ('type', 'value')
-
-    type: str
-    start_pos: Optional[int]
-    value: Any
-    line: Optional[int]
-    column: Optional[int]
-    end_line: Optional[int]
-    end_column: Optional[int]
-    end_pos: Optional[int]
-
-
-    @overload
-    def __new__(
-            cls,
-            type: str,
-            value: Any,
-            start_pos: Optional[int] = None,
-            line: Optional[int] = None,
-            column: Optional[int] = None,
-            end_line: Optional[int] = None,
-            end_column: Optional[int] = None,
-            end_pos: Optional[int] = None
-    ) -> 'Token':
-        ...
-
-    @overload
-    def __new__(
-            cls,
-            type_: str,
-            value: Any,
-            start_pos: Optional[int] = None,
-            line: Optional[int] = None,
-            column: Optional[int] = None,
-            end_line: Optional[int] = None,
-            end_column: Optional[int] = None,
-            end_pos: Optional[int] = None
-    ) -> 'Token':        ...
-
-    def __new__(cls, *args, **kwargs):
-        if "type_" in kwargs:
-            warnings.warn("`type_` is deprecated use `type` instead", DeprecationWarning)
-
-            if "type" in kwargs:
-                raise TypeError("Error: using both 'type' and the deprecated 'type_' as arguments.")
-            kwargs["type"] = kwargs.pop("type_")
-
-        return cls._future_new(*args, **kwargs)
-
-
-    @classmethod
-    def _future_new(cls, type, value, start_pos=None, line=None, column=None, end_line=None, end_column=None, end_pos=None):
-        inst = super(Token, cls).__new__(cls, value)
-
-        inst.type = type
-        inst.start_pos = start_pos
-        inst.value = value
-        inst.line = line
-        inst.column = column
-        inst.end_line = end_line
-        inst.end_column = end_column
-        inst.end_pos = end_pos
-        return inst
-
-    @overload
-    def update(self, type: Optional[str] = None, value: Optional[Any] = None) -> 'Token':
-        ...
-
-    @overload
-    def update(self, type_: Optional[str] = None, value: Optional[Any] = None) -> 'Token':
-        ...
-
-    def update(self, *args, **kwargs):
-        if "type_" in kwargs:
-            warnings.warn("`type_` is deprecated use `type` instead", DeprecationWarning)
-
-            if "type" in kwargs:
-                raise TypeError("Error: using both 'type' and the deprecated 'type_' as arguments.")
-            kwargs["type"] = kwargs.pop("type_")
-
-        return self._future_update(*args, **kwargs)
-
-    def _future_update(self, type: Optional[str] = None, value: Optional[Any] = None) -> 'Token':
-        return Token.new_borrow_pos(
-            type if type is not None else self.type,
-            value if value is not None else self.value,
-            self
-        )
-
-    @classmethod
-    def new_borrow_pos(cls: Type[_T], type_: str, value: Any, borrow_t: 'Token') -> _T:
-        return cls(type_, value, borrow_t.start_pos, borrow_t.line, borrow_t.column, borrow_t.end_line, borrow_t.end_column, borrow_t.end_pos)
-
-    def __reduce__(self):
-        return (self.__class__, (self.type, self.value, self.start_pos, self.line, self.column))
-
-    def __repr__(self):
-        return 'Token(%r, %r)' % (self.type, self.value)
-
-    def __deepcopy__(self, memo):
-        return Token(self.type, self.value, self.start_pos, self.line, self.column)
-
-    def __eq__(self, other):
-        if isinstance(other, Token) and self.type != other.type:
-            return False
-
-        return str.__eq__(self, other)
-
-    __hash__ = str.__hash__
-
-
-class LineCounter:
-    __slots__ = 'char_pos', 'line', 'column', 'line_start_pos', 'newline_char'
-
-    def __init__(self, newline_char):
-        self.newline_char = newline_char
-        self.char_pos = 0
-        self.line = 1
-        self.column = 1
-        self.line_start_pos = 0
-
-    def __eq__(self, other):
-        if not isinstance(other, LineCounter):
-            return NotImplemented
-
-        return self.char_pos == other.char_pos and self.newline_char == other.newline_char
-
-    def feed(self, token: Token, test_newline=True):
-        #--
-        if test_newline:
-            newlines = token.count(self.newline_char)
-            if newlines:
-                self.line += newlines
-                self.line_start_pos = self.char_pos + token.rindex(self.newline_char) + 1
-
-        self.char_pos += len(token)
-        self.column = self.char_pos - self.line_start_pos + 1
-
-
-class UnlessCallback:
-    def __init__(self, scanner):
-        self.scanner = scanner
-
-    def __call__(self, t):
-        res = self.scanner.match(t.value, 0)
-        if res:
-            _value, t.type = res
-        return t
-
-
-class CallChain:
-    def __init__(self, callback1, callback2, cond):
-        self.callback1 = callback1
-        self.callback2 = callback2
-        self.cond = cond
-
-    def __call__(self, t):
-        t2 = self.callback1(t)
-        return self.callback2(t) if self.cond(t2) else t2
-
-
-def _get_match(re_, regexp, s, flags):
-    m = re_.match(regexp, s, flags)
-    if m:
-        return m.group(0)
-
-def _create_unless(terminals, g_regex_flags, re_, use_bytes):
-    tokens_by_type = classify(terminals, lambda t: type(t.pattern))
-    assert len(tokens_by_type) <= 2, tokens_by_type.keys()
-    embedded_strs = set()
-    callback = {}
-    for retok in tokens_by_type.get(PatternRE, []):
-        unless = []
-        for strtok in tokens_by_type.get(PatternStr, []):
-            if strtok.priority != retok.priority:
-                continue
-            s = strtok.pattern.value
-            if s == _get_match(re_, retok.pattern.to_regexp(), s, g_regex_flags):
-                unless.append(strtok)
-                if strtok.pattern.flags <= retok.pattern.flags:
-                    embedded_strs.add(strtok)
-        if unless:
-            callback[retok.name] = UnlessCallback(Scanner(unless, g_regex_flags, re_, match_whole=True, use_bytes=use_bytes))
-
-    new_terminals = [t for t in terminals if t not in embedded_strs]
-    return new_terminals, callback
-
-
-class Scanner:
-    def __init__(self, terminals, g_regex_flags, re_, use_bytes, match_whole=False):
-        self.terminals = terminals
-        self.g_regex_flags = g_regex_flags
-        self.re_ = re_
-        self.use_bytes = use_bytes
-        self.match_whole = match_whole
-
-        self.allowed_types = {t.name for t in self.terminals}
-
-        self._mres = self._build_mres(terminals, len(terminals))
-
-    def _build_mres(self, terminals, max_size):
-        ##
-
-        ##
-
-        ##
-
-        postfix = '$' if self.match_whole else ''
-        mres = []
-        while terminals:
-            pattern = u'|'.join(u'(?P<%s>%s)' % (t.name, t.pattern.to_regexp() + postfix) for t in terminals[:max_size])
-            if self.use_bytes:
-                pattern = pattern.encode('latin-1')
-            try:
-                mre = self.re_.compile(pattern, self.g_regex_flags)
-            except AssertionError:  ##
-
-                return self._build_mres(terminals, max_size // 2)
-
-            mres.append(mre)
-            terminals = terminals[max_size:]
-        return mres
-
-    def match(self, text, pos):
-        for mre in self._mres:
-            m = mre.match(text, pos)
-            if m:
-                return m.group(0), m.lastgroup
-
-
-def _regexp_has_newline(r: str):
-    #--
-    return '\n' in r or '\\n' in r or '\\s' in r or '[^' in r or ('(?s' in r and '.' in r)
-
-
-class LexerState:
-    #--
-
-    __slots__ = 'text', 'line_ctr', 'last_token'
-
-    text: str
-    line_ctr: LineCounter
-    last_token: Optional[Token]
-
-    def __init__(self, text: str, line_ctr: Optional[LineCounter]=None, last_token: Optional[Token]=None):
-        self.text = text
-        self.line_ctr = line_ctr or LineCounter(b'\n' if isinstance(text, bytes) else '\n')
-        self.last_token = last_token
-
-    def __eq__(self, other):
-        if not isinstance(other, LexerState):
-            return NotImplemented
-
-        return self.text is other.text and self.line_ctr == other.line_ctr and self.last_token == other.last_token
-
-    def __copy__(self):
-        return type(self)(self.text, copy(self.line_ctr), self.last_token)
-
-
-class LexerThread:
-    #--
-
-    def __init__(self, lexer: 'Lexer', lexer_state: LexerState):
-        self.lexer = lexer
-        self.state = lexer_state
-
-    @classmethod
-    def from_text(cls, lexer: 'Lexer', text: str):
-        return cls(lexer, LexerState(text))
-
-    def lex(self, parser_state):
-        return self.lexer.lex(self.state, parser_state)
-
-    def __copy__(self):
-        return type(self)(self.lexer, copy(self.state))
-
-    _Token = Token
-
-
-_Callback = Callable[[Token], Token]
-
-class Lexer(ABC):
-    #--
-    @abstractmethod
-    def lex(self, lexer_state: LexerState, parser_state: Any) -> Iterator[Token]:
-        return NotImplemented
-
-    def make_lexer_state(self, text):
-        #--
-        return LexerState(text)
-
-
-def _check_regex_collisions(terminal_to_regexp: Dict[TerminalDef, str], comparator, strict_mode, max_collisions_to_show=8):
-    if not comparator:
-        comparator = interegular.Comparator.from_regexes(terminal_to_regexp)
-
-    ##
-
-    ##
-
-    max_time = 2 if strict_mode else 0.2
-
-    ##
-
-    if comparator.count_marked_pairs() >= max_collisions_to_show:
-        return
-    for group in classify(terminal_to_regexp, lambda t: t.priority).values():
-        for a, b in comparator.check(group, skip_marked=True):
-            assert a.priority == b.priority
-            ##
-
-            comparator.mark(a, b)
-
-            ##
-
-            message = f"Collision between Terminals {a.name} and {b.name}. "
-            try:
-                example = comparator.get_example_overlap(a, b, max_time).format_multiline()
-            except ValueError:
-                ##
-
-                example = "No example could be found fast enough. However, the collision does still exists"
-            if strict_mode:
-                raise LexError(f"{message}\n{example}")
-            logger.warning("%s The lexer will choose between them arbitrarily.\n%s", message, example)
-            if comparator.count_marked_pairs() >= max_collisions_to_show:
-                logger.warning("Found 8 regex collisions, will not check for more.")
-                return
-
-
-class BasicLexer(Lexer):
-    terminals: Collection[TerminalDef]
-    ignore_types: FrozenSet[str]
-    newline_types: FrozenSet[str]
-    user_callbacks: Dict[str, _Callback]
-    callback: Dict[str, _Callback]
-    re: ModuleType
-
-    def __init__(self, conf: 'LexerConf', comparator=None) -> None:
-        terminals = list(conf.terminals)
-        assert all(isinstance(t, TerminalDef) for t in terminals), terminals
-
-        self.re = conf.re_module
-
-        if not conf.skip_validation:
-            ##
-
-            terminal_to_regexp = {}
-            for t in terminals:
-                regexp = t.pattern.to_regexp()
-                try:
-                    self.re.compile(regexp, conf.g_regex_flags)
-                except self.re.error:
-                    raise LexError("Cannot compile token %s: %s" % (t.name, t.pattern))
-
-                if t.pattern.min_width == 0:
-                    raise LexError("Lexer does not allow zero-width terminals. (%s: %s)" % (t.name, t.pattern))
-                if t.pattern.type == "re":
-                    terminal_to_regexp[t] = regexp
-
-            if not (set(conf.ignore) <= {t.name for t in terminals}):
-                raise LexError("Ignore terminals are not defined: %s" % (set(conf.ignore) - {t.name for t in terminals}))
-
-            if has_interegular:
-                _check_regex_collisions(terminal_to_regexp, comparator, conf.strict)
-            elif conf.strict:
-                raise LexError("interegular must be installed for strict mode. Use `pip install 'lark[interegular]'`.")
-
-        ##
-
-        self.newline_types = frozenset(t.name for t in terminals if _regexp_has_newline(t.pattern.to_regexp()))
-        self.ignore_types = frozenset(conf.ignore)
-
-        terminals.sort(key=lambda x: (-x.priority, -x.pattern.max_width, -len(x.pattern.value), x.name))
-        self.terminals = terminals
-        self.user_callbacks = conf.callbacks
-        self.g_regex_flags = conf.g_regex_flags
-        self.use_bytes = conf.use_bytes
-        self.terminals_by_name = conf.terminals_by_name
-
-        self._scanner = None
-
-    def _build_scanner(self):
-        terminals, self.callback = _create_unless(self.terminals, self.g_regex_flags, self.re, self.use_bytes)
-        assert all(self.callback.values())
-
-        for type_, f in self.user_callbacks.items():
-            if type_ in self.callback:
-                ##
-
-                self.callback[type_] = CallChain(self.callback[type_], f, lambda t: t.type == type_)
-            else:
-                self.callback[type_] = f
-
-        self._scanner = Scanner(terminals, self.g_regex_flags, self.re, self.use_bytes)
-
-    @property
-    def scanner(self):
-        if self._scanner is None:
-            self._build_scanner()
-        return self._scanner
-
-    def match(self, text, pos):
-        return self.scanner.match(text, pos)
-
-    def lex(self, state: LexerState, parser_state: Any) -> Iterator[Token]:
-        with suppress(EOFError):
-            while True:
-                yield self.next_token(state, parser_state)
-
-    def next_token(self, lex_state: LexerState, parser_state: Any = None) -> Token:
-        line_ctr = lex_state.line_ctr
-        while line_ctr.char_pos < len(lex_state.text):
-            res = self.match(lex_state.text, line_ctr.char_pos)
-            if not res:
-                allowed = self.scanner.allowed_types - self.ignore_types
-                if not allowed:
-                    allowed = {"<END-OF-FILE>"}
-                raise UnexpectedCharacters(lex_state.text, line_ctr.char_pos, line_ctr.line, line_ctr.column,
-                                           allowed=allowed, token_history=lex_state.last_token and [lex_state.last_token],
-                                           state=parser_state, terminals_by_name=self.terminals_by_name)
-
-            value, type_ = res
-
-            if type_ not in self.ignore_types:
-                t = Token(type_, value, line_ctr.char_pos, line_ctr.line, line_ctr.column)
-                line_ctr.feed(value, type_ in self.newline_types)
-                t.end_line = line_ctr.line
-                t.end_column = line_ctr.column
-                t.end_pos = line_ctr.char_pos
-                if t.type in self.callback:
-                    t = self.callback[t.type](t)
-                    if not isinstance(t, Token):
-                        raise LexError("Callbacks must return a token (returned %r)" % t)
-                lex_state.last_token = t
-                return t
-            else:
-                if type_ in self.callback:
-                    t2 = Token(type_, value, line_ctr.char_pos, line_ctr.line, line_ctr.column)
-                    self.callback[type_](t2)
-                line_ctr.feed(value, type_ in self.newline_types)
-
-        ##
-
-        raise EOFError(self)
-
-
-class ContextualLexer(Lexer):
-
-    lexers: Dict[str, BasicLexer]
-    root_lexer: BasicLexer
-
-    def __init__(self, conf: 'LexerConf', states: Dict[str, Collection[str]], always_accept: Collection[str]=()) -> None:
-        terminals = list(conf.terminals)
-        terminals_by_name = conf.terminals_by_name
-
-        trad_conf = copy(conf)
-        trad_conf.terminals = terminals
-
-        if has_interegular and not conf.skip_validation:
-            comparator = interegular.Comparator.from_regexes({t: t.pattern.to_regexp() for t in terminals})
-        else:
-            comparator = None
-        lexer_by_tokens: Dict[FrozenSet[str], BasicLexer] = {}
-        self.lexers = {}
-        for state, accepts in states.items():
-            key = frozenset(accepts)
-            try:
-                lexer = lexer_by_tokens[key]
-            except KeyError:
-                accepts = set(accepts) | set(conf.ignore) | set(always_accept)
-                lexer_conf = copy(trad_conf)
-                lexer_conf.terminals = [terminals_by_name[n] for n in accepts if n in terminals_by_name]
-                lexer = BasicLexer(lexer_conf, comparator)
-                lexer_by_tokens[key] = lexer
-
-            self.lexers[state] = lexer
-
-        assert trad_conf.terminals is terminals
-        trad_conf.skip_validation = True  ##
-
-        self.root_lexer = BasicLexer(trad_conf, comparator)
-
-    def lex(self, lexer_state: LexerState, parser_state: Any) -> Iterator[Token]:
-        try:
-            while True:
-                lexer = self.lexers[parser_state.position]
-                yield lexer.next_token(lexer_state, parser_state)
-        except EOFError:
-            pass
-        except UnexpectedCharacters as e:
-            ##
-
-            ##
-
-            try:
-                last_token = lexer_state.last_token  ##
-
-                token = self.root_lexer.next_token(lexer_state, parser_state)
-                raise UnexpectedToken(token, e.allowed, state=parser_state, token_history=[last_token], terminals_by_name=self.root_lexer.terminals_by_name)
-            except UnexpectedCharacters:
-                raise e  ##
-
-
-
-
-_ParserArgType: 'TypeAlias' = 'Literal["earley", "lalr", "cyk", "auto"]'
-_LexerArgType: 'TypeAlias' = 'Union[Literal["auto", "basic", "contextual", "dynamic", "dynamic_complete"], Type[Lexer]]'
-_Callback = Callable[[Token], Token]
-
-class LexerConf(Serialize):
-    __serialize_fields__ = 'terminals', 'ignore', 'g_regex_flags', 'use_bytes', 'lexer_type'
-    __serialize_namespace__ = TerminalDef,
-
-    terminals: Collection[TerminalDef]
-    re_module: ModuleType
-    ignore: Collection[str]
-    postlex: 'Optional[PostLex]'
-    callbacks: Dict[str, _Callback]
-    g_regex_flags: int
-    skip_validation: bool
-    use_bytes: bool
-    lexer_type: Optional[_LexerArgType]
-    strict: bool
-
-    def __init__(self, terminals: Collection[TerminalDef], re_module: ModuleType, ignore: Collection[str]=(), postlex: 'Optional[PostLex]'=None,
-                 callbacks: Optional[Dict[str, _Callback]]=None, g_regex_flags: int=0, skip_validation: bool=False, use_bytes: bool=False, strict: bool=False):
-        self.terminals = terminals
-        self.terminals_by_name = {t.name: t for t in self.terminals}
-        assert len(self.terminals) == len(self.terminals_by_name)
-        self.ignore = ignore
-        self.postlex = postlex
-        self.callbacks = callbacks or {}
-        self.g_regex_flags = g_regex_flags
-        self.re_module = re_module
-        self.skip_validation = skip_validation
-        self.use_bytes = use_bytes
-        self.strict = strict
-        self.lexer_type = None
-
-    def _deserialize(self):
-        self.terminals_by_name = {t.name: t for t in self.terminals}
-
-    def __deepcopy__(self, memo=None):
-        return type(self)(
-            deepcopy(self.terminals, memo),
-            self.re_module,
-            deepcopy(self.ignore, memo),
-            deepcopy(self.postlex, memo),
-            deepcopy(self.callbacks, memo),
-            deepcopy(self.g_regex_flags, memo),
-            deepcopy(self.skip_validation, memo),
-            deepcopy(self.use_bytes, memo),
-        )
-
-
-class ParserConf(Serialize):
-    __serialize_fields__ = 'rules', 'start', 'parser_type'
-
-    def __init__(self, rules, callbacks, start):
-        assert isinstance(start, list)
-        self.rules = rules
-        self.callbacks = callbacks
-        self.start = start
-
-        self.parser_type = None
-
-
-from functools import partial, wraps
-from itertools import product
-
-
-class ExpandSingleChild:
-    def __init__(self, node_builder):
-        self.node_builder = node_builder
-
-    def __call__(self, children):
-        if len(children) == 1:
-            return children[0]
-        else:
-            return self.node_builder(children)
-
-
-
-class PropagatePositions:
-    def __init__(self, node_builder, node_filter=None):
-        self.node_builder = node_builder
-        self.node_filter = node_filter
-
-    def __call__(self, children):
-        res = self.node_builder(children)
-
-        if isinstance(res, Tree):
-            ##
-
-            ##
-
-            ##
-
-            ##
-
-
-            res_meta = res.meta
-
-            first_meta = self._pp_get_meta(children)
-            if first_meta is not None:
-                if not hasattr(res_meta, 'line'):
-                    ##
-
-                    res_meta.line = getattr(first_meta, 'container_line', first_meta.line)
-                    res_meta.column = getattr(first_meta, 'container_column', first_meta.column)
-                    res_meta.start_pos = getattr(first_meta, 'container_start_pos', first_meta.start_pos)
-                    res_meta.empty = False
-
-                res_meta.container_line = getattr(first_meta, 'container_line', first_meta.line)
-                res_meta.container_column = getattr(first_meta, 'container_column', first_meta.column)
-                res_meta.container_start_pos = getattr(first_meta, 'container_start_pos', first_meta.start_pos)
-
-            last_meta = self._pp_get_meta(reversed(children))
-            if last_meta is not None:
-                if not hasattr(res_meta, 'end_line'):
-                    res_meta.end_line = getattr(last_meta, 'container_end_line', last_meta.end_line)
-                    res_meta.end_column = getattr(last_meta, 'container_end_column', last_meta.end_column)
-                    res_meta.end_pos = getattr(last_meta, 'container_end_pos', last_meta.end_pos)
-                    res_meta.empty = False
-
-                res_meta.container_end_line = getattr(last_meta, 'container_end_line', last_meta.end_line)
-                res_meta.container_end_column = getattr(last_meta, 'container_end_column', last_meta.end_column)
-                res_meta.container_end_pos = getattr(last_meta, 'container_end_pos', last_meta.end_pos)
-
-        return res
-
-    def _pp_get_meta(self, children):
-        for c in children:
-            if self.node_filter is not None and not self.node_filter(c):
-                continue
-            if isinstance(c, Tree):
-                if not c.meta.empty:
-                    return c.meta
-            elif isinstance(c, Token):
-                return c
-            elif hasattr(c, '__lark_meta__'):
-                return c.__lark_meta__()
-
-def make_propagate_positions(option):
-    if callable(option):
-        return partial(PropagatePositions, node_filter=option)
-    elif option is True:
-        return PropagatePositions
-    elif option is False:
-        return None
-
-    raise ConfigurationError('Invalid option for propagate_positions: %r' % option)
-
-
-class ChildFilter:
-    def __init__(self, to_include, append_none, node_builder):
-        self.node_builder = node_builder
-        self.to_include = to_include
-        self.append_none = append_none
-
-    def __call__(self, children):
-        filtered = []
-
-        for i, to_expand, add_none in self.to_include:
-            if add_none:
-                filtered += [None] * add_none
-            if to_expand:
-                filtered += children[i].children
-            else:
-                filtered.append(children[i])
-
-        if self.append_none:
-            filtered += [None] * self.append_none
-
-        return self.node_builder(filtered)
-
-
-class ChildFilterLALR(ChildFilter):
-    #--
-
-    def __call__(self, children):
-        filtered = []
-        for i, to_expand, add_none in self.to_include:
-            if add_none:
-                filtered += [None] * add_none
-            if to_expand:
-                if filtered:
-                    filtered += children[i].children
-                else:   ##
-
-                    filtered = children[i].children
-            else:
-                filtered.append(children[i])
-
-        if self.append_none:
-            filtered += [None] * self.append_none
-
-        return self.node_builder(filtered)
-
-
-class ChildFilterLALR_NoPlaceholders(ChildFilter):
-    #--
-    def __init__(self, to_include, node_builder):
-        self.node_builder = node_builder
-        self.to_include = to_include
-
-    def __call__(self, children):
-        filtered = []
-        for i, to_expand in self.to_include:
-            if to_expand:
-                if filtered:
-                    filtered += children[i].children
-                else:   ##
-
-                    filtered = children[i].children
-            else:
-                filtered.append(children[i])
-        return self.node_builder(filtered)
-
-
-def _should_expand(sym):
-    return not sym.is_term and sym.name.startswith('_')
-
-
-def maybe_create_child_filter(expansion, keep_all_tokens, ambiguous, _empty_indices: List[bool]):
-    ##
-
-    if _empty_indices:
-        assert _empty_indices.count(False) == len(expansion)
-        s = ''.join(str(int(b)) for b in _empty_indices)
-        empty_indices = [len(ones) for ones in s.split('0')]
-        assert len(empty_indices) == len(expansion)+1, (empty_indices, len(expansion))
-    else:
-        empty_indices = [0] * (len(expansion)+1)
-
-    to_include = []
-    nones_to_add = 0
-    for i, sym in enumerate(expansion):
-        nones_to_add += empty_indices[i]
-        if keep_all_tokens or not (sym.is_term and sym.filter_out):
-            to_include.append((i, _should_expand(sym), nones_to_add))
-            nones_to_add = 0
-
-    nones_to_add += empty_indices[len(expansion)]
-
-    if _empty_indices or len(to_include) < len(expansion) or any(to_expand for i, to_expand,_ in to_include):
-        if _empty_indices or ambiguous:
-            return partial(ChildFilter if ambiguous else ChildFilterLALR, to_include, nones_to_add)
-        else:
-            ##
-
-            return partial(ChildFilterLALR_NoPlaceholders, [(i, x) for i,x,_ in to_include])
-
-
-class AmbiguousExpander:
-    #--
-    def __init__(self, to_expand, tree_class, node_builder):
-        self.node_builder = node_builder
-        self.tree_class = tree_class
-        self.to_expand = to_expand
-
-    def __call__(self, children):
-        def _is_ambig_tree(t):
-            return hasattr(t, 'data') and t.data == '_ambig'
-
-        ##
-
-        ##
-
-        ##
-
-        ##
-
-        ambiguous = []
-        for i, child in enumerate(children):
-            if _is_ambig_tree(child):
-                if i in self.to_expand:
-                    ambiguous.append(i)
-
-                child.expand_kids_by_data('_ambig')
-
-        if not ambiguous:
-            return self.node_builder(children)
-
-        expand = [child.children if i in ambiguous else (child,) for i, child in enumerate(children)]
-        return self.tree_class('_ambig', [self.node_builder(list(f)) for f in product(*expand)])
-
-
-def maybe_create_ambiguous_expander(tree_class, expansion, keep_all_tokens):
-    to_expand = [i for i, sym in enumerate(expansion)
-                 if keep_all_tokens or ((not (sym.is_term and sym.filter_out)) and _should_expand(sym))]
-    if to_expand:
-        return partial(AmbiguousExpander, to_expand, tree_class)
-
-
-class AmbiguousIntermediateExpander:
-    #--
-
-    def __init__(self, tree_class, node_builder):
-        self.node_builder = node_builder
-        self.tree_class = tree_class
-
-    def __call__(self, children):
-        def _is_iambig_tree(child):
-            return hasattr(child, 'data') and child.data == '_iambig'
-
-        def _collapse_iambig(children):
-            #--
-
-            ##
-
-            ##
-
-            if children and _is_iambig_tree(children[0]):
-                iambig_node = children[0]
-                result = []
-                for grandchild in iambig_node.children:
-                    collapsed = _collapse_iambig(grandchild.children)
-                    if collapsed:
-                        for child in collapsed:
-                            child.children += children[1:]
-                        result += collapsed
-                    else:
-                        new_tree = self.tree_class('_inter', grandchild.children + children[1:])
-                        result.append(new_tree)
-                return result
-
-        collapsed = _collapse_iambig(children)
-        if collapsed:
-            processed_nodes = [self.node_builder(c.children) for c in collapsed]
-            return self.tree_class('_ambig', processed_nodes)
-
-        return self.node_builder(children)
-
-
-
-def inplace_transformer(func):
-    @wraps(func)
-    def f(children):
-        ##
-
-        tree = Tree(func.__name__, children)
-        return func(tree)
-    return f
-
-
-def apply_visit_wrapper(func, name, wrapper):
-    if wrapper is _vargs_meta or wrapper is _vargs_meta_inline:
-        raise NotImplementedError("Meta args not supported for internal transformer")
-
-    @wraps(func)
-    def f(children):
-        return wrapper(func, name, children, None)
-    return f
-
-
-class ParseTreeBuilder:
-    def __init__(self, rules, tree_class, propagate_positions=False, ambiguous=False, maybe_placeholders=False):
-        self.tree_class = tree_class
-        self.propagate_positions = propagate_positions
-        self.ambiguous = ambiguous
-        self.maybe_placeholders = maybe_placeholders
-
-        self.rule_builders = list(self._init_builders(rules))
-
-    def _init_builders(self, rules):
-        propagate_positions = make_propagate_positions(self.propagate_positions)
-
-        for rule in rules:
-            options = rule.options
-            keep_all_tokens = options.keep_all_tokens
-            expand_single_child = options.expand1
-
-            wrapper_chain = list(filter(None, [
-                (expand_single_child and not rule.alias) and ExpandSingleChild,
-                maybe_create_child_filter(rule.expansion, keep_all_tokens, self.ambiguous, options.empty_indices if self.maybe_placeholders else None),
-                propagate_positions,
-                self.ambiguous and maybe_create_ambiguous_expander(self.tree_class, rule.expansion, keep_all_tokens),
-                self.ambiguous and partial(AmbiguousIntermediateExpander, self.tree_class)
-            ]))
-
-            yield rule, wrapper_chain
-
-    def create_callback(self, transformer=None):
-        callbacks = {}
-
-        default_handler = getattr(transformer, '__default__', None)
-        if default_handler:
-            def default_callback(data, children):
-                return default_handler(data, children, None)
-        else:
-            default_callback = self.tree_class
-
-        for rule, wrapper_chain in self.rule_builders:
-
-            user_callback_name = rule.alias or rule.options.template_source or rule.origin.name
-            try:
-                f = getattr(transformer, user_callback_name)
-                wrapper = getattr(f, 'visit_wrapper', None)
-                if wrapper is not None:
-                    f = apply_visit_wrapper(f, user_callback_name, wrapper)
-                elif isinstance(transformer, Transformer_InPlace):
-                    f = inplace_transformer(f)
-            except AttributeError:
-                f = partial(default_callback, user_callback_name)
-
-            for w in wrapper_chain:
-                f = w(f)
-
-            if rule in callbacks:
-                raise GrammarError("Rule '%s' already exists" % (rule,))
-
-            callbacks[rule] = f
-
-        return callbacks
-
-
-
-class LALR_Parser(Serialize):
-    def __init__(self, parser_conf, debug=False, strict=False):
-        analysis = LALR_Analyzer(parser_conf, debug=debug, strict=strict)
-        analysis.compute_lalr()
-        callbacks = parser_conf.callbacks
-
-        self._parse_table = analysis.parse_table
-        self.parser_conf = parser_conf
-        self.parser = _Parser(analysis.parse_table, callbacks, debug)
-
-    @classmethod
-    def deserialize(cls, data, memo, callbacks, debug=False):
-        inst = cls.__new__(cls)
-        inst._parse_table = IntParseTable.deserialize(data, memo)
-        inst.parser = _Parser(inst._parse_table, callbacks, debug)
-        return inst
-
-    def serialize(self, memo: Any = None) -> Dict[str, Any]:
-        return self._parse_table.serialize(memo)
-
-    def parse_interactive(self, lexer, start):
-        return self.parser.parse(lexer, start, start_interactive=True)
-
-    def parse(self, lexer, start, on_error=None):
-        try:
-            return self.parser.parse(lexer, start)
-        except UnexpectedInput as e:
-            if on_error is None:
-                raise
-
-            while True:
-                if isinstance(e, UnexpectedCharacters):
-                    s = e.interactive_parser.lexer_thread.state
-                    p = s.line_ctr.char_pos
-
-                if not on_error(e):
-                    raise e
-
-                if isinstance(e, UnexpectedCharacters):
-                    ##
-
-                    if p == s.line_ctr.char_pos:
-                        s.line_ctr.feed(s.text[p:p+1])
-
-                try:
-                    return e.interactive_parser.resume_parse()
-                except UnexpectedToken as e2:
-                    if (isinstance(e, UnexpectedToken)
-                        and e.token.type == e2.token.type == '$END'
-                        and e.interactive_parser == e2.interactive_parser):
-                        ##
-
-                        raise e2
-                    e = e2
-                except UnexpectedCharacters as e2:
-                    e = e2
-
-
-class ParseConf:
-    __slots__ = 'parse_table', 'callbacks', 'start', 'start_state', 'end_state', 'states'
-
-    def __init__(self, parse_table, callbacks, start):
-        self.parse_table = parse_table
-
-        self.start_state = self.parse_table.start_states[start]
-        self.end_state = self.parse_table.end_states[start]
-        self.states = self.parse_table.states
-
-        self.callbacks = callbacks
-        self.start = start
-
-
-class ParserState:
-    __slots__ = 'parse_conf', 'lexer', 'state_stack', 'value_stack'
-
-    def __init__(self, parse_conf, lexer, state_stack=None, value_stack=None):
-        self.parse_conf = parse_conf
-        self.lexer = lexer
-        self.state_stack = state_stack or [self.parse_conf.start_state]
-        self.value_stack = value_stack or []
-
-    @property
-    def position(self):
-        return self.state_stack[-1]
-
-    ##
-
-    def __eq__(self, other):
-        if not isinstance(other, ParserState):
-            return NotImplemented
-        return len(self.state_stack) == len(other.state_stack) and self.position == other.position
-
-    def __copy__(self):
-        return type(self)(
-            self.parse_conf,
-            self.lexer, ##
-
-            copy(self.state_stack),
-            deepcopy(self.value_stack),
-        )
-
-    def copy(self):
-        return copy(self)
-
-    def feed_token(self, token, is_end=False):
-        state_stack = self.state_stack
-        value_stack = self.value_stack
-        states = self.parse_conf.states
-        end_state = self.parse_conf.end_state
-        callbacks = self.parse_conf.callbacks
-
-        while True:
-            state = state_stack[-1]
-            try:
-                action, arg = states[state][token.type]
-            except KeyError:
-                expected = {s for s in states[state].keys() if s.isupper()}
-                raise UnexpectedToken(token, expected, state=self, interactive_parser=None)
-
-            assert arg != end_state
-
-            if action is Shift:
-                ##
-
-                assert not is_end
-                state_stack.append(arg)
-                value_stack.append(token if token.type not in callbacks else callbacks[token.type](token))
-                return
-            else:
-                ##
-
-                rule = arg
-                size = len(rule.expansion)
-                if size:
-                    s = value_stack[-size:]
-                    del state_stack[-size:]
-                    del value_stack[-size:]
-                else:
-                    s = []
-
-                value = callbacks[rule](s)
-
-                _action, new_state = states[state_stack[-1]][rule.origin.name]
-                assert _action is Shift
-                state_stack.append(new_state)
-                value_stack.append(value)
-
-                if is_end and state_stack[-1] == end_state:
-                    return value_stack[-1]
-
-class _Parser:
-    def __init__(self, parse_table, callbacks, debug=False):
-        self.parse_table = parse_table
-        self.callbacks = callbacks
-        self.debug = debug
-
-    def parse(self, lexer, start, value_stack=None, state_stack=None, start_interactive=False):
-        parse_conf = ParseConf(self.parse_table, self.callbacks, start)
-        parser_state = ParserState(parse_conf, lexer, state_stack, value_stack)
-        if start_interactive:
-            return InteractiveParser(self, parser_state, parser_state.lexer)
-        return self.parse_from_state(parser_state)
-
-
-    def parse_from_state(self, state, last_token=None):
-        #--
-        try:
-            token = last_token
-            for token in state.lexer.lex(state):
-                state.feed_token(token)
-
-            end_token = Token.new_borrow_pos('$END', '', token) if token else Token('$END', '', 0, 1, 1)
-            return state.feed_token(end_token, True)
-        except UnexpectedInput as e:
-            try:
-                e.interactive_parser = InteractiveParser(self, state, state.lexer)
-            except NameError:
-                pass
-            raise e
-        except Exception as e:
-            if self.debug:
-                print("")
-                print("STATE STACK DUMP")
-                print("----------------")
-                for i, s in enumerate(state.state_stack):
-                    print('%d)' % i , s)
-                print("")
-
-            raise
-
-
-class Action:
-    def __init__(self, name):
-        self.name = name
-    def __str__(self):
-        return self.name
-    def __repr__(self):
-        return str(self)
-
-Shift = Action('Shift')
-Reduce = Action('Reduce')
-
-
-class ParseTable:
-    def __init__(self, states, start_states, end_states):
-        self.states = states
-        self.start_states = start_states
-        self.end_states = end_states
-
-    def serialize(self, memo):
-        tokens = Enumerator()
-
-        states = {
-            state: {tokens.get(token): ((1, arg.serialize(memo)) if action is Reduce else (0, arg))
-                    for token, (action, arg) in actions.items()}
-            for state, actions in self.states.items()
-        }
-
-        return {
-            'tokens': tokens.reversed(),
-            'states': states,
-            'start_states': self.start_states,
-            'end_states': self.end_states,
-        }
-
-    @classmethod
-    def deserialize(cls, data, memo):
-        tokens = data['tokens']
-        states = {
-            state: {tokens[token]: ((Reduce, Rule.deserialize(arg, memo)) if action==1 else (Shift, arg))
-                    for token, (action, arg) in actions.items()}
-            for state, actions in data['states'].items()
-        }
-        return cls(states, data['start_states'], data['end_states'])
-
-
-class IntParseTable(ParseTable):
-
-    @classmethod
-    def from_ParseTable(cls, parse_table):
-        enum = list(parse_table.states)
-        state_to_idx = {s:i for i,s in enumerate(enum)}
-        int_states = {}
-
-        for s, la in parse_table.states.items():
-            la = {k:(v[0], state_to_idx[v[1]]) if v[0] is Shift else v
-                  for k,v in la.items()}
-            int_states[ state_to_idx[s] ] = la
-
-
-        start_states = {start:state_to_idx[s] for start, s in parse_table.start_states.items()}
-        end_states = {start:state_to_idx[s] for start, s in parse_table.end_states.items()}
-        return cls(int_states, start_states, end_states)
-
-
-
-def _wrap_lexer(lexer_class):
-    future_interface = getattr(lexer_class, '__future_interface__', False)
-    if future_interface:
-        return lexer_class
-    else:
-        class CustomLexerWrapper(Lexer):
-            def __init__(self, lexer_conf):
-                self.lexer = lexer_class(lexer_conf)
-            def lex(self, lexer_state, parser_state):
-                return self.lexer.lex(lexer_state.text)
-        return CustomLexerWrapper
-
-
-def _deserialize_parsing_frontend(data, memo, lexer_conf, callbacks, options):
-    parser_conf = ParserConf.deserialize(data['parser_conf'], memo)
-    cls = (options and options._plugins.get('LALR_Parser')) or LALR_Parser
-    parser = cls.deserialize(data['parser'], memo, callbacks, options.debug)
-    parser_conf.callbacks = callbacks
-    return ParsingFrontend(lexer_conf, parser_conf, options, parser=parser)
-
-
-_parser_creators: 'Dict[str, Callable[[LexerConf, Any, Any], Any]]' = {}
-
-
-class ParsingFrontend(Serialize):
-    __serialize_fields__ = 'lexer_conf', 'parser_conf', 'parser'
-
-    lexer_conf: LexerConf
-    parser_conf: ParserConf
-    options: Any
-
-    def __init__(self, lexer_conf: LexerConf, parser_conf: ParserConf, options, parser=None):
-        self.parser_conf = parser_conf
-        self.lexer_conf = lexer_conf
-        self.options = options
-
-        ##
-
-        if parser:  ##
-
-            self.parser = parser
-        else:
-            create_parser = _parser_creators.get(parser_conf.parser_type)
-            assert create_parser is not None, "{} is not supported in standalone mode".format(
-                    parser_conf.parser_type
-                )
-            self.parser = create_parser(lexer_conf, parser_conf, options)
-
-        ##
-
-        lexer_type = lexer_conf.lexer_type
-        self.skip_lexer = False
-        if lexer_type in ('dynamic', 'dynamic_complete'):
-            assert lexer_conf.postlex is None
-            self.skip_lexer = True
-            return
-
-        if isinstance(lexer_type, type):
-            assert issubclass(lexer_type, Lexer)
-            self.lexer = _wrap_lexer(lexer_type)(lexer_conf)
-        elif isinstance(lexer_type, str):
-            create_lexer = {
-                'basic': create_basic_lexer,
-                'contextual': create_contextual_lexer,
-            }[lexer_type]
-            self.lexer = create_lexer(lexer_conf, self.parser, lexer_conf.postlex, options)
-        else:
-            raise TypeError("Bad value for lexer_type: {lexer_type}")
-
-        if lexer_conf.postlex:
-            self.lexer = PostLexConnector(self.lexer, lexer_conf.postlex)
-
-    def _verify_start(self, start=None):
-        if start is None:
-            start_decls = self.parser_conf.start
-            if len(start_decls) > 1:
-                raise ConfigurationError("Lark initialized with more than 1 possible start rule. Must specify which start rule to parse", start_decls)
-            start ,= start_decls
-        elif start not in self.parser_conf.start:
-            raise ConfigurationError("Unknown start rule %s. Must be one of %r" % (start, self.parser_conf.start))
-        return start
-
-    def _make_lexer_thread(self, text: str):
-        cls = (self.options and self.options._plugins.get('LexerThread')) or LexerThread
-        return text if self.skip_lexer else cls.from_text(self.lexer, text)
-
-    def parse(self, text: str, start=None, on_error=None):
-        chosen_start = self._verify_start(start)
-        kw = {} if on_error is None else {'on_error': on_error}
-        stream = self._make_lexer_thread(text)
-        return self.parser.parse(stream, chosen_start, **kw)
-
-    def parse_interactive(self, text: Optional[str]=None, start=None):
-        ##
-
-        ##
-
-        chosen_start = self._verify_start(start)
-        if self.parser_conf.parser_type != 'lalr':
-            raise ConfigurationError("parse_interactive() currently only works with parser='lalr' ")
-        stream = self._make_lexer_thread(text)  ##
-
-        return self.parser.parse_interactive(stream, chosen_start)
-
-
-def _validate_frontend_args(parser, lexer) -> None:
-    assert_config(parser, ('lalr', 'earley', 'cyk'))
-    if not isinstance(lexer, type):     ##
-
-        expected = {
-            'lalr': ('basic', 'contextual'),
-            'earley': ('basic', 'dynamic', 'dynamic_complete'),
-            'cyk': ('basic', ),
-         }[parser]
-        assert_config(lexer, expected, 'Parser %r does not support lexer %%r, expected one of %%s' % parser)
-
-
-def _get_lexer_callbacks(transformer, terminals):
-    result = {}
-    for terminal in terminals:
-        callback = getattr(transformer, terminal.name, None)
-        if callback is not None:
-            result[terminal.name] = callback
-    return result
-
-class PostLexConnector:
-    def __init__(self, lexer, postlexer):
-        self.lexer = lexer
-        self.postlexer = postlexer
-
-    def lex(self, lexer_state, parser_state):
-        i = self.lexer.lex(lexer_state, parser_state)
-        return self.postlexer.process(i)
-
-
-
-def create_basic_lexer(lexer_conf, parser, postlex, options) -> BasicLexer:
-    cls = (options and options._plugins.get('BasicLexer')) or BasicLexer
-    return cls(lexer_conf)
-
-def create_contextual_lexer(lexer_conf: LexerConf, parser, postlex, options) -> ContextualLexer:
-    cls = (options and options._plugins.get('ContextualLexer')) or ContextualLexer
-    states: Dict[str, Collection[str]] = {idx:list(t.keys()) for idx, t in parser._parse_table.states.items()}
-    always_accept: Collection[str] = postlex.always_accept if postlex else ()
-    return cls(lexer_conf, states, always_accept=always_accept)
-
-def create_lalr_parser(lexer_conf: LexerConf, parser_conf: ParserConf, options=None) -> LALR_Parser:
-    debug = options.debug if options else False
-    strict = options.strict if options else False
-    cls = (options and options._plugins.get('LALR_Parser')) or LALR_Parser
-    return cls(parser_conf, debug=debug, strict=strict)
-
-_parser_creators['lalr'] = create_lalr_parser
-
-
-
-
-class PostLex(ABC):
-    @abstractmethod
-    def process(self, stream: Iterator[Token]) -> Iterator[Token]:
-        return stream
-
-    always_accept: Iterable[str] = ()
-
-class LarkOptions(Serialize):
-    #--
-
-    start: List[str]
-    debug: bool
-    strict: bool
-    transformer: 'Optional[Transformer]'
-    propagate_positions: Union[bool, str]
-    maybe_placeholders: bool
-    cache: Union[bool, str]
-    regex: bool
-    g_regex_flags: int
-    keep_all_tokens: bool
-    tree_class: Any
-    parser: _ParserArgType
-    lexer: _LexerArgType
-    ambiguity: 'Literal["auto", "resolve", "explicit", "forest"]'
-    postlex: Optional[PostLex]
-    priority: 'Optional[Literal["auto", "normal", "invert"]]'
-    lexer_callbacks: Dict[str, Callable[[Token], Token]]
-    use_bytes: bool
-    edit_terminals: Optional[Callable[[TerminalDef], TerminalDef]]
-    import_paths: 'List[Union[str, Callable[[Union[None, str, PackageResource], str], Tuple[str, str]]]]'
-    source_path: Optional[str]
-
-    OPTIONS_DOC = """
-    **===  General Options  ===**
-
-    start
-            The start symbol. Either a string, or a list of strings for multiple possible starts (Default: "start")
-    debug
-            Display debug information and extra warnings. Use only when debugging (Default: ``False``)
-            When used with Earley, it generates a forest graph as "sppf.png", if 'dot' is installed.
-    strict
-            Throw an exception on any potential ambiguity, including shift/reduce conflicts, and regex collisions.
-    transformer
-            Applies the transformer to every parse tree (equivalent to applying it after the parse, but faster)
-    propagate_positions
-            Propagates positional attributes into the 'meta' attribute of all tree branches.
-            Sets attributes: (line, column, end_line, end_column, start_pos, end_pos,
-                              container_line, container_column, container_end_line, container_end_column)
-            Accepts ``False``, ``True``, or a callable, which will filter which nodes to ignore when propagating.
-    maybe_placeholders
-            When ``True``, the ``[]`` operator returns ``None`` when not matched.
-            When ``False``,  ``[]`` behaves like the ``?`` operator, and returns no value at all.
-            (default= ``True``)
-    cache
-            Cache the results of the Lark grammar analysis, for x2 to x3 faster loading. LALR only for now.
-
-            - When ``False``, does nothing (default)
-            - When ``True``, caches to a temporary file in the local directory
-            - When given a string, caches to the path pointed by the string
-    regex
-            When True, uses the ``regex`` module instead of the stdlib ``re``.
-    g_regex_flags
-            Flags that are applied to all terminals (both regex and strings)
-    keep_all_tokens
-            Prevent the tree builder from automagically removing "punctuation" tokens (Default: ``False``)
-    tree_class
-            Lark will produce trees comprised of instances of this class instead of the default ``lark.Tree``.
-
-    **=== Algorithm Options ===**
-
-    parser
-            Decides which parser engine to use. Accepts "earley" or "lalr". (Default: "earley").
-            (there is also a "cyk" option for legacy)
-    lexer
-            Decides whether or not to use a lexer stage
-
-            - "auto" (default): Choose for me based on the parser
-            - "basic": Use a basic lexer
-            - "contextual": Stronger lexer (only works with parser="lalr")
-            - "dynamic": Flexible and powerful (only with parser="earley")
-            - "dynamic_complete": Same as dynamic, but tries *every* variation of tokenizing possible.
-    ambiguity
-            Decides how to handle ambiguity in the parse. Only relevant if parser="earley"
-
-            - "resolve": The parser will automatically choose the simplest derivation
-              (it chooses consistently: greedy for tokens, non-greedy for rules)
-            - "explicit": The parser will return all derivations wrapped in "_ambig" tree nodes (i.e. a forest).
-            - "forest": The parser will return the root of the shared packed parse forest.
-
-    **=== Misc. / Domain Specific Options ===**
-
-    postlex
-            Lexer post-processing (Default: ``None``) Only works with the basic and contextual lexers.
-    priority
-            How priorities should be evaluated - "auto", ``None``, "normal", "invert" (Default: "auto")
-    lexer_callbacks
-            Dictionary of callbacks for the lexer. May alter tokens during lexing. Use with caution.
-    use_bytes
-            Accept an input of type ``bytes`` instead of ``str``.
-    edit_terminals
-            A callback for editing the terminals before parse.
-    import_paths
-            A List of either paths or loader functions to specify from where grammars are imported
-    source_path
-            Override the source of from where the grammar was loaded. Useful for relative imports and unconventional grammar loading
-    **=== End of Options ===**
-    """
-    if __doc__:
-        __doc__ += OPTIONS_DOC
-
-
-    ##
-
-    ##
-
-    ##
-
-    ##
-
-    ##
-
-    ##
-
-    _defaults: Dict[str, Any] = {
-        'debug': False,
-        'strict': False,
-        'keep_all_tokens': False,
-        'tree_class': None,
-        'cache': False,
-        'postlex': None,
-        'parser': 'earley',
-        'lexer': 'auto',
-        'transformer': None,
-        'start': 'start',
-        'priority': 'auto',
-        'ambiguity': 'auto',
-        'regex': False,
-        'propagate_positions': False,
-        'lexer_callbacks': {},
-        'maybe_placeholders': True,
-        'edit_terminals': None,
-        'g_regex_flags': 0,
-        'use_bytes': False,
-        'import_paths': [],
-        'source_path': None,
-        '_plugins': {},
-    }
-
-    def __init__(self, options_dict: Dict[str, Any]) -> None:
-        o = dict(options_dict)
-
-        options = {}
-        for name, default in self._defaults.items():
-            if name in o:
-                value = o.pop(name)
-                if isinstance(default, bool) and name not in ('cache', 'use_bytes', 'propagate_positions'):
-                    value = bool(value)
-            else:
-                value = default
-
-            options[name] = value
-
-        if isinstance(options['start'], str):
-            options['start'] = [options['start']]
-
-        self.__dict__['options'] = options
-
-
-        assert_config(self.parser, ('earley', 'lalr', 'cyk', None))
-
-        if self.parser == 'earley' and self.transformer:
-            raise ConfigurationError('Cannot specify an embedded transformer when using the Earley algorithm. '
-                             'Please use your transformer on the resulting parse tree, or use a different algorithm (i.e. LALR)')
-
-        if o:
-            raise ConfigurationError("Unknown options: %s" % o.keys())
-
-    def __getattr__(self, name: str) -> Any:
-        try:
-            return self.__dict__['options'][name]
-        except KeyError as e:
-            raise AttributeError(e)
-
-    def __setattr__(self, name: str, value: str) -> None:
-        assert_config(name, self.options.keys(), "%r isn't a valid option. Expected one of: %s")
-        self.options[name] = value
-
-    def serialize(self, memo = None) -> Dict[str, Any]:
-        return self.options
-
-    @classmethod
-    def deserialize(cls, data: Dict[str, Any], memo: Dict[int, Union[TerminalDef, Rule]]) -> "LarkOptions":
-        return cls(data)
-
-
-##
-
-##
-
-_LOAD_ALLOWED_OPTIONS = {'postlex', 'transformer', 'lexer_callbacks', 'use_bytes', 'debug', 'g_regex_flags', 'regex', 'propagate_positions', 'tree_class', '_plugins'}
-
-_VALID_PRIORITY_OPTIONS = ('auto', 'normal', 'invert', None)
-_VALID_AMBIGUITY_OPTIONS = ('auto', 'resolve', 'explicit', 'forest')
-
-
-_T = TypeVar('_T', bound="Lark")
-
-class Lark(Serialize):
-    #--
-
-    source_path: str
-    source_grammar: str
-    grammar: 'Grammar'
-    options: LarkOptions
-    lexer: Lexer
-    parser: 'ParsingFrontend'
-    terminals: Collection[TerminalDef]
-
-    def __init__(self, grammar: 'Union[Grammar, str, IO[str]]', **options) -> None:
-        self.options = LarkOptions(options)
-        re_module: types.ModuleType
-
-        ##
-
-        use_regex = self.options.regex
-        if use_regex:
-            if _has_regex:
-                re_module = regex
-            else:
-                raise ImportError('`regex` module must be installed if calling `Lark(regex=True)`.')
-        else:
-            re_module = re
-
-        ##
-
-        if self.options.source_path is None:
-            try:
-                self.source_path = grammar.name  ##
-
-            except AttributeError:
-                self.source_path = '<string>'
-        else:
-            self.source_path = self.options.source_path
-
-        ##
-
-        try:
-            read = grammar.read  ##
-
-        except AttributeError:
-            pass
-        else:
-            grammar = read()
-
-        cache_fn = None
-        cache_sha256 = None
-        if isinstance(grammar, str):
-            self.source_grammar = grammar
-            if self.options.use_bytes:
-                if not isascii(grammar):
-                    raise ConfigurationError("Grammar must be ascii only, when use_bytes=True")
-
-            if self.options.cache:
-                if self.options.parser != 'lalr':
-                    raise ConfigurationError("cache only works with parser='lalr' for now")
-
-                unhashable = ('transformer', 'postlex', 'lexer_callbacks', 'edit_terminals', '_plugins')
-                options_str = ''.join(k+str(v) for k, v in options.items() if k not in unhashable)
-                from . import __version__
-                s = grammar + options_str + __version__ + str(sys.version_info[:2])
-                cache_sha256 = sha256_digest(s)
-
-                if isinstance(self.options.cache, str):
-                    cache_fn = self.options.cache
-                else:
-                    if self.options.cache is not True:
-                        raise ConfigurationError("cache argument must be bool or str")
-
-                    try:
-                        username = getpass.getuser()
-                    except Exception:
-                        ##
-
-                        ##
-
-                        ##
-
-                        username = "unknown"
-
-                    cache_fn = tempfile.gettempdir() + "/.lark_cache_%s_%s_%s_%s.tmp" % (username, cache_sha256, *sys.version_info[:2])
-
-                old_options = self.options
-                try:
-                    with FS.open(cache_fn, 'rb') as f:
-                        logger.debug('Loading grammar from cache: %s', cache_fn)
-                        ##
-
-                        for name in (set(options) - _LOAD_ALLOWED_OPTIONS):
-                            del options[name]
-                        file_sha256 = f.readline().rstrip(b'\n')
-                        cached_used_files = pickle.load(f)
-                        if file_sha256 == cache_sha256.encode('utf8') and verify_used_files(cached_used_files):
-                            cached_parser_data = pickle.load(f)
-                            self._load(cached_parser_data, **options)
-                            return
-                except FileNotFoundError:
-                    ##
-
-                    pass
-                except Exception: ##
-
-                    logger.exception("Failed to load Lark from cache: %r. We will try to carry on.", cache_fn)
-
-                    ##
-
-                    ##
-
-                    self.options = old_options
-
-
-            ##
-
-            self.grammar, used_files = load_grammar(grammar, self.source_path, self.options.import_paths, self.options.keep_all_tokens)
-        else:
-            assert isinstance(grammar, Grammar)
-            self.grammar = grammar
-
-
-        if self.options.lexer == 'auto':
-            if self.options.parser == 'lalr':
-                self.options.lexer = 'contextual'
-            elif self.options.parser == 'earley':
-                if self.options.postlex is not None:
-                    logger.info("postlex can't be used with the dynamic lexer, so we use 'basic' instead. "
-                                "Consider using lalr with contextual instead of earley")
-                    self.options.lexer = 'basic'
-                else:
-                    self.options.lexer = 'dynamic'
-            elif self.options.parser == 'cyk':
-                self.options.lexer = 'basic'
-            else:
-                assert False, self.options.parser
-        lexer = self.options.lexer
-        if isinstance(lexer, type):
-            assert issubclass(lexer, Lexer)     ##
-
-        else:
-            assert_config(lexer, ('basic', 'contextual', 'dynamic', 'dynamic_complete'))
-            if self.options.postlex is not None and 'dynamic' in lexer:
-                raise ConfigurationError("Can't use postlex with a dynamic lexer. Use basic or contextual instead")
-
-        if self.options.ambiguity == 'auto':
-            if self.options.parser == 'earley':
-                self.options.ambiguity = 'resolve'
-        else:
-            assert_config(self.options.parser, ('earley', 'cyk'), "%r doesn't support disambiguation. Use one of these parsers instead: %s")
-
-        if self.options.priority == 'auto':
-            self.options.priority = 'normal'
-
-        if self.options.priority not in _VALID_PRIORITY_OPTIONS:
-            raise ConfigurationError("invalid priority option: %r. Must be one of %r" % (self.options.priority, _VALID_PRIORITY_OPTIONS))
-        if self.options.ambiguity not in _VALID_AMBIGUITY_OPTIONS:
-            raise ConfigurationError("invalid ambiguity option: %r. Must be one of %r" % (self.options.ambiguity, _VALID_AMBIGUITY_OPTIONS))
-
-        if self.options.parser is None:
-            terminals_to_keep = '*'
-        elif self.options.postlex is not None:
-            terminals_to_keep = set(self.options.postlex.always_accept)
-        else:
-            terminals_to_keep = set()
-
-        ##
-
-        self.terminals, self.rules, self.ignore_tokens = self.grammar.compile(self.options.start, terminals_to_keep)
-
-        if self.options.edit_terminals:
-            for t in self.terminals:
-                self.options.edit_terminals(t)
-
-        self._terminals_dict = {t.name: t for t in self.terminals}
-
-        ##
-
-        if self.options.priority == 'invert':
-            for rule in self.rules:
-                if rule.options.priority is not None:
-                    rule.options.priority = -rule.options.priority
-            for term in self.terminals:
-                term.priority = -term.priority
-        ##
-
-        ##
-
-        ##
-
-        elif self.options.priority is None:
-            for rule in self.rules:
-                if rule.options.priority is not None:
-                    rule.options.priority = None
-            for term in self.terminals:
-                term.priority = 0
-
-        ##
-
-        self.lexer_conf = LexerConf(
-                self.terminals, re_module, self.ignore_tokens, self.options.postlex,
-                self.options.lexer_callbacks, self.options.g_regex_flags, use_bytes=self.options.use_bytes, strict=self.options.strict
-            )
-
-        if self.options.parser:
-            self.parser = self._build_parser()
-        elif lexer:
-            self.lexer = self._build_lexer()
-
-        if cache_fn:
-            logger.debug('Saving grammar to cache: %s', cache_fn)
-            try:
-                with FS.open(cache_fn, 'wb') as f:
-                    assert cache_sha256 is not None
-                    f.write(cache_sha256.encode('utf8') + b'\n')
-                    pickle.dump(used_files, f)
-                    self.save(f, _LOAD_ALLOWED_OPTIONS)
-            except IOError as e:
-                logger.exception("Failed to save Lark to cache: %r.", cache_fn, e)
-
-    if __doc__:
-        __doc__ += "\n\n" + LarkOptions.OPTIONS_DOC
-
-    __serialize_fields__ = 'parser', 'rules', 'options'
-
-    def _build_lexer(self, dont_ignore: bool=False) -> BasicLexer:
-        lexer_conf = self.lexer_conf
-        if dont_ignore:
-            from copy import copy
-            lexer_conf = copy(lexer_conf)
-            lexer_conf.ignore = ()
-        return BasicLexer(lexer_conf)
-
-    def _prepare_callbacks(self) -> None:
-        self._callbacks = {}
-        ##
-
-        if self.options.ambiguity != 'forest':
-            self._parse_tree_builder = ParseTreeBuilder(
-                    self.rules,
-                    self.options.tree_class or Tree,
-                    self.options.propagate_positions,
-                    self.options.parser != 'lalr' and self.options.ambiguity == 'explicit',
-                    self.options.maybe_placeholders
-                )
-            self._callbacks = self._parse_tree_builder.create_callback(self.options.transformer)
-        self._callbacks.update(_get_lexer_callbacks(self.options.transformer, self.terminals))
-
-    def _build_parser(self) -> "ParsingFrontend":
-        self._prepare_callbacks()
-        _validate_frontend_args(self.options.parser, self.options.lexer)
-        parser_conf = ParserConf(self.rules, self._callbacks, self.options.start)
-        return _construct_parsing_frontend(
-            self.options.parser,
-            self.options.lexer,
-            self.lexer_conf,
-            parser_conf,
-            options=self.options
-        )
-
-    def save(self, f, exclude_options: Collection[str] = ()) -> None:
-        #--
-        data, m = self.memo_serialize([TerminalDef, Rule])
-        if exclude_options:
-            data["options"] = {n: v for n, v in data["options"].items() if n not in exclude_options}
-        pickle.dump({'data': data, 'memo': m}, f, protocol=pickle.HIGHEST_PROTOCOL)
-
-    @classmethod
-    def load(cls: Type[_T], f) -> _T:
-        #--
-        inst = cls.__new__(cls)
-        return inst._load(f)
-
-    def _deserialize_lexer_conf(self, data: Dict[str, Any], memo: Dict[int, Union[TerminalDef, Rule]], options: LarkOptions) -> LexerConf:
-        lexer_conf = LexerConf.deserialize(data['lexer_conf'], memo)
-        lexer_conf.callbacks = options.lexer_callbacks or {}
-        lexer_conf.re_module = regex if options.regex else re
-        lexer_conf.use_bytes = options.use_bytes
-        lexer_conf.g_regex_flags = options.g_regex_flags
-        lexer_conf.skip_validation = True
-        lexer_conf.postlex = options.postlex
-        return lexer_conf
-
-    def _load(self: _T, f: Any, **kwargs) -> _T:
-        if isinstance(f, dict):
-            d = f
-        else:
-            d = pickle.load(f)
-        memo_json = d['memo']
-        data = d['data']
-
-        assert memo_json
-        memo = SerializeMemoizer.deserialize(memo_json, {'Rule': Rule, 'TerminalDef': TerminalDef}, {})
-        options = dict(data['options'])
-        if (set(kwargs) - _LOAD_ALLOWED_OPTIONS) & set(LarkOptions._defaults):
-            raise ConfigurationError("Some options are not allowed when loading a Parser: {}"
-                             .format(set(kwargs) - _LOAD_ALLOWED_OPTIONS))
-        options.update(kwargs)
-        self.options = LarkOptions.deserialize(options, memo)
-        self.rules = [Rule.deserialize(r, memo) for r in data['rules']]
-        self.source_path = '<deserialized>'
-        _validate_frontend_args(self.options.parser, self.options.lexer)
-        self.lexer_conf = self._deserialize_lexer_conf(data['parser'], memo, self.options)
-        self.terminals = self.lexer_conf.terminals
-        self._prepare_callbacks()
-        self._terminals_dict = {t.name: t for t in self.terminals}
-        self.parser = _deserialize_parsing_frontend(
-            data['parser'],
-            memo,
-            self.lexer_conf,
-            self._callbacks,
-            self.options,  ##
-
-        )
-        return self
-
-    @classmethod
-    def _load_from_dict(cls, data, memo, **kwargs):
-        inst = cls.__new__(cls)
-        return inst._load({'data': data, 'memo': memo}, **kwargs)
-
-    @classmethod
-    def open(cls: Type[_T], grammar_filename: str, rel_to: Optional[str]=None, **options) -> _T:
-        #--
-        if rel_to:
-            basepath = os.path.dirname(rel_to)
-            grammar_filename = os.path.join(basepath, grammar_filename)
-        with open(grammar_filename, encoding='utf8') as f:
-            return cls(f, **options)
-
-    @classmethod
-    def open_from_package(cls: Type[_T], package: str, grammar_path: str, search_paths: 'Sequence[str]'=[""], **options) -> _T:
-        #--
-        package_loader = FromPackageLoader(package, search_paths)
-        full_path, text = package_loader(None, grammar_path)
-        options.setdefault('source_path', full_path)
-        options.setdefault('import_paths', [])
-        options['import_paths'].append(package_loader)
-        return cls(text, **options)
-
-    def __repr__(self):
-        return 'Lark(open(%r), parser=%r, lexer=%r, ...)' % (self.source_path, self.options.parser, self.options.lexer)
-
-
-    def lex(self, text: str, dont_ignore: bool=False) -> Iterator[Token]:
-        #--
-        lexer: Lexer
-        if not hasattr(self, 'lexer') or dont_ignore:
-            lexer = self._build_lexer(dont_ignore)
-        else:
-            lexer = self.lexer
-        lexer_thread = LexerThread.from_text(lexer, text)
-        stream = lexer_thread.lex(None)
-        if self.options.postlex:
-            return self.options.postlex.process(stream)
-        return stream
-
-    def get_terminal(self, name: str) -> TerminalDef:
-        #--
-        return self._terminals_dict[name]
-
-    def parse_interactive(self, text: Optional[str]=None, start: Optional[str]=None) -> 'InteractiveParser':
-        #--
-        return self.parser.parse_interactive(text, start=start)
-
-    def parse(self, text: str, start: Optional[str]=None, on_error: 'Optional[Callable[[UnexpectedInput], bool]]'=None) -> 'ParseTree':
-        #--
-        return self.parser.parse(text, start=start, on_error=on_error)
-
-
-
-
-class DedentError(LarkError):
-    pass
-
-class Indenter(PostLex, ABC):
-    paren_level: int
-    indent_level: List[int]
-
-    def __init__(self) -> None:
-        self.paren_level = 0
-        self.indent_level = [0]
-        assert self.tab_len > 0
-
-    def handle_NL(self, token: Token) -> Iterator[Token]:
-        if self.paren_level > 0:
-            return
-
-        yield token
-
-        indent_str = token.rsplit('\n', 1)[1] ##
-
-        indent = indent_str.count(' ') + indent_str.count('\t') * self.tab_len
-
-        if indent > self.indent_level[-1]:
-            self.indent_level.append(indent)
-            yield Token.new_borrow_pos(self.INDENT_type, indent_str, token)
-        else:
-            while indent < self.indent_level[-1]:
-                self.indent_level.pop()
-                yield Token.new_borrow_pos(self.DEDENT_type, indent_str, token)
-
-            if indent != self.indent_level[-1]:
-                raise DedentError('Unexpected dedent to column %s. Expected dedent to %s' % (indent, self.indent_level[-1]))
-
-    def _process(self, stream):
-        for token in stream:
-            if token.type == self.NL_type:
-                yield from self.handle_NL(token)
-            else:
-                yield token
-
-            if token.type in self.OPEN_PAREN_types:
-                self.paren_level += 1
-            elif token.type in self.CLOSE_PAREN_types:
-                self.paren_level -= 1
-                assert self.paren_level >= 0
-
-        while len(self.indent_level) > 1:
-            self.indent_level.pop()
-            yield Token(self.DEDENT_type, '')
-
-        assert self.indent_level == [0], self.indent_level
-
-    def process(self, stream):
-        self.paren_level = 0
-        self.indent_level = [0]
-        return self._process(stream)
-
-    ##
-
-    @property
-    def always_accept(self):
-        return (self.NL_type,)
-
-    @property
-    @abstractmethod
-    def NL_type(self) -> str:
-        raise NotImplementedError()
-
-    @property
-    @abstractmethod
-    def OPEN_PAREN_types(self) -> List[str]:
-        raise NotImplementedError()
-
-    @property
-    @abstractmethod
-    def CLOSE_PAREN_types(self) -> List[str]:
-        raise NotImplementedError()
-
-    @property
-    @abstractmethod
-    def INDENT_type(self) -> str:
-        raise NotImplementedError()
-
-    @property
-    @abstractmethod
-    def DEDENT_type(self) -> str:
-        raise NotImplementedError()
-
-    @property
-    @abstractmethod
-    def tab_len(self) -> int:
-        raise NotImplementedError()
-
-
-class PythonIndenter(Indenter):
-    NL_type = '_NEWLINE'
-    OPEN_PAREN_types = ['LPAR', 'LSQB', 'LBRACE']
-    CLOSE_PAREN_types = ['RPAR', 'RSQB', 'RBRACE']
-    INDENT_type = '_INDENT'
-    DEDENT_type = '_DEDENT'
-    tab_len = 8
-
-
-import pickle, zlib, base64
-DATA = (
-{'parser': {'lexer_conf': {'terminals': [{'@': 0}, {'@': 1}, {'@': 2}, {'@': 3}, {'@': 4}, {'@': 5}, {'@': 6}, {'@': 7}, {'@': 8}, {'@': 9}, {'@': 10}, {'@': 11}, {'@': 12}, {'@': 13}, {'@': 14}, {'@': 15}, {'@': 16}, {'@': 17}, {'@': 18}, {'@': 19}, {'@': 20}, {'@': 21}, {'@': 22}, {'@': 23}, {'@': 24}, {'@': 25}, {'@': 26}, {'@': 27}, {'@': 28}, {'@': 29}, {'@': 30}, {'@': 31}, {'@': 32}, {'@': 33}, {'@': 34}, {'@': 35}, {'@': 36}, {'@': 37}, {'@': 38}, {'@': 39}, {'@': 40}, {'@': 41}, {'@': 42}, {'@': 43}, {'@': 44}, {'@': 45}, {'@': 46}, {'@': 47}, {'@': 48}, {'@': 49}, {'@': 50}, {'@': 51}, {'@': 52}, {'@': 53}, {'@': 54}, {'@': 55}, {'@': 56}, {'@': 57}, {'@': 58}, {'@': 59}, {'@': 60}, {'@': 61}, {'@': 62}, {'@': 63}, {'@': 64}, {'@': 65}, {'@': 66}, {'@': 67}, {'@': 68}, {'@': 69}, {'@': 70}, {'@': 71}, {'@': 72}, {'@': 73}, {'@': 74}, {'@': 75}, {'@': 76}, {'@': 77}, {'@': 78}, {'@': 79}, {'@': 80}, {'@': 81}, {'@': 82}, {'@': 83}, {'@': 84}, {'@': 85}], 'ignore': ['WS'], 'g_regex_flags': 0, 'use_bytes': False, 'lexer_type': 'contextual', '__type__': 'LexerConf'}, 'parser_conf': {'rules': [{'@': 86}, {'@': 87}, {'@': 88}, {'@': 89}, {'@': 90}, {'@': 91}, {'@': 92}, {'@': 93}, {'@': 94}, {'@': 95}, {'@': 96}, {'@': 97}, {'@': 98}, {'@': 99}, {'@': 100}, {'@': 101}, {'@': 102}, {'@': 103}, {'@': 104}, {'@': 105}, {'@': 106}, {'@': 107}, {'@': 108}, {'@': 109}, {'@': 110}, {'@': 111}, {'@': 112}, {'@': 113}, {'@': 114}, {'@': 115}, {'@': 116}, {'@': 117}, {'@': 118}, {'@': 119}, {'@': 120}, {'@': 121}, {'@': 122}, {'@': 123}, {'@': 124}, {'@': 125}, {'@': 126}, {'@': 127}, {'@': 128}, {'@': 129}, {'@': 130}, {'@': 131}, {'@': 132}, {'@': 133}, {'@': 134}, {'@': 135}, {'@': 136}, {'@': 137}, {'@': 138}, {'@': 139}, {'@': 140}, {'@': 141}, {'@': 142}, {'@': 143}, {'@': 144}, {'@': 145}, {'@': 146}, {'@': 147}, {'@': 148}, {'@': 149}, {'@': 150}, {'@': 151}, {'@': 152}, {'@': 153}, {'@': 154}, {'@': 155}, {'@': 156}, {'@': 157}, {'@': 158}, {'@': 159}, {'@': 160}, {'@': 161}, {'@': 162}, {'@': 163}, {'@': 164}, {'@': 165}, {'@': 166}, {'@': 167}, {'@': 168}, {'@': 169}, {'@': 170}, {'@': 171}, {'@': 172}, {'@': 173}, {'@': 174}, {'@': 175}, {'@': 176}, {'@': 177}, {'@': 178}, {'@': 179}, {'@': 180}, {'@': 181}, {'@': 182}, {'@': 183}, {'@': 184}, {'@': 185}, {'@': 186}, {'@': 187}, {'@': 188}, {'@': 189}, {'@': 190}, {'@': 191}, {'@': 192}, {'@': 193}, {'@': 194}, {'@': 195}, {'@': 196}, {'@': 197}, {'@': 198}, {'@': 199}, {'@': 200}, {'@': 201}, {'@': 202}, {'@': 203}, {'@': 204}, {'@': 205}, {'@': 206}, {'@': 207}, {'@': 208}, {'@': 209}, {'@': 210}, {'@': 211}, {'@': 212}, {'@': 213}, {'@': 214}, {'@': 215}, {'@': 216}, {'@': 217}, {'@': 218}, {'@': 219}, {'@': 220}, {'@': 221}, {'@': 222}, {'@': 223}, {'@': 224}, {'@': 225}, {'@': 226}, {'@': 227}, {'@': 228}, {'@': 229}, {'@': 230}, {'@': 231}, {'@': 232}, {'@': 233}, {'@': 234}, {'@': 235}, {'@': 236}, {'@': 237}, {'@': 238}, {'@': 239}, {'@': 240}, {'@': 241}, {'@': 242}, {'@': 243}, {'@': 244}, {'@': 245}, {'@': 246}, {'@': 247}, {'@': 248}, {'@': 249}, {'@': 250}, {'@': 251}, {'@': 252}, {'@': 253}, {'@': 254}, {'@': 255}, {'@': 256}, {'@': 257}, {'@': 258}, {'@': 259}, {'@': 260}, {'@': 261}, {'@': 262}, {'@': 263}, {'@': 264}, {'@': 265}, {'@': 266}, {'@': 267}, {'@': 268}, {'@': 269}, {'@': 270}, {'@': 271}, {'@': 272}, {'@': 273}, {'@': 274}, {'@': 275}, {'@': 276}, {'@': 277}, {'@': 278}, {'@': 279}, {'@': 280}, {'@': 281}, {'@': 282}, {'@': 283}, {'@': 284}, {'@': 285}, {'@': 286}, {'@': 287}, {'@': 288}, {'@': 289}, {'@': 290}, {'@': 291}, {'@': 292}, {'@': 293}, {'@': 294}, {'@': 295}, {'@': 296}, {'@': 297}, {'@': 298}, {'@': 299}, {'@': 300}, {'@': 301}, {'@': 302}, {'@': 303}, {'@': 304}, {'@': 305}, {'@': 306}, {'@': 307}, {'@': 308}], 'start': ['start'], 'parser_type': 'lalr', '__type__': 'ParserConf'}, 'parser': {'tokens': {0: 'SIGNED_FLOAT', 1: 'BREAK', 2: 'VOLATILE', 3: 'STRING', 4: 'IF', 5: 'DECL', 6: 'DEREF', 7: 'FOR', 8: 'CONTINUE', 9: 'RETURN', 10: 'LPAR', 11: 'BLACKBOX', 12: 'LET', 13: 'INT', 14: 'STATIC', 15: 'ADDR', 16: 'FALSE', 17: 'NONE', 18: 'RBRACE', 19: 'ASSERT', 20: 'CAST', 21: 'PROTECTED', 22: 'IDENT', 23: 'ASM', 24: 'WHILE', 25: 'TRUE', 26: 'AT', 27: 'RPAR', 28: 'SEMICOLON', 29: 'DOT', 30: 'LSQB', 31: 'COMMA', 32: '__black_box_stmt_star_5', 33: 'fn_call_expr', 34: 'expr', 35: 'call_expr', 36: 'get_attr', 37: 'addr_expr', 38: 'fn_name', 39: 'atom', 40: 'if_then_else_expr', 41: 'deref_expr', 42: 'cast_expr', 43: 'get_item_expr', 44: 'let_expr', 45: 'symbol_var', 46: 'var', 47: 'LBRACE', 48: '__ANON_9', 49: '__ANON_7', 50: '__ANON_10', 51: 'SLASH', 52: '__ANON_3', 53: 'MINUS', 54: '__ANON_11', 55: '__ANON_4', 56: 'STAR', 57: '__ANON_6', 58: 'LESSTHAN', 59: 'AMPERSAND', 60: '__ANON_5', 61: 'CIRCUMFLEX', 62: 'VBAR', 63: '__ANON_8', 64: 'PLUS', 65: 'MORETHAN', 66: 'PERCENT', 67: 'THEN', 68: 'RSQB', 69: 'COLON', 70: 'AS', 71: 'EQUAL', 72: 'DOLLAR', 73: 'asm_label', 74: 'asm_group', 75: 'DEF', 76: '__asm_group_star_6', 77: 'IN', 78: 'for_ind_var', 79: 'static_decl', 80: 'declare_stmt', 81: '__module_star_1', 82: 'attribute_name', 83: 'HASH', 84: 'function', 85: 'attribute', 86: '_unary_op', 87: 'BANG', 88: 'name', 89: 'def_var', 90: 'dim3', 91: '__ANON_0', 92: 'I32', 93: 'I64', 94: 'data_type', 95: 'U64', 96: 'tensor_type', 97: 'U8', 98: 'ptr_type', 99: 'F16', 100: 'type', 101: 'U32', 102: 'TILDE', 103: 'BOOL', 104: 'U16', 105: 'VOID', 106: 'I8', 107: 'F64', 108: 'F32', 109: 'fn_type', 110: 'I16', 111: '__fn_args_star_2', 112: 'pair', 113: '$END', 114: 'ELSE', 115: 'break_stmt', 116: 'stmt', 117: 'evaluate_stmt', 118: 'volatile', 119: 'for_stmt', 120: '__stmt_body_star_3', 121: 'black_box_stmt', 122: 'assert_stmt', 123: 'continue_stmt', 124: 'return_stmt', 125: 'asm_stmt', 126: 'launch_kernel_stmt', 127: 'while_stmt', 128: 'stmt_body', 129: 'buffer_store_stmt', 130: 'assign_stmt', 131: 'protected', 132: 'let_stmt', 133: 'if_stmt', 134: '__call_args_star_8', 135: 'ANYWILDCARD', 136: '__list_star_12', 137: 'slice', 138: 'LIST', 139: 'DICT', 140: 'list', 141: 'value', 142: 'dict', 143: 'tensor_layout', 144: '__tensor_type_star_10', 145: 'keyword_arg', 146: '__let_stmt_plus_4', 147: '__ANON_2', 148: '__ANON_1', 149: 'task_mapping', 150: 'ON', 151: '__module_star_0', 152: '__dict_star_11', 153: 'call_args', 154: 'module', 155: 'MODULE', 156: '__get_item_expr_star_7', 157: 'start', 158: 'tensor_tt_shape', 159: '_binary_op', 160: 'fn_args', 161: '__fn_type_star_9'}, 'states': {0: {0: (1, {'@': 113}), 1: (1, {'@': 113}), 2: (1, {'@': 113}), 3: (1, {'@': 113}), 4: (1, {'@': 113}), 5: (1, {'@': 113}), 6: (1, {'@': 113}), 7: (1, {'@': 113}), 8: (1, {'@': 113}), 9: (1, {'@': 113}), 10: (1, {'@': 113}), 11: (1, {'@': 113}), 12: (1, {'@': 113}), 13: (1, {'@': 113}), 14: (1, {'@': 113}), 15: (1, {'@': 113}), 16: (1, {'@': 113}), 17: (1, {'@': 113}), 18: (1, {'@': 113}), 19: (1, {'@': 113}), 20: (1, {'@': 113}), 21: (1, {'@': 113}), 22: (1, {'@': 113}), 23: (1, {'@': 113}), 24: (1, {'@': 113}), 25: (1, {'@': 113}), 26: (1, {'@': 113}), 27: (1, {'@': 113})}, 1: {18: (0, 205)}, 2: {28: (0, 79), 29: (0, 406), 30: (0, 399)}, 3: {29: (0, 406), 30: (0, 399), 27: (0, 264), 31: (0, 336), 32: (0, 254)}, 4: {33: (0, 427), 34: (0, 236), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 5: {28: (0, 242)}, 6: {29: (1, {'@': 219}), 30: (1, {'@': 219}), 47: (1, {'@': 219}), 27: (1, {'@': 219}), 28: (1, {'@': 219}), 31: (1, {'@': 219}), 48: (1, {'@': 219}), 49: (1, {'@': 219}), 50: (1, {'@': 219}), 51: (1, {'@': 219}), 52: (1, {'@': 219}), 53: (1, {'@': 219}), 54: (1, {'@': 219}), 55: (1, {'@': 219}), 56: (1, {'@': 219}), 57: (1, {'@': 219}), 58: (1, {'@': 219}), 59: (1, {'@': 219}), 60: (1, {'@': 219}), 61: (1, {'@': 219}), 62: (1, {'@': 219}), 63: (1, {'@': 219}), 64: (1, {'@': 219}), 65: (1, {'@': 219}), 66: (1, {'@': 219}), 67: (1, {'@': 219}), 68: (1, {'@': 219}), 69: (1, {'@': 219}), 18: (1, {'@': 219}), 70: (1, {'@': 219})}, 7: {28: (0, 217)}, 8: {0: (1, {'@': 133}), 1: (1, {'@': 133}), 2: (1, {'@': 133}), 3: (1, {'@': 133}), 4: (1, {'@': 133}), 5: (1, {'@': 133}), 6: (1, {'@': 133}), 7: (1, {'@': 133}), 8: (1, {'@': 133}), 9: (1, {'@': 133}), 10: (1, {'@': 133}), 11: (1, {'@': 133}), 12: (1, {'@': 133}), 13: (1, {'@': 133}), 14: (1, {'@': 133}), 15: (1, {'@': 133}), 16: (1, {'@': 133}), 17: (1, {'@': 133}), 18: (1, {'@': 133}), 19: (1, {'@': 133}), 20: (1, {'@': 133}), 21: (1, {'@': 133}), 22: (1, {'@': 133}), 23: (1, {'@': 133}), 24: (1, {'@': 133}), 25: (1, {'@': 133}), 26: (1, {'@': 133}), 27: (1, {'@': 133})}, 9: {18: (0, 259)}, 10: {47: (1, {'@': 136})}, 11: {71: (0, 445), 29: (1, {'@': 226}), 30: (1, {'@': 226}), 28: (1, {'@': 226})}, 12: {18: (0, 69)}, 13: {72: (0, 241)}, 14: {18: (0, 246)}, 15: {18: (0, 226)}, 16: {73: (0, 18), 3: (0, 200), 18: (0, 85), 74: (0, 195)}, 17: {75: (1, {'@': 94}), 5: (1, {'@': 94}), 18: (1, {'@': 94}), 14: (1, {'@': 94})}, 18: {76: (0, 211), 31: (0, 230), 18: (1, {'@': 158})}, 19: {0: (1, {'@': 154}), 1: (1, {'@': 154}), 2: (1, {'@': 154}), 3: (1, {'@': 154}), 4: (1, {'@': 154}), 5: (1, {'@': 154}), 6: (1, {'@': 154}), 7: (1, {'@': 154}), 8: (1, {'@': 154}), 9: (1, {'@': 154}), 10: (1, {'@': 154}), 11: (1, {'@': 154}), 12: (1, {'@': 154}), 13: (1, {'@': 154}), 14: (1, {'@': 154}), 15: (1, {'@': 154}), 16: (1, {'@': 154}), 17: (1, {'@': 154}), 18: (1, {'@': 154}), 19: (1, {'@': 154}), 20: (1, {'@': 154}), 21: (1, {'@': 154}), 22: (1, {'@': 154}), 23: (1, {'@': 154}), 24: (1, {'@': 154}), 25: (1, {'@': 154}), 26: (1, {'@': 154}), 27: (1, {'@': 154})}, 20: {23: (0, 233)}, 21: {47: (0, 221)}, 22: {33: (0, 427), 34: (0, 216), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 23: {47: (0, 108)}, 24: {28: (0, 440)}, 25: {0: (1, {'@': 131}), 1: (1, {'@': 131}), 2: (1, {'@': 131}), 3: (1, {'@': 131}), 4: (1, {'@': 131}), 5: (1, {'@': 131}), 6: (1, {'@': 131}), 7: (1, {'@': 131}), 8: (1, {'@': 131}), 9: (1, {'@': 131}), 10: (1, {'@': 131}), 11: (1, {'@': 131}), 12: (1, {'@': 131}), 13: (1, {'@': 131}), 14: (1, {'@': 131}), 15: (1, {'@': 131}), 16: (1, {'@': 131}), 17: (1, {'@': 131}), 18: (1, {'@': 131}), 19: (1, {'@': 131}), 20: (1, {'@': 131}), 21: (1, {'@': 131}), 22: (1, {'@': 131}), 23: (1, {'@': 131}), 24: (1, {'@': 131}), 25: (1, {'@': 131}), 26: (1, {'@': 131}), 27: (1, {'@': 131})}, 26: {33: (0, 427), 34: (0, 255), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 27: {0: (1, {'@': 124}), 1: (1, {'@': 124}), 2: (1, {'@': 124}), 3: (1, {'@': 124}), 4: (1, {'@': 124}), 5: (1, {'@': 124}), 6: (1, {'@': 124}), 7: (1, {'@': 124}), 8: (1, {'@': 124}), 9: (1, {'@': 124}), 10: (1, {'@': 124}), 11: (1, {'@': 124}), 12: (1, {'@': 124}), 13: (1, {'@': 124}), 14: (1, {'@': 124}), 15: (1, {'@': 124}), 16: (1, {'@': 124}), 17: (1, {'@': 124}), 18: (1, {'@': 124}), 19: (1, {'@': 124}), 20: (1, {'@': 124}), 21: (1, {'@': 124}), 22: (1, {'@': 124}), 23: (1, {'@': 124}), 24: (1, {'@': 124}), 25: (1, {'@': 124}), 26: (1, {'@': 124}), 27: (1, {'@': 124})}, 28: {0: (1, {'@': 106}), 1: (1, {'@': 106}), 2: (1, {'@': 106}), 3: (1, {'@': 106}), 4: (1, {'@': 106}), 5: (1, {'@': 106}), 6: (1, {'@': 106}), 7: (1, {'@': 106}), 8: (1, {'@': 106}), 9: (1, {'@': 106}), 10: (1, {'@': 106}), 11: (1, {'@': 106}), 12: (1, {'@': 106}), 13: (1, {'@': 106}), 14: (1, {'@': 106}), 15: (1, {'@': 106}), 16: (1, {'@': 106}), 17: (1, {'@': 106}), 18: (1, {'@': 106}), 19: (1, {'@': 106}), 20: (1, {'@': 106}), 21: (1, {'@': 106}), 22: (1, {'@': 106}), 23: (1, {'@': 106}), 24: (1, {'@': 106}), 25: (1, {'@': 106}), 26: (1, {'@': 106}), 27: (1, {'@': 106})}, 29: {71: (0, 263)}, 30: {0: (1, {'@': 125}), 15: (1, {'@': 125}), 16: (1, {'@': 125}), 3: (1, {'@': 125}), 4: (1, {'@': 125}), 6: (1, {'@': 125}), 17: (1, {'@': 125}), 20: (1, {'@': 125}), 13: (1, {'@': 125}), 22: (1, {'@': 125}), 10: (1, {'@': 125}), 12: (1, {'@': 125}), 25: (1, {'@': 125}), 26: (1, {'@': 125})}, 31: {31: (0, 225)}, 32: {18: (0, 222)}, 33: {0: (1, {'@': 104}), 1: (1, {'@': 104}), 2: (1, {'@': 104}), 3: (1, {'@': 104}), 4: (1, {'@': 104}), 5: (1, {'@': 104}), 6: (1, {'@': 104}), 7: (1, {'@': 104}), 8: (1, {'@': 104}), 9: (1, {'@': 104}), 10: (1, {'@': 104}), 11: (1, {'@': 104}), 12: (1, {'@': 104}), 13: (1, {'@': 104}), 14: (1, {'@': 104}), 15: (1, {'@': 104}), 16: (1, {'@': 104}), 17: (1, {'@': 104}), 18: (1, {'@': 104}), 19: (1, {'@': 104}), 20: (1, {'@': 104}), 21: (1, {'@': 104}), 22: (1, {'@': 104}), 23: (1, {'@': 104}), 24: (1, {'@': 104}), 25: (1, {'@': 104}), 26: (1, {'@': 104}), 27: (1, {'@': 104})}, 34: {27: (0, 306)}, 35: {10: (0, 290)}, 36: {28: (0, 253)}, 37: {0: (1, {'@': 287}), 1: (1, {'@': 287}), 4: (1, {'@': 287}), 8: (1, {'@': 287}), 10: (1, {'@': 287}), 11: (1, {'@': 287}), 15: (1, {'@': 287}), 19: (1, {'@': 287}), 20: (1, {'@': 287}), 22: (1, {'@': 287}), 24: (1, {'@': 287}), 26: (1, {'@': 287}), 2: (1, {'@': 287}), 3: (1, {'@': 287}), 5: (1, {'@': 287}), 6: (1, {'@': 287}), 7: (1, {'@': 287}), 9: (1, {'@': 287}), 12: (1, {'@': 287}), 13: (1, {'@': 287}), 14: (1, {'@': 287}), 16: (1, {'@': 287}), 17: (1, {'@': 287}), 18: (1, {'@': 287}), 21: (1, {'@': 287}), 23: (1, {'@': 287}), 25: (1, {'@': 287}), 27: (1, {'@': 287})}, 38: {28: (0, 54), 77: (0, 272)}, 39: {18: (1, {'@': 156}), 31: (1, {'@': 156})}, 40: {10: (0, 91)}, 41: {77: (1, {'@': 134})}, 42: {33: (0, 427), 34: (0, 219), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 43: {29: (0, 406), 30: (0, 399), 28: (0, 209)}, 44: {78: (0, 103), 10: (0, 67)}, 45: {18: (1, {'@': 294}), 31: (1, {'@': 294})}, 46: {29: (1, {'@': 218}), 30: (1, {'@': 218}), 47: (1, {'@': 218}), 27: (1, {'@': 218}), 28: (1, {'@': 218}), 31: (1, {'@': 218}), 48: (1, {'@': 218}), 49: (1, {'@': 218}), 50: (1, {'@': 218}), 51: (1, {'@': 218}), 52: (1, {'@': 218}), 53: (1, {'@': 218}), 54: (1, {'@': 218}), 55: (1, {'@': 218}), 56: (1, {'@': 218}), 57: (1, {'@': 218}), 58: (1, {'@': 218}), 59: (1, {'@': 218}), 60: (1, {'@': 218}), 61: (1, {'@': 218}), 62: (1, {'@': 218}), 63: (1, {'@': 218}), 64: (1, {'@': 218}), 65: (1, {'@': 218}), 66: (1, {'@': 218}), 67: (1, {'@': 218}), 68: (1, {'@': 218}), 69: (1, {'@': 218}), 18: (1, {'@': 218}), 70: (1, {'@': 218})}, 47: {22: (0, 458)}, 48: {18: (0, 96), 5: (0, 292), 79: (0, 340), 80: (0, 296), 81: (0, 89), 82: (0, 352), 75: (0, 70), 83: (0, 354), 84: (0, 64), 14: (0, 75), 85: (0, 401)}, 49: {27: (1, {'@': 235}), 31: (1, {'@': 235}), 71: (1, {'@': 235}), 28: (1, {'@': 235}), 47: (1, {'@': 235}), 83: (1, {'@': 235})}, 50: {10: (0, 186)}, 51: {33: (0, 427), 86: (0, 57), 34: (0, 412), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 53: (0, 56), 41: (0, 180), 42: (0, 273), 64: (0, 68), 16: (0, 193), 15: (0, 257), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 87: (0, 97), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 52: {88: (0, 95), 89: (0, 120), 22: (0, 99)}, 53: {75: (1, {'@': 114}), 5: (1, {'@': 114}), 18: (1, {'@': 114}), 14: (1, {'@': 114}), 0: (1, {'@': 114}), 1: (1, {'@': 114}), 2: (1, {'@': 114}), 3: (1, {'@': 114}), 4: (1, {'@': 114}), 6: (1, {'@': 114}), 7: (1, {'@': 114}), 8: (1, {'@': 114}), 9: (1, {'@': 114}), 10: (1, {'@': 114}), 11: (1, {'@': 114}), 12: (1, {'@': 114}), 13: (1, {'@': 114}), 15: (1, {'@': 114}), 16: (1, {'@': 114}), 17: (1, {'@': 114}), 19: (1, {'@': 114}), 20: (1, {'@': 114}), 21: (1, {'@': 114}), 22: (1, {'@': 114}), 23: (1, {'@': 114}), 24: (1, {'@': 114}), 25: (1, {'@': 114}), 26: (1, {'@': 114}), 27: (1, {'@': 114})}, 54: {0: (1, {'@': 130}), 1: (1, {'@': 130}), 2: (1, {'@': 130}), 3: (1, {'@': 130}), 4: (1, {'@': 130}), 5: (1, {'@': 130}), 6: (1, {'@': 130}), 7: (1, {'@': 130}), 8: (1, {'@': 130}), 9: (1, {'@': 130}), 10: (1, {'@': 130}), 11: (1, {'@': 130}), 12: (1, {'@': 130}), 13: (1, {'@': 130}), 14: (1, {'@': 130}), 15: (1, {'@': 130}), 16: (1, {'@': 130}), 17: (1, {'@': 130}), 18: (1, {'@': 130}), 19: (1, {'@': 130}), 20: (1, {'@': 130}), 21: (1, {'@': 130}), 22: (1, {'@': 130}), 23: (1, {'@': 130}), 24: (1, {'@': 130}), 25: (1, {'@': 130}), 26: (1, {'@': 130}), 27: (1, {'@': 130})}, 55: {75: (1, {'@': 284}), 5: (1, {'@': 284}), 18: (1, {'@': 284}), 14: (1, {'@': 284})}, 56: {0: (1, {'@': 169}), 15: (1, {'@': 169}), 16: (1, {'@': 169}), 3: (1, {'@': 169}), 4: (1, {'@': 169}), 6: (1, {'@': 169}), 17: (1, {'@': 169}), 20: (1, {'@': 169}), 13: (1, {'@': 169}), 22: (1, {'@': 169}), 10: (1, {'@': 169}), 12: (1, {'@': 169}), 25: (1, {'@': 169}), 26: (1, {'@': 169})}, 57: {33: (0, 427), 34: (0, 358), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 58: {82: (0, 311), 83: (0, 354), 75: (1, {'@': 117}), 5: (1, {'@': 117}), 18: (1, {'@': 117}), 14: (1, {'@': 117}), 0: (1, {'@': 117}), 1: (1, {'@': 117}), 2: (1, {'@': 117}), 3: (1, {'@': 117}), 4: (1, {'@': 117}), 6: (1, {'@': 117}), 7: (1, {'@': 117}), 8: (1, {'@': 117}), 9: (1, {'@': 117}), 10: (1, {'@': 117}), 11: (1, {'@': 117}), 12: (1, {'@': 117}), 13: (1, {'@': 117}), 15: (1, {'@': 117}), 16: (1, {'@': 117}), 17: (1, {'@': 117}), 19: (1, {'@': 117}), 20: (1, {'@': 117}), 21: (1, {'@': 117}), 22: (1, {'@': 117}), 23: (1, {'@': 117}), 24: (1, {'@': 117}), 25: (1, {'@': 117}), 26: (1, {'@': 117}), 27: (1, {'@': 117})}, 59: {90: (0, 467), 10: (0, 392)}, 60: {3: (0, 200), 73: (0, 45)}, 61: {27: (1, {'@': 268}), 31: (1, {'@': 268}), 28: (1, {'@': 268})}, 62: {91: (0, 482)}, 63: {92: (0, 248), 93: (0, 464), 94: (0, 505), 95: (0, 252), 96: (0, 228), 10: (0, 63), 97: (0, 185), 98: (0, 422), 99: (0, 210), 100: (0, 511), 101: (0, 256), 102: (0, 260), 103: (0, 457), 104: (0, 207), 105: (0, 237), 106: (0, 484), 107: (0, 208), 108: (0, 408), 27: (0, 62), 109: (0, 243), 110: (0, 227)}, 64: {75: (1, {'@': 281}), 5: (1, {'@': 281}), 18: (1, {'@': 281}), 14: (1, {'@': 281})}, 65: {31: (1, {'@': 298}), 68: (1, {'@': 298})}, 66: {14: (1, {'@': 279}), 5: (1, {'@': 279}), 75: (1, {'@': 279}), 18: (1, {'@': 279}), 83: (1, {'@': 279})}, 67: {88: (0, 95), 22: (0, 99), 89: (0, 204)}, 68: {0: (1, {'@': 168}), 15: (1, {'@': 168}), 16: (1, {'@': 168}), 3: (1, {'@': 168}), 4: (1, {'@': 168}), 6: (1, {'@': 168}), 17: (1, {'@': 168}), 20: (1, {'@': 168}), 13: (1, {'@': 168}), 22: (1, {'@': 168}), 10: (1, {'@': 168}), 12: (1, {'@': 168}), 25: (1, {'@': 168}), 26: (1, {'@': 168})}, 69: {28: (0, 270)}, 70: {88: (0, 87), 22: (0, 99)}, 71: {27: (0, 373)}, 72: {29: (0, 406), 30: (0, 399), 67: (0, 382)}, 73: {31: (0, 366), 111: (0, 353), 27: (1, {'@': 96})}, 74: {29: (0, 406), 30: (0, 399), 70: (0, 410)}, 75: {5: (1, {'@': 122})}, 76: {29: (0, 406), 69: (0, 421), 30: (0, 399), 31: (1, {'@': 297}), 68: (1, {'@': 297})}, 77: {112: (0, 466), 3: (0, 316)}, 78: {91: (0, 426)}, 79: {0: (1, {'@': 164}), 1: (1, {'@': 164}), 2: (1, {'@': 164}), 3: (1, {'@': 164}), 4: (1, {'@': 164}), 5: (1, {'@': 164}), 6: (1, {'@': 164}), 7: (1, {'@': 164}), 8: (1, {'@': 164}), 9: (1, {'@': 164}), 10: (1, {'@': 164}), 11: (1, {'@': 164}), 12: (1, {'@': 164}), 13: (1, {'@': 164}), 14: (1, {'@': 164}), 15: (1, {'@': 164}), 16: (1, {'@': 164}), 17: (1, {'@': 164}), 18: (1, {'@': 164}), 19: (1, {'@': 164}), 20: (1, {'@': 164}), 21: (1, {'@': 164}), 22: (1, {'@': 164}), 23: (1, {'@': 164}), 24: (1, {'@': 164}), 25: (1, {'@': 164}), 26: (1, {'@': 164}), 27: (1, {'@': 164})}, 80: {113: (1, {'@': 90})}, 81: {0: (1, {'@': 175}), 15: (1, {'@': 175}), 16: (1, {'@': 175}), 3: (1, {'@': 175}), 4: (1, {'@': 175}), 6: (1, {'@': 175}), 17: (1, {'@': 175}), 20: (1, {'@': 175}), 13: (1, {'@': 175}), 22: (1, {'@': 175}), 10: (1, {'@': 175}), 12: (1, {'@': 175}), 25: (1, {'@': 175}), 26: (1, {'@': 175})}, 82: {114: (0, 295)}, 83: {19: (0, 437), 79: (0, 340), 33: (0, 427), 38: (0, 397), 23: (0, 23), 5: (0, 292), 26: (0, 47), 17: (0, 494), 7: (0, 44), 20: (0, 438), 115: (0, 28), 11: (0, 40), 116: (0, 37), 21: (0, 30), 37: (0, 488), 34: (0, 2), 117: (0, 0), 24: (0, 4), 10: (0, 51), 46: (0, 11), 118: (0, 20), 119: (0, 33), 39: (0, 468), 8: (0, 24), 2: (0, 142), 40: (0, 201), 120: (0, 152), 121: (0, 150), 122: (0, 161), 41: (0, 180), 42: (0, 273), 16: (0, 193), 123: (0, 158), 43: (0, 164), 124: (0, 141), 125: (0, 135), 12: (0, 129), 4: (0, 124), 25: (0, 442), 9: (0, 112), 126: (0, 154), 127: (0, 147), 35: (0, 178), 36: (0, 232), 0: (0, 268), 1: (0, 288), 6: (0, 249), 128: (0, 492), 129: (0, 294), 130: (0, 303), 15: (0, 257), 131: (0, 224), 22: (0, 262), 14: (0, 75), 3: (0, 265), 13: (0, 267), 80: (0, 244), 44: (0, 274), 132: (0, 220), 45: (0, 504), 133: (0, 218), 18: (1, {'@': 98})}, 84: {0: (1, {'@': 189}), 15: (1, {'@': 189}), 16: (1, {'@': 189}), 3: (1, {'@': 189}), 4: (1, {'@': 189}), 6: (1, {'@': 189}), 17: (1, {'@': 189}), 20: (1, {'@': 189}), 13: (1, {'@': 189}), 22: (1, {'@': 189}), 10: (1, {'@': 189}), 12: (1, {'@': 189}), 25: (1, {'@': 189}), 26: (1, {'@': 189})}, 85: {18: (0, 231)}, 86: {29: (0, 406), 30: (0, 399), 27: (1, {'@': 292}), 31: (1, {'@': 292})}, 87: {10: (0, 435)}, 88: {74: (0, 1), 18: (0, 21), 73: (0, 18), 3: (0, 200)}, 89: {5: (0, 292), 79: (0, 340), 18: (0, 452), 75: (0, 70), 84: (0, 94), 14: (0, 75), 80: (0, 55)}, 90: {27: (1, {'@': 305}), 31: (1, {'@': 305})}, 91: {33: (0, 427), 34: (0, 478), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 16: (0, 193), 15: (0, 257), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 27: (0, 503), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 92: {27: (1, {'@': 241}), 31: (1, {'@': 241}), 71: (1, {'@': 241}), 28: (1, {'@': 241}), 47: (1, {'@': 241}), 83: (1, {'@': 241})}, 93: {77: (0, 50)}, 94: {75: (1, {'@': 283}), 5: (1, {'@': 283}), 18: (1, {'@': 283}), 14: (1, {'@': 283})}, 95: {69: (0, 454)}, 96: {113: (1, {'@': 88})}, 97: {0: (1, {'@': 170}), 15: (1, {'@': 170}), 16: (1, {'@': 170}), 3: (1, {'@': 170}), 4: (1, {'@': 170}), 6: (1, {'@': 170}), 17: (1, {'@': 170}), 20: (1, {'@': 170}), 13: (1, {'@': 170}), 22: (1, {'@': 170}), 10: (1, {'@': 170}), 12: (1, {'@': 170}), 25: (1, {'@': 170}), 26: (1, {'@': 170})}, 98: {89: (0, 465), 88: (0, 95), 22: (0, 99)}, 99: {69: (1, {'@': 276}), 10: (1, {'@': 276}), 71: (1, {'@': 276}), 48: (1, {'@': 276}), 51: (1, {'@': 276}), 70: (1, {'@': 276}), 29: (1, {'@': 276}), 56: (1, {'@': 276}), 57: (1, {'@': 276}), 30: (1, {'@': 276}), 59: (1, {'@': 276}), 60: (1, {'@': 276}), 28: (1, {'@': 276}), 47: (1, {'@': 276}), 64: (1, {'@': 276}), 65: (1, {'@': 276}), 67: (1, {'@': 276}), 49: (1, {'@': 276}), 50: (1, {'@': 276}), 52: (1, {'@': 276}), 53: (1, {'@': 276}), 27: (1, {'@': 276}), 68: (1, {'@': 276}), 54: (1, {'@': 276}), 55: (1, {'@': 276}), 58: (1, {'@': 276}), 61: (1, {'@': 276}), 62: (1, {'@': 276}), 18: (1, {'@': 276}), 63: (1, {'@': 276}), 31: (1, {'@': 276}), 66: (1, {'@': 276})}, 100: {71: (1, {'@': 275}), 27: (1, {'@': 275}), 31: (1, {'@': 275}), 28: (1, {'@': 275})}, 101: {113: (1, {'@': 89})}, 102: {29: (0, 406), 30: (0, 399), 28: (0, 385)}, 103: {77: (0, 400)}, 104: {71: (0, 417), 28: (0, 475)}, 105: {75: (1, {'@': 120}), 5: (1, {'@': 120}), 18: (1, {'@': 120}), 14: (1, {'@': 120}), 0: (1, {'@': 120}), 1: (1, {'@': 120}), 2: (1, {'@': 120}), 3: (1, {'@': 120}), 4: (1, {'@': 120}), 6: (1, {'@': 120}), 7: (1, {'@': 120}), 8: (1, {'@': 120}), 9: (1, {'@': 120}), 10: (1, {'@': 120}), 11: (1, {'@': 120}), 12: (1, {'@': 120}), 13: (1, {'@': 120}), 15: (1, {'@': 120}), 16: (1, {'@': 120}), 17: (1, {'@': 120}), 19: (1, {'@': 120}), 20: (1, {'@': 120}), 21: (1, {'@': 120}), 22: (1, {'@': 120}), 23: (1, {'@': 120}), 24: (1, {'@': 120}), 25: (1, {'@': 120}), 26: (1, {'@': 120}), 27: (1, {'@': 120})}, 106: {34: (0, 368), 33: (0, 427), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 107: {31: (0, 281), 27: (0, 41)}, 108: {3: (0, 339)}, 109: {28: (0, 329)}, 110: {27: (1, {'@': 238}), 31: (1, {'@': 238}), 71: (1, {'@': 238}), 28: (1, {'@': 238}), 47: (1, {'@': 238}), 83: (1, {'@': 238})}, 111: {33: (0, 427), 34: (0, 493), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 112: {33: (0, 427), 34: (0, 182), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 28: (0, 190), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 113: {27: (1, {'@': 302}), 31: (1, {'@': 302})}, 114: {29: (0, 406), 30: (0, 399), 18: (0, 82)}, 115: {27: (1, {'@': 308}), 31: (1, {'@': 308})}, 116: {27: (1, {'@': 272}), 31: (1, {'@': 272})}, 117: {27: (1, {'@': 236}), 31: (1, {'@': 236}), 71: (1, {'@': 236}), 28: (1, {'@': 236}), 47: (1, {'@': 236}), 83: (1, {'@': 236})}, 118: {29: (0, 406), 30: (0, 399), 27: (0, 10)}, 119: {31: (1, {'@': 296}), 68: (1, {'@': 296})}, 120: {71: (0, 514)}, 121: {65: (0, 472)}, 122: {0: (1, {'@': 138}), 1: (1, {'@': 138}), 2: (1, {'@': 138}), 3: (1, {'@': 138}), 4: (1, {'@': 138}), 5: (1, {'@': 138}), 6: (1, {'@': 138}), 7: (1, {'@': 138}), 8: (1, {'@': 138}), 9: (1, {'@': 138}), 10: (1, {'@': 138}), 11: (1, {'@': 138}), 12: (1, {'@': 138}), 13: (1, {'@': 138}), 14: (1, {'@': 138}), 15: (1, {'@': 138}), 16: (1, {'@': 138}), 17: (1, {'@': 138}), 18: (1, {'@': 138}), 19: (1, {'@': 138}), 20: (1, {'@': 138}), 21: (1, {'@': 138}), 22: (1, {'@': 138}), 23: (1, {'@': 138}), 24: (1, {'@': 138}), 25: (1, {'@': 138}), 26: (1, {'@': 138}), 27: (1, {'@': 138})}, 123: {29: (0, 406), 30: (0, 399), 31: (1, {'@': 202}), 68: (1, {'@': 202})}, 124: {33: (0, 427), 34: (0, 206), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 125: {29: (0, 406), 30: (0, 399), 31: (0, 448), 134: (0, 414), 32: (0, 395), 27: (1, {'@': 209})}, 126: {29: (0, 406), 30: (0, 399), 27: (1, {'@': 291}), 31: (1, {'@': 291})}, 127: {29: (0, 406), 30: (0, 399), 27: (0, 39)}, 128: {72: (0, 14)}, 129: {10: (0, 197)}, 130: {135: (0, 13)}, 131: {31: (0, 364), 27: (0, 483), 136: (0, 449)}, 132: {29: (0, 406), 27: (0, 93), 30: (0, 399)}, 133: {18: (0, 25)}, 134: {29: (0, 406), 69: (0, 421), 30: (0, 399), 31: (1, {'@': 295}), 68: (1, {'@': 295})}, 135: {0: (1, {'@': 111}), 1: (1, {'@': 111}), 2: (1, {'@': 111}), 3: (1, {'@': 111}), 4: (1, {'@': 111}), 5: (1, {'@': 111}), 6: (1, {'@': 111}), 7: (1, {'@': 111}), 8: (1, {'@': 111}), 9: (1, {'@': 111}), 10: (1, {'@': 111}), 11: (1, {'@': 111}), 12: (1, {'@': 111}), 13: (1, {'@': 111}), 14: (1, {'@': 111}), 15: (1, {'@': 111}), 16: (1, {'@': 111}), 17: (1, {'@': 111}), 18: (1, {'@': 111}), 19: (1, {'@': 111}), 20: (1, {'@': 111}), 21: (1, {'@': 111}), 22: (1, {'@': 111}), 23: (1, {'@': 111}), 24: (1, {'@': 111}), 25: (1, {'@': 111}), 26: (1, {'@': 111}), 27: (1, {'@': 111})}, 136: {19: (0, 437), 79: (0, 340), 33: (0, 427), 38: (0, 397), 23: (0, 23), 5: (0, 292), 26: (0, 47), 17: (0, 494), 7: (0, 44), 20: (0, 438), 115: (0, 28), 11: (0, 40), 116: (0, 37), 21: (0, 30), 37: (0, 488), 34: (0, 2), 117: (0, 0), 24: (0, 4), 10: (0, 51), 46: (0, 11), 118: (0, 20), 119: (0, 33), 39: (0, 468), 8: (0, 24), 2: (0, 142), 40: (0, 201), 120: (0, 152), 121: (0, 150), 122: (0, 161), 41: (0, 180), 42: (0, 273), 16: (0, 193), 123: (0, 158), 43: (0, 164), 124: (0, 141), 125: (0, 135), 12: (0, 129), 4: (0, 124), 25: (0, 442), 9: (0, 112), 126: (0, 154), 127: (0, 147), 35: (0, 178), 36: (0, 232), 0: (0, 268), 1: (0, 288), 6: (0, 249), 129: (0, 294), 128: (0, 515), 130: (0, 303), 15: (0, 257), 131: (0, 224), 22: (0, 262), 14: (0, 75), 3: (0, 265), 13: (0, 267), 80: (0, 244), 44: (0, 274), 132: (0, 220), 45: (0, 504), 133: (0, 218), 18: (1, {'@': 98})}, 137: {27: (1, {'@': 239}), 31: (1, {'@': 239}), 71: (1, {'@': 239}), 28: (1, {'@': 239}), 47: (1, {'@': 239}), 83: (1, {'@': 239})}, 138: {18: (0, 8)}, 139: {19: (0, 437), 79: (0, 340), 33: (0, 427), 38: (0, 397), 23: (0, 23), 5: (0, 292), 26: (0, 47), 17: (0, 494), 7: (0, 44), 20: (0, 438), 115: (0, 28), 11: (0, 40), 116: (0, 37), 21: (0, 30), 37: (0, 488), 34: (0, 2), 117: (0, 0), 24: (0, 4), 10: (0, 51), 46: (0, 11), 118: (0, 20), 119: (0, 33), 39: (0, 468), 8: (0, 24), 2: (0, 142), 40: (0, 201), 120: (0, 152), 121: (0, 150), 122: (0, 161), 41: (0, 180), 42: (0, 273), 16: (0, 193), 123: (0, 158), 43: (0, 164), 124: (0, 141), 125: (0, 135), 12: (0, 129), 4: (0, 124), 25: (0, 442), 9: (0, 112), 128: (0, 160), 126: (0, 154), 127: (0, 147), 35: (0, 178), 36: (0, 232), 0: (0, 268), 1: (0, 288), 6: (0, 249), 129: (0, 294), 130: (0, 303), 15: (0, 257), 131: (0, 224), 22: (0, 262), 14: (0, 75), 3: (0, 265), 13: (0, 267), 80: (0, 244), 44: (0, 274), 132: (0, 220), 45: (0, 504), 133: (0, 218), 18: (1, {'@': 98})}, 140: {47: (0, 16)}, 141: {0: (1, {'@': 102}), 1: (1, {'@': 102}), 2: (1, {'@': 102}), 3: (1, {'@': 102}), 4: (1, {'@': 102}), 5: (1, {'@': 102}), 6: (1, {'@': 102}), 7: (1, {'@': 102}), 8: (1, {'@': 102}), 9: (1, {'@': 102}), 10: (1, {'@': 102}), 11: (1, {'@': 102}), 12: (1, {'@': 102}), 13: (1, {'@': 102}), 14: (1, {'@': 102}), 15: (1, {'@': 102}), 16: (1, {'@': 102}), 17: (1, {'@': 102}), 18: (1, {'@': 102}), 19: (1, {'@': 102}), 20: (1, {'@': 102}), 21: (1, {'@': 102}), 22: (1, {'@': 102}), 23: (1, {'@': 102}), 24: (1, {'@': 102}), 25: (1, {'@': 102}), 26: (1, {'@': 102}), 27: (1, {'@': 102})}, 142: {23: (1, {'@': 155})}, 143: {27: (1, {'@': 299}), 31: (1, {'@': 299})}, 144: {10: (0, 26)}, 145: {29: (1, {'@': 198}), 30: (1, {'@': 198}), 47: (1, {'@': 198}), 27: (1, {'@': 198}), 28: (1, {'@': 198}), 31: (1, {'@': 198}), 48: (1, {'@': 198}), 49: (1, {'@': 198}), 50: (1, {'@': 198}), 51: (1, {'@': 198}), 52: (1, {'@': 198}), 53: (1, {'@': 198}), 54: (1, {'@': 198}), 55: (1, {'@': 198}), 56: (1, {'@': 198}), 57: (1, {'@': 198}), 58: (1, {'@': 198}), 59: (1, {'@': 198}), 60: (1, {'@': 198}), 61: (1, {'@': 198}), 62: (1, {'@': 198}), 63: (1, {'@': 198}), 64: (1, {'@': 198}), 65: (1, {'@': 198}), 66: (1, {'@': 198}), 71: (1, {'@': 198}), 67: (1, {'@': 198}), 68: (1, {'@': 198}), 69: (1, {'@': 198}), 18: (1, {'@': 198}), 70: (1, {'@': 198})}, 146: {29: (0, 406), 30: (0, 399), 28: (1, {'@': 245}), 31: (1, {'@': 245}), 65: (1, {'@': 245})}, 147: {0: (1, {'@': 105}), 1: (1, {'@': 105}), 2: (1, {'@': 105}), 3: (1, {'@': 105}), 4: (1, {'@': 105}), 5: (1, {'@': 105}), 6: (1, {'@': 105}), 7: (1, {'@': 105}), 8: (1, {'@': 105}), 9: (1, {'@': 105}), 10: (1, {'@': 105}), 11: (1, {'@': 105}), 12: (1, {'@': 105}), 13: (1, {'@': 105}), 14: (1, {'@': 105}), 15: (1, {'@': 105}), 16: (1, {'@': 105}), 17: (1, {'@': 105}), 18: (1, {'@': 105}), 19: (1, {'@': 105}), 20: (1, {'@': 105}), 21: (1, {'@': 105}), 22: (1, {'@': 105}), 23: (1, {'@': 105}), 24: (1, {'@': 105}), 25: (1, {'@': 105}), 26: (1, {'@': 105}), 27: (1, {'@': 105})}, 148: {0: (1, {'@': 132}), 1: (1, {'@': 132}), 2: (1, {'@': 132}), 3: (1, {'@': 132}), 4: (1, {'@': 132}), 5: (1, {'@': 132}), 6: (1, {'@': 132}), 7: (1, {'@': 132}), 8: (1, {'@': 132}), 9: (1, {'@': 132}), 10: (1, {'@': 132}), 11: (1, {'@': 132}), 12: (1, {'@': 132}), 13: (1, {'@': 132}), 14: (1, {'@': 132}), 15: (1, {'@': 132}), 16: (1, {'@': 132}), 17: (1, {'@': 132}), 18: (1, {'@': 132}), 19: (1, {'@': 132}), 20: (1, {'@': 132}), 21: (1, {'@': 132}), 22: (1, {'@': 132}), 23: (1, {'@': 132}), 24: (1, {'@': 132}), 25: (1, {'@': 132}), 26: (1, {'@': 132}), 27: (1, {'@': 132})}, 149: {27: (1, {'@': 244}), 31: (1, {'@': 244}), 71: (1, {'@': 244}), 28: (1, {'@': 244}), 47: (1, {'@': 244}), 83: (1, {'@': 244})}, 150: {0: (1, {'@': 110}), 1: (1, {'@': 110}), 2: (1, {'@': 110}), 3: (1, {'@': 110}), 4: (1, {'@': 110}), 5: (1, {'@': 110}), 6: (1, {'@': 110}), 7: (1, {'@': 110}), 8: (1, {'@': 110}), 9: (1, {'@': 110}), 10: (1, {'@': 110}), 11: (1, {'@': 110}), 12: (1, {'@': 110}), 13: (1, {'@': 110}), 14: (1, {'@': 110}), 15: (1, {'@': 110}), 16: (1, {'@': 110}), 17: (1, {'@': 110}), 18: (1, {'@': 110}), 19: (1, {'@': 110}), 20: (1, {'@': 110}), 21: (1, {'@': 110}), 22: (1, {'@': 110}), 23: (1, {'@': 110}), 24: (1, {'@': 110}), 25: (1, {'@': 110}), 26: (1, {'@': 110}), 27: (1, {'@': 110})}, 151: {33: (0, 427), 34: (0, 76), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 137: (0, 65), 69: (0, 376), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 152: {19: (0, 437), 79: (0, 340), 33: (0, 427), 38: (0, 397), 23: (0, 23), 5: (0, 292), 26: (0, 47), 17: (0, 494), 7: (0, 44), 20: (0, 438), 115: (0, 28), 11: (0, 40), 21: (0, 30), 37: (0, 488), 10: (0, 51), 34: (0, 2), 117: (0, 0), 24: (0, 4), 46: (0, 11), 118: (0, 20), 119: (0, 33), 39: (0, 468), 8: (0, 24), 2: (0, 142), 40: (0, 201), 121: (0, 150), 122: (0, 161), 41: (0, 180), 42: (0, 273), 16: (0, 193), 123: (0, 158), 43: (0, 164), 124: (0, 141), 125: (0, 135), 12: (0, 129), 4: (0, 124), 25: (0, 442), 9: (0, 112), 126: (0, 154), 127: (0, 147), 116: (0, 194), 35: (0, 178), 36: (0, 232), 0: (0, 268), 1: (0, 288), 6: (0, 249), 129: (0, 294), 130: (0, 303), 15: (0, 257), 131: (0, 224), 22: (0, 262), 14: (0, 75), 3: (0, 265), 13: (0, 267), 80: (0, 244), 44: (0, 274), 132: (0, 220), 45: (0, 504), 133: (0, 218), 18: (1, {'@': 97}), 27: (1, {'@': 97})}, 153: {27: (1, {'@': 300}), 31: (1, {'@': 300})}, 154: {0: (1, {'@': 112}), 1: (1, {'@': 112}), 2: (1, {'@': 112}), 3: (1, {'@': 112}), 4: (1, {'@': 112}), 5: (1, {'@': 112}), 6: (1, {'@': 112}), 7: (1, {'@': 112}), 8: (1, {'@': 112}), 9: (1, {'@': 112}), 10: (1, {'@': 112}), 11: (1, {'@': 112}), 12: (1, {'@': 112}), 13: (1, {'@': 112}), 14: (1, {'@': 112}), 15: (1, {'@': 112}), 16: (1, {'@': 112}), 17: (1, {'@': 112}), 18: (1, {'@': 112}), 19: (1, {'@': 112}), 20: (1, {'@': 112}), 21: (1, {'@': 112}), 22: (1, {'@': 112}), 23: (1, {'@': 112}), 24: (1, {'@': 112}), 25: (1, {'@': 112}), 26: (1, {'@': 112}), 27: (1, {'@': 112})}, 155: {29: (0, 406), 30: (0, 399), 28: (0, 349)}, 156: {18: (0, 36)}, 157: {33: (0, 427), 34: (0, 347), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 158: {0: (1, {'@': 107}), 1: (1, {'@': 107}), 2: (1, {'@': 107}), 3: (1, {'@': 107}), 4: (1, {'@': 107}), 5: (1, {'@': 107}), 6: (1, {'@': 107}), 7: (1, {'@': 107}), 8: (1, {'@': 107}), 9: (1, {'@': 107}), 10: (1, {'@': 107}), 11: (1, {'@': 107}), 12: (1, {'@': 107}), 13: (1, {'@': 107}), 14: (1, {'@': 107}), 15: (1, {'@': 107}), 16: (1, {'@': 107}), 17: (1, {'@': 107}), 18: (1, {'@': 107}), 19: (1, {'@': 107}), 20: (1, {'@': 107}), 21: (1, {'@': 107}), 22: (1, {'@': 107}), 23: (1, {'@': 107}), 24: (1, {'@': 107}), 25: (1, {'@': 107}), 26: (1, {'@': 107}), 27: (1, {'@': 107})}, 159: {82: (0, 509), 47: (0, 136), 83: (0, 354)}, 160: {18: (0, 17)}, 161: {0: (1, {'@': 109}), 1: (1, {'@': 109}), 2: (1, {'@': 109}), 3: (1, {'@': 109}), 4: (1, {'@': 109}), 5: (1, {'@': 109}), 6: (1, {'@': 109}), 7: (1, {'@': 109}), 8: (1, {'@': 109}), 9: (1, {'@': 109}), 10: (1, {'@': 109}), 11: (1, {'@': 109}), 12: (1, {'@': 109}), 13: (1, {'@': 109}), 14: (1, {'@': 109}), 15: (1, {'@': 109}), 16: (1, {'@': 109}), 17: (1, {'@': 109}), 18: (1, {'@': 109}), 19: (1, {'@': 109}), 20: (1, {'@': 109}), 21: (1, {'@': 109}), 22: (1, {'@': 109}), 23: (1, {'@': 109}), 24: (1, {'@': 109}), 25: (1, {'@': 109}), 26: (1, {'@': 109}), 27: (1, {'@': 109})}, 162: {93: (0, 464), 138: (0, 510), 33: (0, 427), 98: (0, 422), 26: (0, 47), 17: (0, 494), 20: (0, 438), 94: (0, 505), 37: (0, 488), 10: (0, 460), 103: (0, 457), 34: (0, 430), 38: (0, 402), 106: (0, 484), 108: (0, 408), 39: (0, 468), 110: (0, 227), 40: (0, 201), 92: (0, 248), 41: (0, 180), 42: (0, 273), 16: (0, 193), 97: (0, 185), 139: (0, 293), 101: (0, 256), 43: (0, 223), 12: (0, 203), 4: (0, 188), 96: (0, 228), 140: (0, 213), 35: (0, 178), 99: (0, 210), 36: (0, 232), 104: (0, 207), 105: (0, 237), 0: (0, 268), 6: (0, 249), 141: (0, 115), 95: (0, 252), 15: (0, 257), 102: (0, 260), 22: (0, 262), 3: (0, 265), 13: (0, 267), 142: (0, 271), 107: (0, 208), 44: (0, 274), 109: (0, 243), 25: (0, 442), 45: (0, 504), 100: (0, 516), 46: (0, 416)}, 163: {47: (0, 83)}, 164: {71: (0, 179), 29: (1, {'@': 190}), 30: (1, {'@': 190}), 28: (1, {'@': 190})}, 165: {22: (1, {'@': 290}), 27: (1, {'@': 290})}, 166: {33: (0, 427), 34: (0, 502), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 167: {143: (0, 418), 144: (0, 500), 31: (0, 461), 65: (0, 512), 28: (0, 166)}, 168: {0: (1, {'@': 179}), 15: (1, {'@': 179}), 16: (1, {'@': 179}), 3: (1, {'@': 179}), 4: (1, {'@': 179}), 6: (1, {'@': 179}), 17: (1, {'@': 179}), 20: (1, {'@': 179}), 13: (1, {'@': 179}), 22: (1, {'@': 179}), 10: (1, {'@': 179}), 12: (1, {'@': 179}), 25: (1, {'@': 179}), 26: (1, {'@': 179})}, 169: {19: (0, 437), 79: (0, 340), 33: (0, 427), 38: (0, 397), 23: (0, 23), 5: (0, 292), 26: (0, 47), 17: (0, 494), 7: (0, 44), 20: (0, 438), 115: (0, 28), 11: (0, 40), 116: (0, 37), 21: (0, 30), 37: (0, 488), 34: (0, 2), 117: (0, 0), 24: (0, 4), 10: (0, 51), 46: (0, 11), 118: (0, 20), 119: (0, 33), 39: (0, 468), 8: (0, 24), 2: (0, 142), 40: (0, 201), 120: (0, 152), 121: (0, 150), 122: (0, 161), 41: (0, 180), 42: (0, 273), 16: (0, 193), 123: (0, 158), 43: (0, 164), 124: (0, 141), 125: (0, 135), 12: (0, 129), 4: (0, 124), 25: (0, 442), 9: (0, 112), 128: (0, 258), 126: (0, 154), 127: (0, 147), 35: (0, 178), 36: (0, 232), 0: (0, 268), 1: (0, 288), 6: (0, 249), 129: (0, 294), 130: (0, 303), 15: (0, 257), 131: (0, 224), 22: (0, 262), 14: (0, 75), 3: (0, 265), 13: (0, 267), 80: (0, 244), 44: (0, 274), 132: (0, 220), 45: (0, 504), 133: (0, 218), 18: (1, {'@': 98})}, 170: {75: (1, {'@': 91}), 5: (1, {'@': 91}), 18: (1, {'@': 91}), 14: (1, {'@': 91})}, 171: {18: (0, 5)}, 172: {31: (0, 419), 27: (1, {'@': 212})}, 173: {28: (1, {'@': 304}), 31: (1, {'@': 304}), 65: (1, {'@': 304})}, 174: {29: (0, 406), 30: (0, 399), 27: (1, {'@': 214}), 31: (1, {'@': 214})}, 175: {33: (0, 427), 34: (0, 86), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 145: (0, 143), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 88: (0, 362), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 370), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 176: {33: (0, 427), 34: (0, 3), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 177: {18: (0, 12), 73: (0, 18), 3: (0, 200), 74: (0, 32)}, 178: {29: (1, {'@': 194}), 30: (1, {'@': 194}), 47: (1, {'@': 194}), 27: (1, {'@': 194}), 28: (1, {'@': 194}), 31: (1, {'@': 194}), 48: (1, {'@': 194}), 49: (1, {'@': 194}), 50: (1, {'@': 194}), 51: (1, {'@': 194}), 52: (1, {'@': 194}), 53: (1, {'@': 194}), 54: (1, {'@': 194}), 55: (1, {'@': 194}), 56: (1, {'@': 194}), 57: (1, {'@': 194}), 58: (1, {'@': 194}), 59: (1, {'@': 194}), 60: (1, {'@': 194}), 61: (1, {'@': 194}), 62: (1, {'@': 194}), 63: (1, {'@': 194}), 64: (1, {'@': 194}), 65: (1, {'@': 194}), 66: (1, {'@': 194}), 67: (1, {'@': 194}), 68: (1, {'@': 194}), 69: (1, {'@': 194}), 18: (1, {'@': 194}), 70: (1, {'@': 194})}, 179: {33: (0, 427), 34: (0, 451), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 180: {29: (1, {'@': 192}), 30: (1, {'@': 192}), 47: (1, {'@': 192}), 27: (1, {'@': 192}), 28: (1, {'@': 192}), 31: (1, {'@': 192}), 48: (1, {'@': 192}), 49: (1, {'@': 192}), 50: (1, {'@': 192}), 51: (1, {'@': 192}), 52: (1, {'@': 192}), 53: (1, {'@': 192}), 54: (1, {'@': 192}), 55: (1, {'@': 192}), 56: (1, {'@': 192}), 57: (1, {'@': 192}), 58: (1, {'@': 192}), 59: (1, {'@': 192}), 60: (1, {'@': 192}), 61: (1, {'@': 192}), 62: (1, {'@': 192}), 63: (1, {'@': 192}), 64: (1, {'@': 192}), 65: (1, {'@': 192}), 66: (1, {'@': 192}), 67: (1, {'@': 192}), 68: (1, {'@': 192}), 69: (1, {'@': 192}), 18: (1, {'@': 192}), 70: (1, {'@': 192})}, 181: {29: (0, 406), 30: (0, 399)}, 182: {29: (0, 406), 30: (0, 399), 28: (0, 183)}, 183: {0: (1, {'@': 127}), 1: (1, {'@': 127}), 2: (1, {'@': 127}), 3: (1, {'@': 127}), 4: (1, {'@': 127}), 5: (1, {'@': 127}), 6: (1, {'@': 127}), 7: (1, {'@': 127}), 8: (1, {'@': 127}), 9: (1, {'@': 127}), 10: (1, {'@': 127}), 11: (1, {'@': 127}), 12: (1, {'@': 127}), 13: (1, {'@': 127}), 14: (1, {'@': 127}), 15: (1, {'@': 127}), 16: (1, {'@': 127}), 17: (1, {'@': 127}), 18: (1, {'@': 127}), 19: (1, {'@': 127}), 20: (1, {'@': 127}), 21: (1, {'@': 127}), 22: (1, {'@': 127}), 23: (1, {'@': 127}), 24: (1, {'@': 127}), 25: (1, {'@': 127}), 26: (1, {'@': 127}), 27: (1, {'@': 127})}, 184: {74: (0, 315), 73: (0, 18), 18: (0, 278), 3: (0, 200)}, 185: {27: (1, {'@': 251}), 31: (1, {'@': 251}), 58: (1, {'@': 251}), 71: (1, {'@': 251}), 28: (1, {'@': 251}), 47: (1, {'@': 251}), 83: (1, {'@': 251})}, 186: {33: (0, 427), 34: (0, 323), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 187: {47: (0, 240)}, 188: {33: (0, 427), 34: (0, 72), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 189: {77: (1, {'@': 135})}, 190: {0: (1, {'@': 128}), 1: (1, {'@': 128}), 2: (1, {'@': 128}), 3: (1, {'@': 128}), 4: (1, {'@': 128}), 5: (1, {'@': 128}), 6: (1, {'@': 128}), 7: (1, {'@': 128}), 8: (1, {'@': 128}), 9: (1, {'@': 128}), 10: (1, {'@': 128}), 11: (1, {'@': 128}), 12: (1, {'@': 128}), 13: (1, {'@': 128}), 14: (1, {'@': 128}), 15: (1, {'@': 128}), 16: (1, {'@': 128}), 17: (1, {'@': 128}), 18: (1, {'@': 128}), 19: (1, {'@': 128}), 20: (1, {'@': 128}), 21: (1, {'@': 128}), 22: (1, {'@': 128}), 23: (1, {'@': 128}), 24: (1, {'@': 128}), 25: (1, {'@': 128}), 26: (1, {'@': 128}), 27: (1, {'@': 128})}, 191: {19: (0, 437), 79: (0, 340), 33: (0, 427), 38: (0, 397), 23: (0, 23), 5: (0, 292), 26: (0, 47), 17: (0, 494), 7: (0, 44), 20: (0, 438), 115: (0, 28), 11: (0, 40), 116: (0, 37), 21: (0, 30), 37: (0, 488), 34: (0, 2), 117: (0, 0), 24: (0, 4), 10: (0, 51), 46: (0, 11), 118: (0, 20), 119: (0, 33), 39: (0, 468), 8: (0, 24), 2: (0, 142), 40: (0, 201), 120: (0, 152), 121: (0, 150), 122: (0, 161), 41: (0, 180), 42: (0, 273), 16: (0, 193), 123: (0, 158), 128: (0, 321), 43: (0, 164), 124: (0, 141), 125: (0, 135), 12: (0, 129), 4: (0, 124), 25: (0, 442), 9: (0, 112), 126: (0, 154), 127: (0, 147), 35: (0, 178), 36: (0, 232), 0: (0, 268), 1: (0, 288), 6: (0, 249), 129: (0, 294), 130: (0, 303), 15: (0, 257), 131: (0, 224), 22: (0, 262), 14: (0, 75), 3: (0, 265), 13: (0, 267), 80: (0, 244), 44: (0, 274), 132: (0, 220), 45: (0, 504), 133: (0, 218), 18: (1, {'@': 98})}, 192: {71: (0, 363), 29: (1, {'@': 190}), 30: (1, {'@': 190})}, 193: {29: (1, {'@': 223}), 30: (1, {'@': 223}), 47: (1, {'@': 223}), 27: (1, {'@': 223}), 28: (1, {'@': 223}), 31: (1, {'@': 223}), 48: (1, {'@': 223}), 49: (1, {'@': 223}), 50: (1, {'@': 223}), 51: (1, {'@': 223}), 52: (1, {'@': 223}), 53: (1, {'@': 223}), 54: (1, {'@': 223}), 55: (1, {'@': 223}), 56: (1, {'@': 223}), 57: (1, {'@': 223}), 58: (1, {'@': 223}), 59: (1, {'@': 223}), 60: (1, {'@': 223}), 61: (1, {'@': 223}), 62: (1, {'@': 223}), 63: (1, {'@': 223}), 64: (1, {'@': 223}), 65: (1, {'@': 223}), 66: (1, {'@': 223}), 67: (1, {'@': 223}), 68: (1, {'@': 223}), 69: (1, {'@': 223}), 18: (1, {'@': 223}), 70: (1, {'@': 223})}, 194: {0: (1, {'@': 288}), 1: (1, {'@': 288}), 4: (1, {'@': 288}), 8: (1, {'@': 288}), 10: (1, {'@': 288}), 11: (1, {'@': 288}), 15: (1, {'@': 288}), 19: (1, {'@': 288}), 20: (1, {'@': 288}), 22: (1, {'@': 288}), 24: (1, {'@': 288}), 26: (1, {'@': 288}), 2: (1, {'@': 288}), 3: (1, {'@': 288}), 5: (1, {'@': 288}), 6: (1, {'@': 288}), 7: (1, {'@': 288}), 9: (1, {'@': 288}), 12: (1, {'@': 288}), 13: (1, {'@': 288}), 14: (1, {'@': 288}), 16: (1, {'@': 288}), 17: (1, {'@': 288}), 18: (1, {'@': 288}), 21: (1, {'@': 288}), 23: (1, {'@': 288}), 25: (1, {'@': 288}), 27: (1, {'@': 288})}, 195: {18: (0, 235)}, 196: {0: (1, {'@': 129}), 1: (1, {'@': 129}), 2: (1, {'@': 129}), 3: (1, {'@': 129}), 4: (1, {'@': 129}), 5: (1, {'@': 129}), 6: (1, {'@': 129}), 7: (1, {'@': 129}), 8: (1, {'@': 129}), 9: (1, {'@': 129}), 10: (1, {'@': 129}), 11: (1, {'@': 129}), 12: (1, {'@': 129}), 13: (1, {'@': 129}), 14: (1, {'@': 129}), 15: (1, {'@': 129}), 16: (1, {'@': 129}), 17: (1, {'@': 129}), 18: (1, {'@': 129}), 19: (1, {'@': 129}), 20: (1, {'@': 129}), 21: (1, {'@': 129}), 22: (1, {'@': 129}), 23: (1, {'@': 129}), 24: (1, {'@': 129}), 25: (1, {'@': 129}), 26: (1, {'@': 129}), 27: (1, {'@': 129})}, 197: {88: (0, 95), 146: (0, 470), 89: (0, 507), 22: (0, 99)}, 198: {0: (1, {'@': 145}), 1: (1, {'@': 145}), 2: (1, {'@': 145}), 3: (1, {'@': 145}), 4: (1, {'@': 145}), 5: (1, {'@': 145}), 6: (1, {'@': 145}), 7: (1, {'@': 145}), 8: (1, {'@': 145}), 9: (1, {'@': 145}), 10: (1, {'@': 145}), 11: (1, {'@': 145}), 12: (1, {'@': 145}), 13: (1, {'@': 145}), 14: (1, {'@': 145}), 15: (1, {'@': 145}), 16: (1, {'@': 145}), 17: (1, {'@': 145}), 18: (1, {'@': 145}), 19: (1, {'@': 145}), 20: (1, {'@': 145}), 21: (1, {'@': 145}), 22: (1, {'@': 145}), 23: (1, {'@': 145}), 24: (1, {'@': 145}), 25: (1, {'@': 145}), 26: (1, {'@': 145}), 27: (1, {'@': 145})}, 199: {18: (0, 109)}, 200: {10: (0, 234)}, 201: {29: (1, {'@': 196}), 30: (1, {'@': 196}), 47: (1, {'@': 196}), 27: (1, {'@': 196}), 28: (1, {'@': 196}), 31: (1, {'@': 196}), 48: (1, {'@': 196}), 49: (1, {'@': 196}), 50: (1, {'@': 196}), 51: (1, {'@': 196}), 52: (1, {'@': 196}), 53: (1, {'@': 196}), 54: (1, {'@': 196}), 55: (1, {'@': 196}), 56: (1, {'@': 196}), 57: (1, {'@': 196}), 58: (1, {'@': 196}), 59: (1, {'@': 196}), 60: (1, {'@': 196}), 61: (1, {'@': 196}), 62: (1, {'@': 196}), 63: (1, {'@': 196}), 64: (1, {'@': 196}), 65: (1, {'@': 196}), 66: (1, {'@': 196}), 67: (1, {'@': 196}), 68: (1, {'@': 196}), 69: (1, {'@': 196}), 18: (1, {'@': 196}), 70: (1, {'@': 196})}, 202: {29: (0, 406), 30: (0, 399), 27: (0, 279)}, 203: {10: (0, 52)}, 204: {111: (0, 107), 27: (0, 189), 31: (0, 366)}, 205: {47: (0, 324)}, 206: {29: (0, 406), 30: (0, 399), 47: (0, 447), 67: (0, 382)}, 207: {27: (1, {'@': 252}), 31: (1, {'@': 252}), 58: (1, {'@': 252}), 71: (1, {'@': 252}), 28: (1, {'@': 252}), 47: (1, {'@': 252}), 83: (1, {'@': 252})}, 208: {27: (1, {'@': 257}), 31: (1, {'@': 257}), 58: (1, {'@': 257}), 71: (1, {'@': 257}), 28: (1, {'@': 257}), 47: (1, {'@': 257}), 83: (1, {'@': 257})}, 209: {0: (1, {'@': 123}), 1: (1, {'@': 123}), 2: (1, {'@': 123}), 3: (1, {'@': 123}), 4: (1, {'@': 123}), 5: (1, {'@': 123}), 6: (1, {'@': 123}), 7: (1, {'@': 123}), 8: (1, {'@': 123}), 9: (1, {'@': 123}), 10: (1, {'@': 123}), 11: (1, {'@': 123}), 12: (1, {'@': 123}), 13: (1, {'@': 123}), 14: (1, {'@': 123}), 15: (1, {'@': 123}), 16: (1, {'@': 123}), 17: (1, {'@': 123}), 18: (1, {'@': 123}), 19: (1, {'@': 123}), 20: (1, {'@': 123}), 21: (1, {'@': 123}), 22: (1, {'@': 123}), 23: (1, {'@': 123}), 24: (1, {'@': 123}), 25: (1, {'@': 123}), 26: (1, {'@': 123}), 27: (1, {'@': 123})}, 210: {27: (1, {'@': 255}), 31: (1, {'@': 255}), 58: (1, {'@': 255}), 71: (1, {'@': 255}), 28: (1, {'@': 255}), 47: (1, {'@': 255}), 83: (1, {'@': 255})}, 211: {31: (0, 60), 18: (1, {'@': 157})}, 212: {28: (0, 317)}, 213: {27: (1, {'@': 263}), 31: (1, {'@': 263})}, 214: {47: (0, 269)}, 215: {0: (1, {'@': 162}), 1: (1, {'@': 162}), 2: (1, {'@': 162}), 3: (1, {'@': 162}), 4: (1, {'@': 162}), 5: (1, {'@': 162}), 6: (1, {'@': 162}), 7: (1, {'@': 162}), 8: (1, {'@': 162}), 9: (1, {'@': 162}), 10: (1, {'@': 162}), 11: (1, {'@': 162}), 12: (1, {'@': 162}), 13: (1, {'@': 162}), 14: (1, {'@': 162}), 15: (1, {'@': 162}), 16: (1, {'@': 162}), 17: (1, {'@': 162}), 18: (1, {'@': 162}), 19: (1, {'@': 162}), 20: (1, {'@': 162}), 21: (1, {'@': 162}), 22: (1, {'@': 162}), 23: (1, {'@': 162}), 24: (1, {'@': 162}), 25: (1, {'@': 162}), 26: (1, {'@': 162}), 27: (1, {'@': 162})}, 216: {29: (0, 406), 27: (0, 93), 30: (0, 399), 28: (0, 286)}, 217: {0: (1, {'@': 152}), 1: (1, {'@': 152}), 2: (1, {'@': 152}), 3: (1, {'@': 152}), 4: (1, {'@': 152}), 5: (1, {'@': 152}), 6: (1, {'@': 152}), 7: (1, {'@': 152}), 8: (1, {'@': 152}), 9: (1, {'@': 152}), 10: (1, {'@': 152}), 11: (1, {'@': 152}), 12: (1, {'@': 152}), 13: (1, {'@': 152}), 14: (1, {'@': 152}), 15: (1, {'@': 152}), 16: (1, {'@': 152}), 17: (1, {'@': 152}), 18: (1, {'@': 152}), 19: (1, {'@': 152}), 20: (1, {'@': 152}), 21: (1, {'@': 152}), 22: (1, {'@': 152}), 23: (1, {'@': 152}), 24: (1, {'@': 152}), 25: (1, {'@': 152}), 26: (1, {'@': 152}), 27: (1, {'@': 152})}, 218: {0: (1, {'@': 108}), 1: (1, {'@': 108}), 2: (1, {'@': 108}), 3: (1, {'@': 108}), 4: (1, {'@': 108}), 5: (1, {'@': 108}), 6: (1, {'@': 108}), 7: (1, {'@': 108}), 8: (1, {'@': 108}), 9: (1, {'@': 108}), 10: (1, {'@': 108}), 11: (1, {'@': 108}), 12: (1, {'@': 108}), 13: (1, {'@': 108}), 14: (1, {'@': 108}), 15: (1, {'@': 108}), 16: (1, {'@': 108}), 17: (1, {'@': 108}), 18: (1, {'@': 108}), 19: (1, {'@': 108}), 20: (1, {'@': 108}), 21: (1, {'@': 108}), 22: (1, {'@': 108}), 23: (1, {'@': 108}), 24: (1, {'@': 108}), 25: (1, {'@': 108}), 26: (1, {'@': 108}), 27: (1, {'@': 108})}, 219: {29: (0, 406), 30: (0, 399), 31: (0, 330)}, 220: {0: (1, {'@': 103}), 1: (1, {'@': 103}), 2: (1, {'@': 103}), 3: (1, {'@': 103}), 4: (1, {'@': 103}), 5: (1, {'@': 103}), 6: (1, {'@': 103}), 7: (1, {'@': 103}), 8: (1, {'@': 103}), 9: (1, {'@': 103}), 10: (1, {'@': 103}), 11: (1, {'@': 103}), 12: (1, {'@': 103}), 13: (1, {'@': 103}), 14: (1, {'@': 103}), 15: (1, {'@': 103}), 16: (1, {'@': 103}), 17: (1, {'@': 103}), 18: (1, {'@': 103}), 19: (1, {'@': 103}), 20: (1, {'@': 103}), 21: (1, {'@': 103}), 22: (1, {'@': 103}), 23: (1, {'@': 103}), 24: (1, {'@': 103}), 25: (1, {'@': 103}), 26: (1, {'@': 103}), 27: (1, {'@': 103})}, 221: {74: (0, 328), 18: (0, 304), 73: (0, 18), 3: (0, 200)}, 222: {18: (0, 212)}, 223: {29: (1, {'@': 190}), 30: (1, {'@': 190}), 47: (1, {'@': 190}), 27: (1, {'@': 190}), 28: (1, {'@': 190}), 31: (1, {'@': 190}), 48: (1, {'@': 190}), 49: (1, {'@': 190}), 50: (1, {'@': 190}), 51: (1, {'@': 190}), 52: (1, {'@': 190}), 53: (1, {'@': 190}), 54: (1, {'@': 190}), 55: (1, {'@': 190}), 56: (1, {'@': 190}), 57: (1, {'@': 190}), 58: (1, {'@': 190}), 59: (1, {'@': 190}), 60: (1, {'@': 190}), 61: (1, {'@': 190}), 62: (1, {'@': 190}), 63: (1, {'@': 190}), 64: (1, {'@': 190}), 65: (1, {'@': 190}), 66: (1, {'@': 190}), 67: (1, {'@': 190}), 68: (1, {'@': 190}), 69: (1, {'@': 190}), 18: (1, {'@': 190}), 70: (1, {'@': 190})}, 224: {33: (0, 427), 34: (0, 181), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 43: (0, 192), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 225: {13: (0, 342), 147: (0, 337)}, 226: {114: (0, 429), 0: (1, {'@': 141}), 1: (1, {'@': 141}), 2: (1, {'@': 141}), 3: (1, {'@': 141}), 4: (1, {'@': 141}), 5: (1, {'@': 141}), 6: (1, {'@': 141}), 7: (1, {'@': 141}), 8: (1, {'@': 141}), 9: (1, {'@': 141}), 10: (1, {'@': 141}), 11: (1, {'@': 141}), 12: (1, {'@': 141}), 13: (1, {'@': 141}), 14: (1, {'@': 141}), 15: (1, {'@': 141}), 16: (1, {'@': 141}), 17: (1, {'@': 141}), 18: (1, {'@': 141}), 19: (1, {'@': 141}), 20: (1, {'@': 141}), 21: (1, {'@': 141}), 22: (1, {'@': 141}), 23: (1, {'@': 141}), 24: (1, {'@': 141}), 25: (1, {'@': 141}), 26: (1, {'@': 141}), 27: (1, {'@': 141})}, 227: {27: (1, {'@': 248}), 31: (1, {'@': 248}), 58: (1, {'@': 248}), 71: (1, {'@': 248}), 28: (1, {'@': 248}), 47: (1, {'@': 248}), 83: (1, {'@': 248})}, 228: {27: (1, {'@': 232}), 31: (1, {'@': 232}), 71: (1, {'@': 232}), 28: (1, {'@': 232}), 47: (1, {'@': 232}), 83: (1, {'@': 232})}, 229: {28: (0, 305)}, 230: {3: (0, 200), 73: (0, 334)}, 231: {28: (0, 344)}, 232: {29: (1, {'@': 221}), 30: (1, {'@': 221}), 47: (1, {'@': 221}), 27: (1, {'@': 221}), 28: (1, {'@': 221}), 31: (1, {'@': 221}), 48: (1, {'@': 221}), 49: (1, {'@': 221}), 50: (1, {'@': 221}), 51: (1, {'@': 221}), 52: (1, {'@': 221}), 53: (1, {'@': 221}), 54: (1, {'@': 221}), 55: (1, {'@': 221}), 56: (1, {'@': 221}), 57: (1, {'@': 221}), 58: (1, {'@': 221}), 59: (1, {'@': 221}), 60: (1, {'@': 221}), 61: (1, {'@': 221}), 62: (1, {'@': 221}), 63: (1, {'@': 221}), 64: (1, {'@': 221}), 65: (1, {'@': 221}), 66: (1, {'@': 221}), 67: (1, {'@': 221}), 68: (1, {'@': 221}), 69: (1, {'@': 221}), 18: (1, {'@': 221}), 70: (1, {'@': 221})}, 233: {47: (0, 365)}, 234: {33: (0, 427), 34: (0, 127), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 235: {18: (0, 326)}, 236: {29: (0, 406), 30: (0, 399), 47: (0, 506)}, 237: {27: (1, {'@': 259}), 31: (1, {'@': 259}), 58: (1, {'@': 259}), 71: (1, {'@': 259}), 28: (1, {'@': 259}), 47: (1, {'@': 259}), 83: (1, {'@': 259})}, 238: {28: (0, 291)}, 239: {27: (1, {'@': 266}), 31: (1, {'@': 266})}, 240: {72: (0, 130)}, 241: {18: (0, 238)}, 242: {0: (1, {'@': 153}), 1: (1, {'@': 153}), 2: (1, {'@': 153}), 3: (1, {'@': 153}), 4: (1, {'@': 153}), 5: (1, {'@': 153}), 6: (1, {'@': 153}), 7: (1, {'@': 153}), 8: (1, {'@': 153}), 9: (1, {'@': 153}), 10: (1, {'@': 153}), 11: (1, {'@': 153}), 12: (1, {'@': 153}), 13: (1, {'@': 153}), 14: (1, {'@': 153}), 15: (1, {'@': 153}), 16: (1, {'@': 153}), 17: (1, {'@': 153}), 18: (1, {'@': 153}), 19: (1, {'@': 153}), 20: (1, {'@': 153}), 21: (1, {'@': 153}), 22: (1, {'@': 153}), 23: (1, {'@': 153}), 24: (1, {'@': 153}), 25: (1, {'@': 153}), 26: (1, {'@': 153}), 27: (1, {'@': 153})}, 243: {27: (1, {'@': 234}), 31: (1, {'@': 234}), 71: (1, {'@': 234}), 28: (1, {'@': 234}), 47: (1, {'@': 234}), 83: (1, {'@': 234})}, 244: {0: (1, {'@': 99}), 1: (1, {'@': 99}), 2: (1, {'@': 99}), 3: (1, {'@': 99}), 4: (1, {'@': 99}), 5: (1, {'@': 99}), 6: (1, {'@': 99}), 7: (1, {'@': 99}), 8: (1, {'@': 99}), 9: (1, {'@': 99}), 10: (1, {'@': 99}), 11: (1, {'@': 99}), 12: (1, {'@': 99}), 13: (1, {'@': 99}), 14: (1, {'@': 99}), 15: (1, {'@': 99}), 16: (1, {'@': 99}), 17: (1, {'@': 99}), 18: (1, {'@': 99}), 19: (1, {'@': 99}), 20: (1, {'@': 99}), 21: (1, {'@': 99}), 22: (1, {'@': 99}), 23: (1, {'@': 99}), 24: (1, {'@': 99}), 25: (1, {'@': 99}), 26: (1, {'@': 99}), 27: (1, {'@': 99})}, 245: {33: (0, 427), 34: (0, 86), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 246: {28: (0, 198)}, 247: {19: (0, 437), 79: (0, 340), 33: (0, 427), 38: (0, 397), 23: (0, 23), 5: (0, 292), 26: (0, 47), 17: (0, 494), 7: (0, 44), 20: (0, 438), 115: (0, 28), 11: (0, 40), 116: (0, 37), 21: (0, 30), 37: (0, 488), 34: (0, 2), 117: (0, 0), 24: (0, 4), 10: (0, 51), 46: (0, 11), 118: (0, 20), 119: (0, 33), 39: (0, 468), 8: (0, 24), 2: (0, 142), 40: (0, 201), 120: (0, 152), 121: (0, 150), 122: (0, 161), 41: (0, 180), 42: (0, 273), 16: (0, 193), 123: (0, 158), 43: (0, 164), 124: (0, 141), 125: (0, 135), 12: (0, 129), 4: (0, 124), 25: (0, 442), 9: (0, 112), 126: (0, 154), 127: (0, 147), 35: (0, 178), 36: (0, 232), 128: (0, 138), 0: (0, 268), 1: (0, 288), 6: (0, 249), 129: (0, 294), 130: (0, 303), 15: (0, 257), 131: (0, 224), 22: (0, 262), 14: (0, 75), 3: (0, 265), 13: (0, 267), 80: (0, 244), 44: (0, 274), 132: (0, 220), 45: (0, 504), 133: (0, 218), 18: (1, {'@': 98})}, 248: {27: (1, {'@': 249}), 31: (1, {'@': 249}), 58: (1, {'@': 249}), 71: (1, {'@': 249}), 28: (1, {'@': 249}), 47: (1, {'@': 249}), 83: (1, {'@': 249})}, 249: {10: (0, 403)}, 250: {18: (0, 199)}, 251: {27: (0, 318), 31: (0, 245)}, 252: {27: (1, {'@': 254}), 31: (1, {'@': 254}), 58: (1, {'@': 254}), 71: (1, {'@': 254}), 28: (1, {'@': 254}), 47: (1, {'@': 254}), 83: (1, {'@': 254})}, 253: {0: (1, {'@': 146}), 1: (1, {'@': 146}), 2: (1, {'@': 146}), 3: (1, {'@': 146}), 4: (1, {'@': 146}), 5: (1, {'@': 146}), 6: (1, {'@': 146}), 7: (1, {'@': 146}), 8: (1, {'@': 146}), 9: (1, {'@': 146}), 10: (1, {'@': 146}), 11: (1, {'@': 146}), 12: (1, {'@': 146}), 13: (1, {'@': 146}), 14: (1, {'@': 146}), 15: (1, {'@': 146}), 16: (1, {'@': 146}), 17: (1, {'@': 146}), 18: (1, {'@': 146}), 19: (1, {'@': 146}), 20: (1, {'@': 146}), 21: (1, {'@': 146}), 22: (1, {'@': 146}), 23: (1, {'@': 146}), 24: (1, {'@': 146}), 25: (1, {'@': 146}), 26: (1, {'@': 146}), 27: (1, {'@': 146})}, 254: {27: (0, 345), 31: (0, 245)}, 255: {29: (0, 406), 30: (0, 399), 27: (0, 229), 31: (0, 336), 32: (0, 251)}, 256: {27: (1, {'@': 253}), 31: (1, {'@': 253}), 58: (1, {'@': 253}), 71: (1, {'@': 253}), 28: (1, {'@': 253}), 47: (1, {'@': 253}), 83: (1, {'@': 253})}, 257: {10: (0, 359)}, 258: {18: (0, 266)}, 259: {0: (1, {'@': 137}), 1: (1, {'@': 137}), 2: (1, {'@': 137}), 3: (1, {'@': 137}), 4: (1, {'@': 137}), 5: (1, {'@': 137}), 6: (1, {'@': 137}), 7: (1, {'@': 137}), 8: (1, {'@': 137}), 9: (1, {'@': 137}), 10: (1, {'@': 137}), 11: (1, {'@': 137}), 12: (1, {'@': 137}), 13: (1, {'@': 137}), 14: (1, {'@': 137}), 15: (1, {'@': 137}), 16: (1, {'@': 137}), 17: (1, {'@': 137}), 18: (1, {'@': 137}), 19: (1, {'@': 137}), 20: (1, {'@': 137}), 21: (1, {'@': 137}), 22: (1, {'@': 137}), 23: (1, {'@': 137}), 24: (1, {'@': 137}), 25: (1, {'@': 137}), 26: (1, {'@': 137}), 27: (1, {'@': 137})}, 260: {92: (0, 248), 93: (0, 464), 94: (0, 505), 95: (0, 252), 96: (0, 228), 10: (0, 63), 97: (0, 185), 98: (0, 422), 99: (0, 210), 101: (0, 256), 102: (0, 260), 103: (0, 457), 110: (0, 227), 104: (0, 207), 105: (0, 237), 106: (0, 484), 107: (0, 208), 109: (0, 243), 108: (0, 408), 100: (0, 110)}, 261: {135: (0, 289)}, 262: {82: (0, 356), 83: (0, 354), 10: (1, {'@': 278}), 29: (1, {'@': 274}), 30: (1, {'@': 274}), 47: (1, {'@': 274}), 27: (1, {'@': 274}), 28: (1, {'@': 274}), 31: (1, {'@': 274}), 48: (1, {'@': 274}), 49: (1, {'@': 274}), 50: (1, {'@': 274}), 51: (1, {'@': 274}), 52: (1, {'@': 274}), 53: (1, {'@': 274}), 54: (1, {'@': 274}), 55: (1, {'@': 274}), 56: (1, {'@': 274}), 57: (1, {'@': 274}), 58: (1, {'@': 274}), 59: (1, {'@': 274}), 60: (1, {'@': 274}), 61: (1, {'@': 274}), 62: (1, {'@': 274}), 63: (1, {'@': 274}), 64: (1, {'@': 274}), 65: (1, {'@': 274}), 66: (1, {'@': 274}), 148: (1, {'@': 278}), 71: (1, {'@': 274}), 67: (1, {'@': 274}), 68: (1, {'@': 274}), 69: (1, {'@': 274}), 18: (1, {'@': 274}), 70: (1, {'@': 274})}, 263: {33: (0, 427), 34: (0, 282), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 264: {28: (0, 215)}, 265: {29: (1, {'@': 227}), 30: (1, {'@': 227}), 47: (1, {'@': 227}), 27: (1, {'@': 227}), 28: (1, {'@': 227}), 31: (1, {'@': 227}), 48: (1, {'@': 227}), 49: (1, {'@': 227}), 50: (1, {'@': 227}), 51: (1, {'@': 227}), 52: (1, {'@': 227}), 53: (1, {'@': 227}), 54: (1, {'@': 227}), 55: (1, {'@': 227}), 56: (1, {'@': 227}), 57: (1, {'@': 227}), 58: (1, {'@': 227}), 59: (1, {'@': 227}), 60: (1, {'@': 227}), 61: (1, {'@': 227}), 62: (1, {'@': 227}), 63: (1, {'@': 227}), 64: (1, {'@': 227}), 65: (1, {'@': 227}), 66: (1, {'@': 227}), 67: (1, {'@': 227}), 68: (1, {'@': 227}), 69: (1, {'@': 227}), 18: (1, {'@': 227}), 70: (1, {'@': 227})}, 266: {0: (1, {'@': 140}), 1: (1, {'@': 140}), 2: (1, {'@': 140}), 3: (1, {'@': 140}), 4: (1, {'@': 140}), 5: (1, {'@': 140}), 6: (1, {'@': 140}), 7: (1, {'@': 140}), 8: (1, {'@': 140}), 9: (1, {'@': 140}), 10: (1, {'@': 140}), 11: (1, {'@': 140}), 12: (1, {'@': 140}), 13: (1, {'@': 140}), 14: (1, {'@': 140}), 15: (1, {'@': 140}), 16: (1, {'@': 140}), 17: (1, {'@': 140}), 18: (1, {'@': 140}), 19: (1, {'@': 140}), 20: (1, {'@': 140}), 21: (1, {'@': 140}), 22: (1, {'@': 140}), 23: (1, {'@': 140}), 24: (1, {'@': 140}), 25: (1, {'@': 140}), 26: (1, {'@': 140}), 27: (1, {'@': 140})}, 267: {29: (1, {'@': 228}), 30: (1, {'@': 228}), 47: (1, {'@': 228}), 27: (1, {'@': 228}), 28: (1, {'@': 228}), 31: (1, {'@': 228}), 48: (1, {'@': 228}), 49: (1, {'@': 228}), 50: (1, {'@': 228}), 51: (1, {'@': 228}), 52: (1, {'@': 228}), 53: (1, {'@': 228}), 54: (1, {'@': 228}), 55: (1, {'@': 228}), 56: (1, {'@': 228}), 57: (1, {'@': 228}), 58: (1, {'@': 228}), 59: (1, {'@': 228}), 60: (1, {'@': 228}), 61: (1, {'@': 228}), 62: (1, {'@': 228}), 63: (1, {'@': 228}), 64: (1, {'@': 228}), 65: (1, {'@': 228}), 66: (1, {'@': 228}), 67: (1, {'@': 228}), 68: (1, {'@': 228}), 69: (1, {'@': 228}), 18: (1, {'@': 228}), 70: (1, {'@': 228})}, 268: {29: (1, {'@': 229}), 30: (1, {'@': 229}), 47: (1, {'@': 229}), 27: (1, {'@': 229}), 28: (1, {'@': 229}), 31: (1, {'@': 229}), 48: (1, {'@': 229}), 49: (1, {'@': 229}), 50: (1, {'@': 229}), 51: (1, {'@': 229}), 52: (1, {'@': 229}), 53: (1, {'@': 229}), 54: (1, {'@': 229}), 55: (1, {'@': 229}), 56: (1, {'@': 229}), 57: (1, {'@': 229}), 58: (1, {'@': 229}), 59: (1, {'@': 229}), 60: (1, {'@': 229}), 61: (1, {'@': 229}), 62: (1, {'@': 229}), 63: (1, {'@': 229}), 64: (1, {'@': 229}), 65: (1, {'@': 229}), 66: (1, {'@': 229}), 67: (1, {'@': 229}), 68: (1, {'@': 229}), 69: (1, {'@': 229}), 18: (1, {'@': 229}), 70: (1, {'@': 229})}, 269: {19: (0, 437), 79: (0, 340), 33: (0, 427), 38: (0, 397), 23: (0, 23), 5: (0, 292), 26: (0, 47), 17: (0, 494), 7: (0, 44), 20: (0, 438), 115: (0, 28), 11: (0, 40), 116: (0, 37), 21: (0, 30), 37: (0, 488), 34: (0, 2), 117: (0, 0), 24: (0, 4), 10: (0, 51), 46: (0, 11), 118: (0, 20), 119: (0, 33), 39: (0, 468), 8: (0, 24), 2: (0, 142), 40: (0, 201), 120: (0, 152), 121: (0, 150), 122: (0, 161), 41: (0, 180), 42: (0, 273), 16: (0, 193), 123: (0, 158), 43: (0, 164), 124: (0, 141), 125: (0, 135), 12: (0, 129), 4: (0, 124), 25: (0, 442), 9: (0, 112), 126: (0, 154), 127: (0, 147), 35: (0, 178), 36: (0, 232), 0: (0, 268), 1: (0, 288), 6: (0, 249), 128: (0, 133), 129: (0, 294), 130: (0, 303), 15: (0, 257), 131: (0, 224), 22: (0, 262), 14: (0, 75), 3: (0, 265), 13: (0, 267), 80: (0, 244), 44: (0, 274), 132: (0, 220), 45: (0, 504), 133: (0, 218), 18: (1, {'@': 98})}, 270: {0: (1, {'@': 150}), 1: (1, {'@': 150}), 2: (1, {'@': 150}), 3: (1, {'@': 150}), 4: (1, {'@': 150}), 5: (1, {'@': 150}), 6: (1, {'@': 150}), 7: (1, {'@': 150}), 8: (1, {'@': 150}), 9: (1, {'@': 150}), 10: (1, {'@': 150}), 11: (1, {'@': 150}), 12: (1, {'@': 150}), 13: (1, {'@': 150}), 14: (1, {'@': 150}), 15: (1, {'@': 150}), 16: (1, {'@': 150}), 17: (1, {'@': 150}), 18: (1, {'@': 150}), 19: (1, {'@': 150}), 20: (1, {'@': 150}), 21: (1, {'@': 150}), 22: (1, {'@': 150}), 23: (1, {'@': 150}), 24: (1, {'@': 150}), 25: (1, {'@': 150}), 26: (1, {'@': 150}), 27: (1, {'@': 150})}, 271: {27: (1, {'@': 262}), 31: (1, {'@': 262})}, 272: {10: (0, 348)}, 273: {29: (1, {'@': 191}), 30: (1, {'@': 191}), 47: (1, {'@': 191}), 27: (1, {'@': 191}), 28: (1, {'@': 191}), 31: (1, {'@': 191}), 48: (1, {'@': 191}), 49: (1, {'@': 191}), 50: (1, {'@': 191}), 51: (1, {'@': 191}), 52: (1, {'@': 191}), 53: (1, {'@': 191}), 54: (1, {'@': 191}), 55: (1, {'@': 191}), 56: (1, {'@': 191}), 57: (1, {'@': 191}), 58: (1, {'@': 191}), 59: (1, {'@': 191}), 60: (1, {'@': 191}), 61: (1, {'@': 191}), 62: (1, {'@': 191}), 63: (1, {'@': 191}), 64: (1, {'@': 191}), 65: (1, {'@': 191}), 66: (1, {'@': 191}), 67: (1, {'@': 191}), 68: (1, {'@': 191}), 69: (1, {'@': 191}), 18: (1, {'@': 191}), 70: (1, {'@': 191})}, 274: {29: (1, {'@': 195}), 30: (1, {'@': 195}), 47: (1, {'@': 195}), 27: (1, {'@': 195}), 28: (1, {'@': 195}), 31: (1, {'@': 195}), 48: (1, {'@': 195}), 49: (1, {'@': 195}), 50: (1, {'@': 195}), 51: (1, {'@': 195}), 52: (1, {'@': 195}), 53: (1, {'@': 195}), 54: (1, {'@': 195}), 55: (1, {'@': 195}), 56: (1, {'@': 195}), 57: (1, {'@': 195}), 58: (1, {'@': 195}), 59: (1, {'@': 195}), 60: (1, {'@': 195}), 61: (1, {'@': 195}), 62: (1, {'@': 195}), 63: (1, {'@': 195}), 64: (1, {'@': 195}), 65: (1, {'@': 195}), 66: (1, {'@': 195}), 67: (1, {'@': 195}), 68: (1, {'@': 195}), 69: (1, {'@': 195}), 18: (1, {'@': 195}), 70: (1, {'@': 195})}, 275: {29: (1, {'@': 217}), 30: (1, {'@': 217}), 47: (1, {'@': 217}), 27: (1, {'@': 217}), 28: (1, {'@': 217}), 31: (1, {'@': 217}), 48: (1, {'@': 217}), 49: (1, {'@': 217}), 50: (1, {'@': 217}), 51: (1, {'@': 217}), 52: (1, {'@': 217}), 53: (1, {'@': 217}), 54: (1, {'@': 217}), 55: (1, {'@': 217}), 56: (1, {'@': 217}), 57: (1, {'@': 217}), 58: (1, {'@': 217}), 59: (1, {'@': 217}), 60: (1, {'@': 217}), 61: (1, {'@': 217}), 62: (1, {'@': 217}), 63: (1, {'@': 217}), 64: (1, {'@': 217}), 65: (1, {'@': 217}), 66: (1, {'@': 217}), 67: (1, {'@': 217}), 68: (1, {'@': 217}), 69: (1, {'@': 217}), 18: (1, {'@': 217}), 70: (1, {'@': 217})}, 276: {135: (0, 128)}, 277: {93: (0, 464), 138: (0, 510), 33: (0, 427), 98: (0, 422), 26: (0, 47), 17: (0, 494), 20: (0, 438), 94: (0, 505), 37: (0, 488), 10: (0, 460), 103: (0, 457), 34: (0, 430), 38: (0, 402), 106: (0, 484), 108: (0, 408), 39: (0, 468), 110: (0, 227), 40: (0, 201), 92: (0, 248), 41: (0, 180), 42: (0, 273), 16: (0, 193), 97: (0, 185), 139: (0, 293), 101: (0, 256), 43: (0, 223), 12: (0, 203), 4: (0, 188), 96: (0, 228), 140: (0, 213), 35: (0, 178), 141: (0, 239), 99: (0, 210), 36: (0, 232), 104: (0, 207), 105: (0, 237), 0: (0, 268), 6: (0, 249), 95: (0, 252), 15: (0, 257), 102: (0, 260), 22: (0, 262), 3: (0, 265), 13: (0, 267), 142: (0, 271), 107: (0, 208), 44: (0, 274), 109: (0, 243), 25: (0, 442), 45: (0, 504), 100: (0, 516), 46: (0, 416)}, 278: {47: (0, 177)}, 279: {31: (1, {'@': 163})}, 280: {82: (0, 313), 83: (0, 354)}, 281: {88: (0, 95), 89: (0, 367), 22: (0, 99)}, 282: {29: (0, 406), 30: (0, 399), 28: (0, 165)}, 283: {112: (0, 90), 3: (0, 316)}, 284: {82: (0, 356), 83: (0, 354), 10: (1, {'@': 278})}, 285: {18: (0, 170)}, 286: {22: (1, {'@': 289}), 27: (1, {'@': 289})}, 287: {113: (1, {'@': 86})}, 288: {28: (0, 122)}, 289: {72: (0, 156)}, 290: {33: (0, 427), 34: (0, 118), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 291: {0: (1, {'@': 144}), 1: (1, {'@': 144}), 2: (1, {'@': 144}), 3: (1, {'@': 144}), 4: (1, {'@': 144}), 5: (1, {'@': 144}), 6: (1, {'@': 144}), 7: (1, {'@': 144}), 8: (1, {'@': 144}), 9: (1, {'@': 144}), 10: (1, {'@': 144}), 11: (1, {'@': 144}), 12: (1, {'@': 144}), 13: (1, {'@': 144}), 14: (1, {'@': 144}), 15: (1, {'@': 144}), 16: (1, {'@': 144}), 17: (1, {'@': 144}), 18: (1, {'@': 144}), 19: (1, {'@': 144}), 20: (1, {'@': 144}), 21: (1, {'@': 144}), 22: (1, {'@': 144}), 23: (1, {'@': 144}), 24: (1, {'@': 144}), 25: (1, {'@': 144}), 26: (1, {'@': 144}), 27: (1, {'@': 144})}, 292: {88: (0, 95), 89: (0, 104), 22: (0, 99)}, 293: {10: (0, 310)}, 294: {0: (1, {'@': 100}), 1: (1, {'@': 100}), 2: (1, {'@': 100}), 3: (1, {'@': 100}), 4: (1, {'@': 100}), 5: (1, {'@': 100}), 6: (1, {'@': 100}), 7: (1, {'@': 100}), 8: (1, {'@': 100}), 9: (1, {'@': 100}), 10: (1, {'@': 100}), 11: (1, {'@': 100}), 12: (1, {'@': 100}), 13: (1, {'@': 100}), 14: (1, {'@': 100}), 15: (1, {'@': 100}), 16: (1, {'@': 100}), 17: (1, {'@': 100}), 18: (1, {'@': 100}), 19: (1, {'@': 100}), 20: (1, {'@': 100}), 21: (1, {'@': 100}), 22: (1, {'@': 100}), 23: (1, {'@': 100}), 24: (1, {'@': 100}), 25: (1, {'@': 100}), 26: (1, {'@': 100}), 27: (1, {'@': 100})}, 295: {47: (0, 157)}, 296: {75: (1, {'@': 282}), 5: (1, {'@': 282}), 18: (1, {'@': 282}), 14: (1, {'@': 282})}, 297: {0: (1, {'@': 159}), 1: (1, {'@': 159}), 2: (1, {'@': 159}), 3: (1, {'@': 159}), 4: (1, {'@': 159}), 5: (1, {'@': 159}), 6: (1, {'@': 159}), 7: (1, {'@': 159}), 8: (1, {'@': 159}), 9: (1, {'@': 159}), 10: (1, {'@': 159}), 11: (1, {'@': 159}), 12: (1, {'@': 159}), 13: (1, {'@': 159}), 14: (1, {'@': 159}), 15: (1, {'@': 159}), 16: (1, {'@': 159}), 17: (1, {'@': 159}), 18: (1, {'@': 159}), 19: (1, {'@': 159}), 20: (1, {'@': 159}), 21: (1, {'@': 159}), 22: (1, {'@': 159}), 23: (1, {'@': 159}), 24: (1, {'@': 159}), 25: (1, {'@': 159}), 26: (1, {'@': 159}), 27: (1, {'@': 159})}, 298: {149: (0, 343), 83: (0, 354), 150: (0, 35), 82: (0, 214), 47: (0, 191)}, 299: {28: (0, 196)}, 300: {83: (0, 354), 47: (0, 139), 82: (0, 163)}, 301: {27: (0, 187), 31: (0, 245)}, 302: {29: (1, {'@': 220}), 30: (1, {'@': 220}), 47: (1, {'@': 220}), 27: (1, {'@': 220}), 28: (1, {'@': 220}), 31: (1, {'@': 220}), 48: (1, {'@': 220}), 49: (1, {'@': 220}), 50: (1, {'@': 220}), 51: (1, {'@': 220}), 52: (1, {'@': 220}), 53: (1, {'@': 220}), 54: (1, {'@': 220}), 55: (1, {'@': 220}), 56: (1, {'@': 220}), 57: (1, {'@': 220}), 58: (1, {'@': 220}), 59: (1, {'@': 220}), 60: (1, {'@': 220}), 61: (1, {'@': 220}), 62: (1, {'@': 220}), 63: (1, {'@': 220}), 64: (1, {'@': 220}), 65: (1, {'@': 220}), 66: (1, {'@': 220}), 67: (1, {'@': 220}), 68: (1, {'@': 220}), 69: (1, {'@': 220}), 18: (1, {'@': 220}), 70: (1, {'@': 220})}, 303: {0: (1, {'@': 101}), 1: (1, {'@': 101}), 2: (1, {'@': 101}), 3: (1, {'@': 101}), 4: (1, {'@': 101}), 5: (1, {'@': 101}), 6: (1, {'@': 101}), 7: (1, {'@': 101}), 8: (1, {'@': 101}), 9: (1, {'@': 101}), 10: (1, {'@': 101}), 11: (1, {'@': 101}), 12: (1, {'@': 101}), 13: (1, {'@': 101}), 14: (1, {'@': 101}), 15: (1, {'@': 101}), 16: (1, {'@': 101}), 17: (1, {'@': 101}), 18: (1, {'@': 101}), 19: (1, {'@': 101}), 20: (1, {'@': 101}), 21: (1, {'@': 101}), 22: (1, {'@': 101}), 23: (1, {'@': 101}), 24: (1, {'@': 101}), 25: (1, {'@': 101}), 26: (1, {'@': 101}), 27: (1, {'@': 101})}, 304: {18: (0, 387)}, 305: {0: (1, {'@': 160}), 1: (1, {'@': 160}), 2: (1, {'@': 160}), 3: (1, {'@': 160}), 4: (1, {'@': 160}), 5: (1, {'@': 160}), 6: (1, {'@': 160}), 7: (1, {'@': 160}), 8: (1, {'@': 160}), 9: (1, {'@': 160}), 10: (1, {'@': 160}), 11: (1, {'@': 160}), 12: (1, {'@': 160}), 13: (1, {'@': 160}), 14: (1, {'@': 160}), 15: (1, {'@': 160}), 16: (1, {'@': 160}), 17: (1, {'@': 160}), 18: (1, {'@': 160}), 19: (1, {'@': 160}), 20: (1, {'@': 160}), 21: (1, {'@': 160}), 22: (1, {'@': 160}), 23: (1, {'@': 160}), 24: (1, {'@': 160}), 25: (1, {'@': 160}), 26: (1, {'@': 160}), 27: (1, {'@': 160})}, 306: {28: (0, 346)}, 307: {0: (1, {'@': 161}), 1: (1, {'@': 161}), 2: (1, {'@': 161}), 3: (1, {'@': 161}), 4: (1, {'@': 161}), 5: (1, {'@': 161}), 6: (1, {'@': 161}), 7: (1, {'@': 161}), 8: (1, {'@': 161}), 9: (1, {'@': 161}), 10: (1, {'@': 161}), 11: (1, {'@': 161}), 12: (1, {'@': 161}), 13: (1, {'@': 161}), 14: (1, {'@': 161}), 15: (1, {'@': 161}), 16: (1, {'@': 161}), 17: (1, {'@': 161}), 18: (1, {'@': 161}), 19: (1, {'@': 161}), 20: (1, {'@': 161}), 21: (1, {'@': 161}), 22: (1, {'@': 161}), 23: (1, {'@': 161}), 24: (1, {'@': 161}), 25: (1, {'@': 161}), 26: (1, {'@': 161}), 27: (1, {'@': 161})}, 308: {0: (1, {'@': 171}), 15: (1, {'@': 171}), 16: (1, {'@': 171}), 3: (1, {'@': 171}), 4: (1, {'@': 171}), 6: (1, {'@': 171}), 17: (1, {'@': 171}), 20: (1, {'@': 171}), 13: (1, {'@': 171}), 22: (1, {'@': 171}), 10: (1, {'@': 171}), 12: (1, {'@': 171}), 25: (1, {'@': 171}), 26: (1, {'@': 171})}, 309: {28: (0, 374)}, 310: {112: (0, 322), 3: (0, 316), 27: (0, 320)}, 311: {75: (1, {'@': 116}), 5: (1, {'@': 116}), 18: (1, {'@': 116}), 14: (1, {'@': 116}), 0: (1, {'@': 116}), 1: (1, {'@': 116}), 2: (1, {'@': 116}), 3: (1, {'@': 116}), 4: (1, {'@': 116}), 6: (1, {'@': 116}), 7: (1, {'@': 116}), 8: (1, {'@': 116}), 9: (1, {'@': 116}), 10: (1, {'@': 116}), 11: (1, {'@': 116}), 12: (1, {'@': 116}), 13: (1, {'@': 116}), 15: (1, {'@': 116}), 16: (1, {'@': 116}), 17: (1, {'@': 116}), 19: (1, {'@': 116}), 20: (1, {'@': 116}), 21: (1, {'@': 116}), 22: (1, {'@': 116}), 23: (1, {'@': 116}), 24: (1, {'@': 116}), 25: (1, {'@': 116}), 26: (1, {'@': 116}), 27: (1, {'@': 116})}, 312: {29: (0, 406), 30: (0, 399), 31: (0, 42)}, 313: {47: (0, 319)}, 314: {47: (0, 184)}, 315: {18: (0, 140)}, 316: {69: (0, 277)}, 317: {0: (1, {'@': 149}), 1: (1, {'@': 149}), 2: (1, {'@': 149}), 3: (1, {'@': 149}), 4: (1, {'@': 149}), 5: (1, {'@': 149}), 6: (1, {'@': 149}), 7: (1, {'@': 149}), 8: (1, {'@': 149}), 9: (1, {'@': 149}), 10: (1, {'@': 149}), 11: (1, {'@': 149}), 12: (1, {'@': 149}), 13: (1, {'@': 149}), 14: (1, {'@': 149}), 15: (1, {'@': 149}), 16: (1, {'@': 149}), 17: (1, {'@': 149}), 18: (1, {'@': 149}), 19: (1, {'@': 149}), 20: (1, {'@': 149}), 21: (1, {'@': 149}), 22: (1, {'@': 149}), 23: (1, {'@': 149}), 24: (1, {'@': 149}), 25: (1, {'@': 149}), 26: (1, {'@': 149}), 27: (1, {'@': 149})}, 318: {28: (0, 297)}, 319: {5: (0, 292), 79: (0, 340), 81: (0, 341), 80: (0, 296), 82: (0, 352), 75: (0, 70), 83: (0, 354), 85: (0, 66), 151: (0, 48), 18: (0, 80), 84: (0, 64), 14: (0, 75)}, 320: {27: (1, {'@': 269}), 31: (1, {'@': 269}), 28: (1, {'@': 269})}, 321: {18: (0, 148)}, 322: {31: (0, 283), 152: (0, 335), 27: (0, 61)}, 323: {29: (0, 406), 30: (0, 399), 27: (0, 46)}, 324: {73: (0, 18), 74: (0, 250), 18: (0, 355), 3: (0, 200)}, 325: {3: (0, 34)}, 326: {28: (0, 333)}, 327: {47: (0, 332)}, 328: {18: (0, 171)}, 329: {0: (1, {'@': 151}), 1: (1, {'@': 151}), 2: (1, {'@': 151}), 3: (1, {'@': 151}), 4: (1, {'@': 151}), 5: (1, {'@': 151}), 6: (1, {'@': 151}), 7: (1, {'@': 151}), 8: (1, {'@': 151}), 9: (1, {'@': 151}), 10: (1, {'@': 151}), 11: (1, {'@': 151}), 12: (1, {'@': 151}), 13: (1, {'@': 151}), 14: (1, {'@': 151}), 15: (1, {'@': 151}), 16: (1, {'@': 151}), 17: (1, {'@': 151}), 18: (1, {'@': 151}), 19: (1, {'@': 151}), 20: (1, {'@': 151}), 21: (1, {'@': 151}), 22: (1, {'@': 151}), 23: (1, {'@': 151}), 24: (1, {'@': 151}), 25: (1, {'@': 151}), 26: (1, {'@': 151}), 27: (1, {'@': 151})}, 330: {33: (0, 427), 34: (0, 202), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 331: {33: (0, 427), 34: (0, 125), 145: (0, 371), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 88: (0, 362), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 153: (0, 389), 22: (0, 370), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 27: (0, 377), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 332: {72: (0, 276)}, 333: {0: (1, {'@': 147}), 1: (1, {'@': 147}), 2: (1, {'@': 147}), 3: (1, {'@': 147}), 4: (1, {'@': 147}), 5: (1, {'@': 147}), 6: (1, {'@': 147}), 7: (1, {'@': 147}), 8: (1, {'@': 147}), 9: (1, {'@': 147}), 10: (1, {'@': 147}), 11: (1, {'@': 147}), 12: (1, {'@': 147}), 13: (1, {'@': 147}), 14: (1, {'@': 147}), 15: (1, {'@': 147}), 16: (1, {'@': 147}), 17: (1, {'@': 147}), 18: (1, {'@': 147}), 19: (1, {'@': 147}), 20: (1, {'@': 147}), 21: (1, {'@': 147}), 22: (1, {'@': 147}), 23: (1, {'@': 147}), 24: (1, {'@': 147}), 25: (1, {'@': 147}), 26: (1, {'@': 147}), 27: (1, {'@': 147})}, 334: {18: (1, {'@': 293}), 31: (1, {'@': 293})}, 335: {31: (0, 77), 27: (0, 499)}, 336: {34: (0, 126), 33: (0, 427), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 337: {10: (0, 176)}, 338: {29: (1, {'@': 216}), 30: (1, {'@': 216}), 47: (1, {'@': 216}), 27: (1, {'@': 216}), 28: (1, {'@': 216}), 31: (1, {'@': 216}), 48: (1, {'@': 216}), 49: (1, {'@': 216}), 50: (1, {'@': 216}), 51: (1, {'@': 216}), 52: (1, {'@': 216}), 53: (1, {'@': 216}), 54: (1, {'@': 216}), 55: (1, {'@': 216}), 56: (1, {'@': 216}), 57: (1, {'@': 216}), 58: (1, {'@': 216}), 59: (1, {'@': 216}), 60: (1, {'@': 216}), 61: (1, {'@': 216}), 62: (1, {'@': 216}), 63: (1, {'@': 216}), 64: (1, {'@': 216}), 65: (1, {'@': 216}), 66: (1, {'@': 216}), 67: (1, {'@': 216}), 68: (1, {'@': 216}), 69: (1, {'@': 216}), 18: (1, {'@': 216}), 70: (1, {'@': 216})}, 339: {47: (0, 88)}, 340: {5: (0, 98)}, 341: {5: (0, 292), 79: (0, 340), 75: (0, 70), 18: (0, 101), 84: (0, 94), 14: (0, 75), 80: (0, 55)}, 342: {147: (0, 144)}, 343: {47: (0, 247)}, 344: {0: (1, {'@': 148}), 1: (1, {'@': 148}), 2: (1, {'@': 148}), 3: (1, {'@': 148}), 4: (1, {'@': 148}), 5: (1, {'@': 148}), 6: (1, {'@': 148}), 7: (1, {'@': 148}), 8: (1, {'@': 148}), 9: (1, {'@': 148}), 10: (1, {'@': 148}), 11: (1, {'@': 148}), 12: (1, {'@': 148}), 13: (1, {'@': 148}), 14: (1, {'@': 148}), 15: (1, {'@': 148}), 16: (1, {'@': 148}), 17: (1, {'@': 148}), 18: (1, {'@': 148}), 19: (1, {'@': 148}), 20: (1, {'@': 148}), 21: (1, {'@': 148}), 22: (1, {'@': 148}), 23: (1, {'@': 148}), 24: (1, {'@': 148}), 25: (1, {'@': 148}), 26: (1, {'@': 148}), 27: (1, {'@': 148})}, 345: {28: (0, 307)}, 346: {0: (1, {'@': 142}), 1: (1, {'@': 142}), 2: (1, {'@': 142}), 3: (1, {'@': 142}), 4: (1, {'@': 142}), 5: (1, {'@': 142}), 6: (1, {'@': 142}), 7: (1, {'@': 142}), 8: (1, {'@': 142}), 9: (1, {'@': 142}), 10: (1, {'@': 142}), 11: (1, {'@': 142}), 12: (1, {'@': 142}), 13: (1, {'@': 142}), 14: (1, {'@': 142}), 15: (1, {'@': 142}), 16: (1, {'@': 142}), 17: (1, {'@': 142}), 18: (1, {'@': 142}), 19: (1, {'@': 142}), 20: (1, {'@': 142}), 21: (1, {'@': 142}), 22: (1, {'@': 142}), 23: (1, {'@': 142}), 24: (1, {'@': 142}), 25: (1, {'@': 142}), 26: (1, {'@': 142}), 27: (1, {'@': 142})}, 347: {29: (0, 406), 18: (0, 6), 30: (0, 399)}, 348: {19: (0, 437), 79: (0, 340), 33: (0, 427), 38: (0, 397), 23: (0, 23), 5: (0, 292), 26: (0, 47), 17: (0, 494), 7: (0, 44), 20: (0, 438), 115: (0, 28), 11: (0, 40), 116: (0, 37), 21: (0, 30), 37: (0, 488), 34: (0, 2), 117: (0, 0), 24: (0, 4), 10: (0, 51), 46: (0, 11), 118: (0, 20), 119: (0, 33), 39: (0, 468), 8: (0, 24), 2: (0, 142), 40: (0, 201), 120: (0, 152), 121: (0, 150), 122: (0, 161), 41: (0, 180), 42: (0, 273), 16: (0, 193), 123: (0, 158), 43: (0, 164), 124: (0, 141), 125: (0, 135), 12: (0, 129), 4: (0, 124), 25: (0, 442), 9: (0, 112), 126: (0, 154), 127: (0, 147), 35: (0, 178), 36: (0, 232), 0: (0, 268), 1: (0, 288), 6: (0, 249), 129: (0, 294), 130: (0, 303), 15: (0, 257), 131: (0, 224), 22: (0, 262), 14: (0, 75), 3: (0, 265), 13: (0, 267), 80: (0, 244), 44: (0, 274), 128: (0, 361), 132: (0, 220), 45: (0, 504), 133: (0, 218), 27: (1, {'@': 98})}, 349: {0: (1, {'@': 126}), 1: (1, {'@': 126}), 2: (1, {'@': 126}), 3: (1, {'@': 126}), 4: (1, {'@': 126}), 5: (1, {'@': 126}), 6: (1, {'@': 126}), 7: (1, {'@': 126}), 8: (1, {'@': 126}), 9: (1, {'@': 126}), 10: (1, {'@': 126}), 11: (1, {'@': 126}), 12: (1, {'@': 126}), 13: (1, {'@': 126}), 14: (1, {'@': 126}), 15: (1, {'@': 126}), 16: (1, {'@': 126}), 17: (1, {'@': 126}), 18: (1, {'@': 126}), 19: (1, {'@': 126}), 20: (1, {'@': 126}), 21: (1, {'@': 126}), 22: (1, {'@': 126}), 23: (1, {'@': 126}), 24: (1, {'@': 126}), 25: (1, {'@': 126}), 26: (1, {'@': 126}), 27: (1, {'@': 126})}, 350: {154: (0, 287), 155: (0, 280)}, 351: {}, 352: {71: (0, 379)}, 353: {31: (0, 281), 27: (1, {'@': 95})}, 354: {22: (0, 369)}, 355: {18: (0, 7)}, 356: {10: (1, {'@': 277}), 148: (1, {'@': 277})}, 357: {27: (0, 407), 31: (0, 375)}, 358: {29: (0, 406), 30: (0, 399), 27: (0, 436)}, 359: {33: (0, 427), 34: (0, 384), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 360: {91: (0, 388)}, 361: {27: (0, 299)}, 362: {71: (0, 434)}, 363: {33: (0, 427), 34: (0, 43), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 364: {93: (0, 464), 138: (0, 510), 33: (0, 427), 98: (0, 422), 26: (0, 47), 17: (0, 494), 20: (0, 438), 94: (0, 505), 37: (0, 488), 10: (0, 460), 103: (0, 457), 34: (0, 430), 38: (0, 402), 106: (0, 484), 108: (0, 408), 39: (0, 468), 110: (0, 227), 40: (0, 201), 92: (0, 248), 41: (0, 180), 42: (0, 273), 16: (0, 193), 97: (0, 185), 139: (0, 293), 101: (0, 256), 43: (0, 223), 12: (0, 203), 4: (0, 188), 96: (0, 228), 140: (0, 213), 35: (0, 178), 99: (0, 210), 141: (0, 508), 36: (0, 232), 104: (0, 207), 105: (0, 237), 0: (0, 268), 6: (0, 249), 95: (0, 252), 15: (0, 257), 102: (0, 260), 22: (0, 262), 3: (0, 265), 13: (0, 267), 142: (0, 271), 107: (0, 208), 44: (0, 274), 109: (0, 243), 25: (0, 442), 45: (0, 504), 100: (0, 516), 46: (0, 416)}, 365: {3: (0, 314)}, 366: {88: (0, 95), 89: (0, 390), 22: (0, 99)}, 367: {27: (1, {'@': 286}), 31: (1, {'@': 286})}, 368: {29: (0, 406), 28: (0, 396), 30: (0, 399)}, 369: {71: (1, {'@': 260}), 0: (1, {'@': 260}), 1: (1, {'@': 260}), 4: (1, {'@': 260}), 8: (1, {'@': 260}), 10: (1, {'@': 260}), 11: (1, {'@': 260}), 15: (1, {'@': 260}), 19: (1, {'@': 260}), 20: (1, {'@': 260}), 22: (1, {'@': 260}), 24: (1, {'@': 260}), 26: (1, {'@': 260}), 2: (1, {'@': 260}), 3: (1, {'@': 260}), 5: (1, {'@': 260}), 6: (1, {'@': 260}), 7: (1, {'@': 260}), 27: (1, {'@': 260}), 9: (1, {'@': 260}), 12: (1, {'@': 260}), 13: (1, {'@': 260}), 14: (1, {'@': 260}), 16: (1, {'@': 260}), 75: (1, {'@': 260}), 17: (1, {'@': 260}), 18: (1, {'@': 260}), 21: (1, {'@': 260}), 23: (1, {'@': 260}), 25: (1, {'@': 260}), 47: (1, {'@': 260}), 148: (1, {'@': 260})}, 370: {82: (0, 356), 83: (0, 354), 10: (1, {'@': 278}), 71: (1, {'@': 276}), 29: (1, {'@': 274}), 30: (1, {'@': 274}), 27: (1, {'@': 274}), 31: (1, {'@': 274})}, 371: {134: (0, 490), 31: (0, 405), 27: (1, {'@': 211})}, 372: {29: (0, 406), 69: (0, 421), 30: (0, 399), 156: (0, 439), 31: (0, 428), 68: (0, 462)}, 373: {91: (0, 425)}, 374: {14: (1, {'@': 261}), 18: (1, {'@': 261}), 83: (1, {'@': 261}), 5: (1, {'@': 261}), 75: (1, {'@': 261}), 155: (1, {'@': 261})}, 375: {92: (0, 248), 93: (0, 464), 94: (0, 505), 95: (0, 252), 96: (0, 228), 10: (0, 63), 97: (0, 185), 98: (0, 422), 99: (0, 210), 101: (0, 256), 102: (0, 260), 103: (0, 457), 104: (0, 207), 105: (0, 237), 106: (0, 484), 100: (0, 113), 107: (0, 208), 109: (0, 243), 108: (0, 408), 110: (0, 227)}, 376: {33: (0, 427), 34: (0, 443), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416), 31: (1, {'@': 205}), 68: (1, {'@': 205})}, 377: {29: (1, {'@': 207}), 30: (1, {'@': 207}), 47: (1, {'@': 207}), 27: (1, {'@': 207}), 28: (1, {'@': 207}), 31: (1, {'@': 207}), 48: (1, {'@': 207}), 49: (1, {'@': 207}), 50: (1, {'@': 207}), 51: (1, {'@': 207}), 52: (1, {'@': 207}), 53: (1, {'@': 207}), 54: (1, {'@': 207}), 55: (1, {'@': 207}), 56: (1, {'@': 207}), 57: (1, {'@': 207}), 58: (1, {'@': 207}), 59: (1, {'@': 207}), 60: (1, {'@': 207}), 61: (1, {'@': 207}), 62: (1, {'@': 207}), 63: (1, {'@': 207}), 64: (1, {'@': 207}), 65: (1, {'@': 207}), 66: (1, {'@': 207}), 67: (1, {'@': 207}), 68: (1, {'@': 207}), 69: (1, {'@': 207}), 18: (1, {'@': 207}), 150: (1, {'@': 207}), 83: (1, {'@': 207}), 70: (1, {'@': 207})}, 378: {29: (0, 406), 30: (0, 399), 27: (0, 338)}, 379: {139: (0, 293), 142: (0, 309)}, 380: {29: (1, {'@': 215}), 30: (1, {'@': 215}), 47: (1, {'@': 215}), 27: (1, {'@': 215}), 28: (1, {'@': 215}), 31: (1, {'@': 215}), 48: (1, {'@': 215}), 49: (1, {'@': 215}), 50: (1, {'@': 215}), 51: (1, {'@': 215}), 52: (1, {'@': 215}), 53: (1, {'@': 215}), 54: (1, {'@': 215}), 55: (1, {'@': 215}), 56: (1, {'@': 215}), 57: (1, {'@': 215}), 58: (1, {'@': 215}), 59: (1, {'@': 215}), 60: (1, {'@': 215}), 61: (1, {'@': 215}), 62: (1, {'@': 215}), 63: (1, {'@': 215}), 64: (1, {'@': 215}), 65: (1, {'@': 215}), 66: (1, {'@': 215}), 67: (1, {'@': 215}), 68: (1, {'@': 215}), 69: (1, {'@': 215}), 18: (1, {'@': 215}), 70: (1, {'@': 215})}, 381: {156: (0, 386), 31: (0, 428), 68: (0, 450)}, 382: {47: (0, 469)}, 383: {157: (0, 351), 83: (0, 354), 82: (0, 352), 85: (0, 350)}, 384: {29: (0, 406), 27: (0, 275), 30: (0, 399)}, 385: {82: (0, 411), 83: (0, 354), 75: (1, {'@': 119}), 5: (1, {'@': 119}), 18: (1, {'@': 119}), 14: (1, {'@': 119}), 0: (1, {'@': 119}), 1: (1, {'@': 119}), 2: (1, {'@': 119}), 3: (1, {'@': 119}), 4: (1, {'@': 119}), 6: (1, {'@': 119}), 7: (1, {'@': 119}), 8: (1, {'@': 119}), 9: (1, {'@': 119}), 10: (1, {'@': 119}), 11: (1, {'@': 119}), 12: (1, {'@': 119}), 13: (1, {'@': 119}), 15: (1, {'@': 119}), 16: (1, {'@': 119}), 17: (1, {'@': 119}), 19: (1, {'@': 119}), 20: (1, {'@': 119}), 21: (1, {'@': 119}), 22: (1, {'@': 119}), 23: (1, {'@': 119}), 24: (1, {'@': 119}), 25: (1, {'@': 119}), 26: (1, {'@': 119}), 27: (1, {'@': 119})}, 386: {31: (0, 151), 68: (0, 476)}, 387: {28: (0, 19)}, 388: {92: (0, 248), 93: (0, 464), 94: (0, 505), 95: (0, 252), 96: (0, 228), 10: (0, 63), 97: (0, 185), 98: (0, 422), 99: (0, 210), 101: (0, 256), 102: (0, 260), 103: (0, 457), 104: (0, 207), 105: (0, 237), 106: (0, 484), 107: (0, 208), 109: (0, 243), 108: (0, 408), 100: (0, 117), 110: (0, 227)}, 389: {27: (0, 456)}, 390: {27: (1, {'@': 285}), 31: (1, {'@': 285})}, 391: {0: (1, {'@': 187}), 15: (1, {'@': 187}), 16: (1, {'@': 187}), 3: (1, {'@': 187}), 4: (1, {'@': 187}), 6: (1, {'@': 187}), 17: (1, {'@': 187}), 20: (1, {'@': 187}), 13: (1, {'@': 187}), 22: (1, {'@': 187}), 10: (1, {'@': 187}), 12: (1, {'@': 187}), 25: (1, {'@': 187}), 26: (1, {'@': 187})}, 392: {33: (0, 427), 34: (0, 312), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 393: {0: (1, {'@': 185}), 15: (1, {'@': 185}), 16: (1, {'@': 185}), 3: (1, {'@': 185}), 4: (1, {'@': 185}), 6: (1, {'@': 185}), 17: (1, {'@': 185}), 20: (1, {'@': 185}), 13: (1, {'@': 185}), 22: (1, {'@': 185}), 10: (1, {'@': 185}), 12: (1, {'@': 185}), 25: (1, {'@': 185}), 26: (1, {'@': 185})}, 394: {75: (1, {'@': 92}), 5: (1, {'@': 92}), 18: (1, {'@': 92}), 14: (1, {'@': 92})}, 395: {31: (0, 175), 134: (0, 172), 27: (1, {'@': 208})}, 396: {83: (0, 354), 82: (0, 53), 75: (1, {'@': 115}), 5: (1, {'@': 115}), 18: (1, {'@': 115}), 14: (1, {'@': 115}), 0: (1, {'@': 115}), 1: (1, {'@': 115}), 2: (1, {'@': 115}), 3: (1, {'@': 115}), 4: (1, {'@': 115}), 6: (1, {'@': 115}), 7: (1, {'@': 115}), 8: (1, {'@': 115}), 9: (1, {'@': 115}), 10: (1, {'@': 115}), 11: (1, {'@': 115}), 12: (1, {'@': 115}), 13: (1, {'@': 115}), 15: (1, {'@': 115}), 16: (1, {'@': 115}), 17: (1, {'@': 115}), 19: (1, {'@': 115}), 20: (1, {'@': 115}), 21: (1, {'@': 115}), 22: (1, {'@': 115}), 23: (1, {'@': 115}), 24: (1, {'@': 115}), 25: (1, {'@': 115}), 26: (1, {'@': 115}), 27: (1, {'@': 115})}, 397: {10: (0, 331), 148: (0, 59)}, 398: {0: (1, {'@': 172}), 15: (1, {'@': 172}), 16: (1, {'@': 172}), 3: (1, {'@': 172}), 4: (1, {'@': 172}), 6: (1, {'@': 172}), 17: (1, {'@': 172}), 20: (1, {'@': 172}), 13: (1, {'@': 172}), 22: (1, {'@': 172}), 10: (1, {'@': 172}), 12: (1, {'@': 172}), 25: (1, {'@': 172}), 26: (1, {'@': 172})}, 399: {33: (0, 427), 34: (0, 372), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 69: (0, 376), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 137: (0, 381), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 400: {35: (0, 298), 22: (0, 284), 38: (0, 402)}, 401: {14: (1, {'@': 280}), 5: (1, {'@': 280}), 75: (1, {'@': 280}), 18: (1, {'@': 280}), 83: (1, {'@': 280})}, 402: {10: (0, 331)}, 403: {33: (0, 427), 34: (0, 378), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 404: {0: (1, {'@': 186}), 15: (1, {'@': 186}), 16: (1, {'@': 186}), 3: (1, {'@': 186}), 4: (1, {'@': 186}), 6: (1, {'@': 186}), 17: (1, {'@': 186}), 20: (1, {'@': 186}), 13: (1, {'@': 186}), 22: (1, {'@': 186}), 10: (1, {'@': 186}), 12: (1, {'@': 186}), 25: (1, {'@': 186}), 26: (1, {'@': 186})}, 405: {145: (0, 143), 22: (0, 99), 88: (0, 362)}, 406: {88: (0, 423), 22: (0, 99)}, 407: {91: (0, 480)}, 408: {27: (1, {'@': 256}), 31: (1, {'@': 256}), 58: (1, {'@': 256}), 71: (1, {'@': 256}), 28: (1, {'@': 256}), 47: (1, {'@': 256}), 83: (1, {'@': 256})}, 409: {143: (0, 121), 33: (0, 427), 34: (0, 146), 35: (0, 178), 26: (0, 47), 28: (0, 166), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 158: (0, 167), 0: (0, 268), 65: (0, 149), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 410: {92: (0, 248), 93: (0, 464), 94: (0, 505), 95: (0, 252), 96: (0, 228), 10: (0, 63), 97: (0, 185), 98: (0, 422), 99: (0, 210), 101: (0, 256), 102: (0, 260), 103: (0, 457), 104: (0, 207), 105: (0, 237), 106: (0, 484), 107: (0, 208), 100: (0, 491), 109: (0, 243), 108: (0, 408), 110: (0, 227)}, 411: {75: (1, {'@': 118}), 5: (1, {'@': 118}), 18: (1, {'@': 118}), 14: (1, {'@': 118}), 0: (1, {'@': 118}), 1: (1, {'@': 118}), 2: (1, {'@': 118}), 3: (1, {'@': 118}), 4: (1, {'@': 118}), 6: (1, {'@': 118}), 7: (1, {'@': 118}), 8: (1, {'@': 118}), 9: (1, {'@': 118}), 10: (1, {'@': 118}), 11: (1, {'@': 118}), 12: (1, {'@': 118}), 13: (1, {'@': 118}), 15: (1, {'@': 118}), 16: (1, {'@': 118}), 17: (1, {'@': 118}), 19: (1, {'@': 118}), 20: (1, {'@': 118}), 21: (1, {'@': 118}), 22: (1, {'@': 118}), 23: (1, {'@': 118}), 24: (1, {'@': 118}), 25: (1, {'@': 118}), 26: (1, {'@': 118}), 27: (1, {'@': 118})}, 412: {29: (0, 406), 59: (0, 84), 66: (0, 81), 30: (0, 399), 64: (0, 308), 27: (0, 302), 65: (0, 168), 57: (0, 455), 61: (0, 391), 50: (0, 393), 55: (0, 453), 159: (0, 413), 48: (0, 431), 49: (0, 420), 54: (0, 404), 52: (0, 485), 60: (0, 479), 63: (0, 495), 56: (0, 459), 53: (0, 398), 62: (0, 481), 51: (0, 497), 58: (0, 471)}, 413: {33: (0, 427), 34: (0, 498), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 414: {31: (0, 419), 27: (1, {'@': 213})}, 415: {28: (1, {'@': 303}), 31: (1, {'@': 303}), 65: (1, {'@': 303})}, 416: {29: (1, {'@': 226}), 30: (1, {'@': 226}), 47: (1, {'@': 226}), 27: (1, {'@': 226}), 28: (1, {'@': 226}), 31: (1, {'@': 226}), 48: (1, {'@': 226}), 49: (1, {'@': 226}), 50: (1, {'@': 226}), 51: (1, {'@': 226}), 52: (1, {'@': 226}), 53: (1, {'@': 226}), 54: (1, {'@': 226}), 55: (1, {'@': 226}), 56: (1, {'@': 226}), 57: (1, {'@': 226}), 58: (1, {'@': 226}), 59: (1, {'@': 226}), 60: (1, {'@': 226}), 61: (1, {'@': 226}), 62: (1, {'@': 226}), 63: (1, {'@': 226}), 64: (1, {'@': 226}), 65: (1, {'@': 226}), 66: (1, {'@': 226}), 67: (1, {'@': 226}), 68: (1, {'@': 226}), 69: (1, {'@': 226}), 18: (1, {'@': 226}), 70: (1, {'@': 226})}, 417: {33: (0, 427), 34: (0, 102), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 418: {65: (0, 92)}, 419: {145: (0, 153), 88: (0, 362), 22: (0, 99)}, 420: {0: (1, {'@': 182}), 15: (1, {'@': 182}), 16: (1, {'@': 182}), 3: (1, {'@': 182}), 4: (1, {'@': 182}), 6: (1, {'@': 182}), 17: (1, {'@': 182}), 20: (1, {'@': 182}), 13: (1, {'@': 182}), 22: (1, {'@': 182}), 10: (1, {'@': 182}), 12: (1, {'@': 182}), 25: (1, {'@': 182}), 26: (1, {'@': 182})}, 421: {33: (0, 427), 34: (0, 123), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416), 31: (1, {'@': 203}), 68: (1, {'@': 203})}, 422: {27: (1, {'@': 233}), 31: (1, {'@': 233}), 71: (1, {'@': 233}), 28: (1, {'@': 233}), 47: (1, {'@': 233}), 83: (1, {'@': 233})}, 423: {29: (1, {'@': 230}), 30: (1, {'@': 230}), 47: (1, {'@': 230}), 27: (1, {'@': 230}), 28: (1, {'@': 230}), 31: (1, {'@': 230}), 48: (1, {'@': 230}), 49: (1, {'@': 230}), 50: (1, {'@': 230}), 51: (1, {'@': 230}), 52: (1, {'@': 230}), 53: (1, {'@': 230}), 54: (1, {'@': 230}), 55: (1, {'@': 230}), 56: (1, {'@': 230}), 57: (1, {'@': 230}), 58: (1, {'@': 230}), 59: (1, {'@': 230}), 60: (1, {'@': 230}), 61: (1, {'@': 230}), 62: (1, {'@': 230}), 63: (1, {'@': 230}), 64: (1, {'@': 230}), 65: (1, {'@': 230}), 66: (1, {'@': 230}), 67: (1, {'@': 230}), 68: (1, {'@': 230}), 69: (1, {'@': 230}), 18: (1, {'@': 230}), 70: (1, {'@': 230})}, 424: {27: (1, {'@': 237}), 31: (1, {'@': 237}), 71: (1, {'@': 237}), 28: (1, {'@': 237}), 47: (1, {'@': 237}), 83: (1, {'@': 237})}, 425: {92: (0, 248), 93: (0, 464), 94: (0, 505), 95: (0, 252), 96: (0, 228), 10: (0, 63), 97: (0, 185), 98: (0, 422), 100: (0, 159), 99: (0, 210), 101: (0, 256), 102: (0, 260), 103: (0, 457), 104: (0, 207), 105: (0, 237), 106: (0, 484), 107: (0, 208), 109: (0, 243), 108: (0, 408), 110: (0, 227)}, 426: {92: (0, 248), 93: (0, 464), 94: (0, 505), 100: (0, 300), 96: (0, 228), 95: (0, 252), 10: (0, 63), 97: (0, 185), 98: (0, 422), 99: (0, 210), 101: (0, 256), 102: (0, 260), 103: (0, 457), 104: (0, 207), 105: (0, 237), 106: (0, 484), 107: (0, 208), 109: (0, 243), 108: (0, 408), 110: (0, 227)}, 427: {29: (1, {'@': 167}), 30: (1, {'@': 167}), 47: (1, {'@': 167}), 27: (1, {'@': 167}), 28: (1, {'@': 167}), 31: (1, {'@': 167}), 48: (1, {'@': 167}), 51: (1, {'@': 167}), 56: (1, {'@': 167}), 57: (1, {'@': 167}), 59: (1, {'@': 167}), 60: (1, {'@': 167}), 64: (1, {'@': 167}), 65: (1, {'@': 167}), 49: (1, {'@': 167}), 50: (1, {'@': 167}), 52: (1, {'@': 167}), 53: (1, {'@': 167}), 54: (1, {'@': 167}), 55: (1, {'@': 167}), 58: (1, {'@': 167}), 61: (1, {'@': 167}), 62: (1, {'@': 167}), 63: (1, {'@': 167}), 66: (1, {'@': 167}), 67: (1, {'@': 167}), 68: (1, {'@': 167}), 69: (1, {'@': 167}), 18: (1, {'@': 167}), 70: (1, {'@': 167})}, 428: {33: (0, 427), 34: (0, 134), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 69: (0, 376), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 137: (0, 119), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 429: {47: (0, 169)}, 430: {29: (0, 406), 30: (0, 399), 27: (1, {'@': 264}), 31: (1, {'@': 264})}, 431: {0: (1, {'@': 184}), 15: (1, {'@': 184}), 16: (1, {'@': 184}), 3: (1, {'@': 184}), 4: (1, {'@': 184}), 6: (1, {'@': 184}), 17: (1, {'@': 184}), 20: (1, {'@': 184}), 13: (1, {'@': 184}), 22: (1, {'@': 184}), 10: (1, {'@': 184}), 12: (1, {'@': 184}), 25: (1, {'@': 184}), 26: (1, {'@': 184})}, 432: {29: (1, {'@': 166}), 30: (1, {'@': 166}), 47: (1, {'@': 166}), 27: (1, {'@': 166}), 28: (1, {'@': 166}), 31: (1, {'@': 166}), 48: (1, {'@': 166}), 51: (1, {'@': 166}), 56: (1, {'@': 166}), 57: (1, {'@': 166}), 59: (1, {'@': 166}), 60: (1, {'@': 166}), 64: (1, {'@': 166}), 65: (1, {'@': 166}), 49: (1, {'@': 166}), 50: (1, {'@': 166}), 52: (1, {'@': 166}), 53: (1, {'@': 166}), 54: (1, {'@': 166}), 55: (1, {'@': 166}), 58: (1, {'@': 166}), 61: (1, {'@': 166}), 62: (1, {'@': 166}), 63: (1, {'@': 166}), 66: (1, {'@': 166}), 67: (1, {'@': 166}), 68: (1, {'@': 166}), 69: (1, {'@': 166}), 18: (1, {'@': 166}), 70: (1, {'@': 166})}, 433: {28: (0, 487)}, 434: {33: (0, 427), 34: (0, 174), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 435: {88: (0, 95), 27: (0, 78), 89: (0, 73), 160: (0, 71), 22: (0, 99)}, 436: {29: (1, {'@': 165}), 30: (1, {'@': 165}), 47: (1, {'@': 165}), 27: (1, {'@': 165}), 28: (1, {'@': 165}), 31: (1, {'@': 165}), 48: (1, {'@': 165}), 51: (1, {'@': 165}), 56: (1, {'@': 165}), 57: (1, {'@': 165}), 59: (1, {'@': 165}), 60: (1, {'@': 165}), 64: (1, {'@': 165}), 65: (1, {'@': 165}), 49: (1, {'@': 165}), 50: (1, {'@': 165}), 52: (1, {'@': 165}), 53: (1, {'@': 165}), 54: (1, {'@': 165}), 55: (1, {'@': 165}), 58: (1, {'@': 165}), 61: (1, {'@': 165}), 62: (1, {'@': 165}), 63: (1, {'@': 165}), 66: (1, {'@': 165}), 67: (1, {'@': 165}), 68: (1, {'@': 165}), 69: (1, {'@': 165}), 18: (1, {'@': 165}), 70: (1, {'@': 165})}, 437: {10: (0, 111)}, 438: {10: (0, 489)}, 439: {31: (0, 151), 68: (0, 145)}, 440: {0: (1, {'@': 139}), 1: (1, {'@': 139}), 2: (1, {'@': 139}), 3: (1, {'@': 139}), 4: (1, {'@': 139}), 5: (1, {'@': 139}), 6: (1, {'@': 139}), 7: (1, {'@': 139}), 8: (1, {'@': 139}), 9: (1, {'@': 139}), 10: (1, {'@': 139}), 11: (1, {'@': 139}), 12: (1, {'@': 139}), 13: (1, {'@': 139}), 14: (1, {'@': 139}), 15: (1, {'@': 139}), 16: (1, {'@': 139}), 17: (1, {'@': 139}), 18: (1, {'@': 139}), 19: (1, {'@': 139}), 20: (1, {'@': 139}), 21: (1, {'@': 139}), 22: (1, {'@': 139}), 23: (1, {'@': 139}), 24: (1, {'@': 139}), 25: (1, {'@': 139}), 26: (1, {'@': 139}), 27: (1, {'@': 139})}, 441: {10: (0, 392), 90: (0, 31)}, 442: {29: (1, {'@': 222}), 30: (1, {'@': 222}), 47: (1, {'@': 222}), 27: (1, {'@': 222}), 28: (1, {'@': 222}), 31: (1, {'@': 222}), 48: (1, {'@': 222}), 49: (1, {'@': 222}), 50: (1, {'@': 222}), 51: (1, {'@': 222}), 52: (1, {'@': 222}), 53: (1, {'@': 222}), 54: (1, {'@': 222}), 55: (1, {'@': 222}), 56: (1, {'@': 222}), 57: (1, {'@': 222}), 58: (1, {'@': 222}), 59: (1, {'@': 222}), 60: (1, {'@': 222}), 61: (1, {'@': 222}), 62: (1, {'@': 222}), 63: (1, {'@': 222}), 64: (1, {'@': 222}), 65: (1, {'@': 222}), 66: (1, {'@': 222}), 67: (1, {'@': 222}), 68: (1, {'@': 222}), 69: (1, {'@': 222}), 18: (1, {'@': 222}), 70: (1, {'@': 222})}, 443: {29: (0, 406), 30: (0, 399), 31: (1, {'@': 204}), 68: (1, {'@': 204})}, 444: {93: (0, 464), 138: (0, 510), 33: (0, 427), 98: (0, 422), 26: (0, 47), 17: (0, 494), 20: (0, 438), 94: (0, 505), 37: (0, 488), 10: (0, 460), 103: (0, 457), 34: (0, 430), 38: (0, 402), 27: (0, 116), 106: (0, 484), 108: (0, 408), 39: (0, 468), 110: (0, 227), 40: (0, 201), 92: (0, 248), 41: (0, 180), 42: (0, 273), 16: (0, 193), 97: (0, 185), 139: (0, 293), 101: (0, 256), 43: (0, 223), 12: (0, 203), 4: (0, 188), 96: (0, 228), 140: (0, 213), 35: (0, 178), 99: (0, 210), 36: (0, 232), 141: (0, 131), 104: (0, 207), 105: (0, 237), 0: (0, 268), 6: (0, 249), 95: (0, 252), 15: (0, 257), 102: (0, 260), 22: (0, 262), 3: (0, 265), 13: (0, 267), 142: (0, 271), 107: (0, 208), 44: (0, 274), 109: (0, 243), 25: (0, 442), 45: (0, 504), 100: (0, 516), 46: (0, 416)}, 445: {33: (0, 427), 34: (0, 155), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 446: {92: (0, 248), 93: (0, 464), 94: (0, 505), 95: (0, 252), 96: (0, 228), 10: (0, 63), 97: (0, 185), 98: (0, 422), 99: (0, 210), 101: (0, 256), 102: (0, 260), 103: (0, 457), 100: (0, 473), 104: (0, 207), 105: (0, 237), 106: (0, 484), 107: (0, 208), 109: (0, 243), 108: (0, 408), 110: (0, 227)}, 447: {19: (0, 437), 79: (0, 340), 33: (0, 427), 38: (0, 397), 23: (0, 23), 5: (0, 292), 26: (0, 47), 17: (0, 494), 7: (0, 44), 20: (0, 438), 115: (0, 28), 11: (0, 40), 116: (0, 37), 21: (0, 30), 128: (0, 15), 34: (0, 2), 117: (0, 0), 24: (0, 4), 37: (0, 488), 10: (0, 51), 46: (0, 11), 118: (0, 20), 119: (0, 33), 39: (0, 468), 8: (0, 24), 2: (0, 142), 40: (0, 201), 120: (0, 152), 121: (0, 150), 122: (0, 161), 41: (0, 180), 42: (0, 273), 16: (0, 193), 123: (0, 158), 43: (0, 164), 124: (0, 141), 125: (0, 135), 12: (0, 129), 4: (0, 124), 9: (0, 112), 126: (0, 154), 127: (0, 147), 35: (0, 178), 36: (0, 232), 132: (0, 220), 0: (0, 268), 1: (0, 288), 6: (0, 249), 129: (0, 294), 130: (0, 303), 15: (0, 257), 131: (0, 224), 22: (0, 262), 14: (0, 75), 3: (0, 265), 13: (0, 267), 80: (0, 244), 44: (0, 274), 25: (0, 442), 45: (0, 504), 133: (0, 218), 18: (1, {'@': 98})}, 448: {34: (0, 126), 33: (0, 427), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 145: (0, 143), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 88: (0, 362), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 370), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 449: {31: (0, 162), 27: (0, 513)}, 450: {29: (1, {'@': 201}), 30: (1, {'@': 201}), 47: (1, {'@': 201}), 27: (1, {'@': 201}), 28: (1, {'@': 201}), 31: (1, {'@': 201}), 48: (1, {'@': 201}), 49: (1, {'@': 201}), 50: (1, {'@': 201}), 51: (1, {'@': 201}), 52: (1, {'@': 201}), 53: (1, {'@': 201}), 54: (1, {'@': 201}), 55: (1, {'@': 201}), 56: (1, {'@': 201}), 57: (1, {'@': 201}), 58: (1, {'@': 201}), 59: (1, {'@': 201}), 60: (1, {'@': 201}), 61: (1, {'@': 201}), 62: (1, {'@': 201}), 63: (1, {'@': 201}), 64: (1, {'@': 201}), 65: (1, {'@': 201}), 66: (1, {'@': 201}), 71: (1, {'@': 201}), 67: (1, {'@': 201}), 68: (1, {'@': 201}), 69: (1, {'@': 201}), 18: (1, {'@': 201}), 70: (1, {'@': 201})}, 451: {30: (0, 399), 29: (0, 406), 28: (0, 27)}, 452: {113: (1, {'@': 87})}, 453: {0: (1, {'@': 177}), 15: (1, {'@': 177}), 16: (1, {'@': 177}), 3: (1, {'@': 177}), 4: (1, {'@': 177}), 6: (1, {'@': 177}), 17: (1, {'@': 177}), 20: (1, {'@': 177}), 13: (1, {'@': 177}), 22: (1, {'@': 177}), 10: (1, {'@': 177}), 12: (1, {'@': 177}), 25: (1, {'@': 177}), 26: (1, {'@': 177})}, 454: {92: (0, 248), 93: (0, 464), 94: (0, 505), 95: (0, 252), 96: (0, 228), 100: (0, 100), 10: (0, 63), 97: (0, 185), 98: (0, 422), 99: (0, 210), 101: (0, 256), 102: (0, 260), 103: (0, 457), 104: (0, 207), 105: (0, 237), 106: (0, 484), 107: (0, 208), 109: (0, 243), 108: (0, 408), 110: (0, 227)}, 455: {0: (1, {'@': 181}), 15: (1, {'@': 181}), 16: (1, {'@': 181}), 3: (1, {'@': 181}), 4: (1, {'@': 181}), 6: (1, {'@': 181}), 17: (1, {'@': 181}), 20: (1, {'@': 181}), 13: (1, {'@': 181}), 22: (1, {'@': 181}), 10: (1, {'@': 181}), 12: (1, {'@': 181}), 25: (1, {'@': 181}), 26: (1, {'@': 181})}, 456: {29: (1, {'@': 206}), 30: (1, {'@': 206}), 47: (1, {'@': 206}), 27: (1, {'@': 206}), 28: (1, {'@': 206}), 31: (1, {'@': 206}), 48: (1, {'@': 206}), 49: (1, {'@': 206}), 50: (1, {'@': 206}), 51: (1, {'@': 206}), 52: (1, {'@': 206}), 53: (1, {'@': 206}), 54: (1, {'@': 206}), 55: (1, {'@': 206}), 56: (1, {'@': 206}), 57: (1, {'@': 206}), 58: (1, {'@': 206}), 59: (1, {'@': 206}), 60: (1, {'@': 206}), 61: (1, {'@': 206}), 62: (1, {'@': 206}), 63: (1, {'@': 206}), 64: (1, {'@': 206}), 65: (1, {'@': 206}), 66: (1, {'@': 206}), 67: (1, {'@': 206}), 68: (1, {'@': 206}), 69: (1, {'@': 206}), 18: (1, {'@': 206}), 150: (1, {'@': 206}), 83: (1, {'@': 206}), 70: (1, {'@': 206})}, 457: {27: (1, {'@': 258}), 31: (1, {'@': 258}), 58: (1, {'@': 258}), 71: (1, {'@': 258}), 28: (1, {'@': 258}), 47: (1, {'@': 258}), 83: (1, {'@': 258})}, 458: {29: (1, {'@': 273}), 30: (1, {'@': 273}), 47: (1, {'@': 273}), 27: (1, {'@': 273}), 28: (1, {'@': 273}), 31: (1, {'@': 273}), 48: (1, {'@': 273}), 49: (1, {'@': 273}), 50: (1, {'@': 273}), 51: (1, {'@': 273}), 52: (1, {'@': 273}), 53: (1, {'@': 273}), 54: (1, {'@': 273}), 55: (1, {'@': 273}), 56: (1, {'@': 273}), 57: (1, {'@': 273}), 58: (1, {'@': 273}), 59: (1, {'@': 273}), 60: (1, {'@': 273}), 61: (1, {'@': 273}), 62: (1, {'@': 273}), 63: (1, {'@': 273}), 64: (1, {'@': 273}), 65: (1, {'@': 273}), 66: (1, {'@': 273}), 67: (1, {'@': 273}), 68: (1, {'@': 273}), 69: (1, {'@': 273}), 18: (1, {'@': 273}), 70: (1, {'@': 273})}, 459: {0: (1, {'@': 173}), 15: (1, {'@': 173}), 16: (1, {'@': 173}), 3: (1, {'@': 173}), 4: (1, {'@': 173}), 6: (1, {'@': 173}), 17: (1, {'@': 173}), 20: (1, {'@': 173}), 13: (1, {'@': 173}), 22: (1, {'@': 173}), 10: (1, {'@': 173}), 12: (1, {'@': 173}), 25: (1, {'@': 173}), 26: (1, {'@': 173})}, 460: {93: (0, 464), 33: (0, 427), 98: (0, 422), 34: (0, 412), 26: (0, 47), 17: (0, 494), 20: (0, 438), 100: (0, 511), 94: (0, 505), 37: (0, 488), 10: (0, 460), 103: (0, 457), 38: (0, 402), 106: (0, 484), 108: (0, 408), 39: (0, 468), 110: (0, 227), 40: (0, 201), 92: (0, 248), 41: (0, 180), 42: (0, 273), 16: (0, 193), 97: (0, 185), 101: (0, 256), 43: (0, 223), 12: (0, 203), 4: (0, 188), 96: (0, 228), 86: (0, 57), 35: (0, 178), 99: (0, 210), 36: (0, 232), 0: (0, 268), 104: (0, 207), 105: (0, 237), 6: (0, 249), 53: (0, 56), 95: (0, 252), 15: (0, 257), 64: (0, 68), 102: (0, 260), 22: (0, 262), 3: (0, 265), 13: (0, 267), 87: (0, 97), 107: (0, 208), 44: (0, 274), 27: (0, 62), 109: (0, 243), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 461: {33: (0, 427), 34: (0, 146), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 158: (0, 415), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 462: {29: (1, {'@': 199}), 30: (1, {'@': 199}), 47: (1, {'@': 199}), 27: (1, {'@': 199}), 28: (1, {'@': 199}), 31: (1, {'@': 199}), 48: (1, {'@': 199}), 49: (1, {'@': 199}), 50: (1, {'@': 199}), 51: (1, {'@': 199}), 52: (1, {'@': 199}), 53: (1, {'@': 199}), 54: (1, {'@': 199}), 55: (1, {'@': 199}), 56: (1, {'@': 199}), 57: (1, {'@': 199}), 58: (1, {'@': 199}), 59: (1, {'@': 199}), 60: (1, {'@': 199}), 61: (1, {'@': 199}), 62: (1, {'@': 199}), 63: (1, {'@': 199}), 64: (1, {'@': 199}), 65: (1, {'@': 199}), 66: (1, {'@': 199}), 71: (1, {'@': 199}), 67: (1, {'@': 199}), 68: (1, {'@': 199}), 69: (1, {'@': 199}), 18: (1, {'@': 199}), 70: (1, {'@': 199})}, 463: {75: (1, {'@': 93}), 5: (1, {'@': 93}), 18: (1, {'@': 93}), 14: (1, {'@': 93})}, 464: {27: (1, {'@': 250}), 31: (1, {'@': 250}), 58: (1, {'@': 250}), 71: (1, {'@': 250}), 28: (1, {'@': 250}), 47: (1, {'@': 250}), 83: (1, {'@': 250})}, 465: {71: (0, 106), 28: (0, 58)}, 466: {27: (1, {'@': 306}), 31: (1, {'@': 306})}, 467: {31: (0, 441)}, 468: {29: (1, {'@': 197}), 30: (1, {'@': 197}), 47: (1, {'@': 197}), 27: (1, {'@': 197}), 28: (1, {'@': 197}), 31: (1, {'@': 197}), 48: (1, {'@': 197}), 49: (1, {'@': 197}), 50: (1, {'@': 197}), 51: (1, {'@': 197}), 52: (1, {'@': 197}), 53: (1, {'@': 197}), 54: (1, {'@': 197}), 55: (1, {'@': 197}), 56: (1, {'@': 197}), 57: (1, {'@': 197}), 58: (1, {'@': 197}), 59: (1, {'@': 197}), 60: (1, {'@': 197}), 61: (1, {'@': 197}), 62: (1, {'@': 197}), 63: (1, {'@': 197}), 64: (1, {'@': 197}), 65: (1, {'@': 197}), 66: (1, {'@': 197}), 67: (1, {'@': 197}), 68: (1, {'@': 197}), 69: (1, {'@': 197}), 18: (1, {'@': 197}), 70: (1, {'@': 197})}, 469: {33: (0, 427), 34: (0, 114), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 470: {88: (0, 95), 27: (0, 38), 89: (0, 29), 22: (0, 99)}, 471: {0: (1, {'@': 178}), 15: (1, {'@': 178}), 16: (1, {'@': 178}), 3: (1, {'@': 178}), 4: (1, {'@': 178}), 6: (1, {'@': 178}), 17: (1, {'@': 178}), 20: (1, {'@': 178}), 13: (1, {'@': 178}), 22: (1, {'@': 178}), 10: (1, {'@': 178}), 12: (1, {'@': 178}), 25: (1, {'@': 178}), 26: (1, {'@': 178})}, 472: {27: (1, {'@': 243}), 31: (1, {'@': 243}), 71: (1, {'@': 243}), 28: (1, {'@': 243}), 47: (1, {'@': 243}), 83: (1, {'@': 243})}, 473: {27: (1, {'@': 301}), 31: (1, {'@': 301})}, 474: {27: (1, {'@': 240}), 31: (1, {'@': 240}), 71: (1, {'@': 240}), 28: (1, {'@': 240}), 47: (1, {'@': 240}), 83: (1, {'@': 240})}, 475: {83: (0, 354), 82: (0, 105), 75: (1, {'@': 121}), 5: (1, {'@': 121}), 18: (1, {'@': 121}), 14: (1, {'@': 121}), 0: (1, {'@': 121}), 1: (1, {'@': 121}), 2: (1, {'@': 121}), 3: (1, {'@': 121}), 4: (1, {'@': 121}), 6: (1, {'@': 121}), 7: (1, {'@': 121}), 8: (1, {'@': 121}), 9: (1, {'@': 121}), 10: (1, {'@': 121}), 11: (1, {'@': 121}), 12: (1, {'@': 121}), 13: (1, {'@': 121}), 15: (1, {'@': 121}), 16: (1, {'@': 121}), 17: (1, {'@': 121}), 19: (1, {'@': 121}), 20: (1, {'@': 121}), 21: (1, {'@': 121}), 22: (1, {'@': 121}), 23: (1, {'@': 121}), 24: (1, {'@': 121}), 25: (1, {'@': 121}), 26: (1, {'@': 121}), 27: (1, {'@': 121})}, 476: {29: (1, {'@': 200}), 30: (1, {'@': 200}), 47: (1, {'@': 200}), 27: (1, {'@': 200}), 28: (1, {'@': 200}), 31: (1, {'@': 200}), 48: (1, {'@': 200}), 49: (1, {'@': 200}), 50: (1, {'@': 200}), 51: (1, {'@': 200}), 52: (1, {'@': 200}), 53: (1, {'@': 200}), 54: (1, {'@': 200}), 55: (1, {'@': 200}), 56: (1, {'@': 200}), 57: (1, {'@': 200}), 58: (1, {'@': 200}), 59: (1, {'@': 200}), 60: (1, {'@': 200}), 61: (1, {'@': 200}), 62: (1, {'@': 200}), 63: (1, {'@': 200}), 64: (1, {'@': 200}), 65: (1, {'@': 200}), 66: (1, {'@': 200}), 71: (1, {'@': 200}), 67: (1, {'@': 200}), 68: (1, {'@': 200}), 69: (1, {'@': 200}), 18: (1, {'@': 200}), 70: (1, {'@': 200})}, 477: {19: (0, 437), 79: (0, 340), 33: (0, 427), 38: (0, 397), 23: (0, 23), 5: (0, 292), 26: (0, 47), 17: (0, 494), 7: (0, 44), 20: (0, 438), 115: (0, 28), 11: (0, 40), 116: (0, 37), 21: (0, 30), 37: (0, 488), 34: (0, 2), 117: (0, 0), 24: (0, 4), 10: (0, 51), 46: (0, 11), 118: (0, 20), 119: (0, 33), 39: (0, 468), 8: (0, 24), 2: (0, 142), 40: (0, 201), 120: (0, 152), 121: (0, 150), 122: (0, 161), 41: (0, 180), 42: (0, 273), 16: (0, 193), 123: (0, 158), 128: (0, 285), 43: (0, 164), 124: (0, 141), 125: (0, 135), 12: (0, 129), 4: (0, 124), 25: (0, 442), 9: (0, 112), 126: (0, 154), 127: (0, 147), 35: (0, 178), 36: (0, 232), 0: (0, 268), 1: (0, 288), 6: (0, 249), 129: (0, 294), 130: (0, 303), 15: (0, 257), 131: (0, 224), 22: (0, 262), 14: (0, 75), 3: (0, 265), 13: (0, 267), 80: (0, 244), 44: (0, 274), 132: (0, 220), 45: (0, 504), 133: (0, 218), 18: (1, {'@': 98})}, 478: {29: (0, 406), 30: (0, 399), 32: (0, 301), 31: (0, 336), 27: (0, 327)}, 479: {0: (1, {'@': 180}), 15: (1, {'@': 180}), 16: (1, {'@': 180}), 3: (1, {'@': 180}), 4: (1, {'@': 180}), 6: (1, {'@': 180}), 17: (1, {'@': 180}), 20: (1, {'@': 180}), 13: (1, {'@': 180}), 22: (1, {'@': 180}), 10: (1, {'@': 180}), 12: (1, {'@': 180}), 25: (1, {'@': 180}), 26: (1, {'@': 180})}, 480: {92: (0, 248), 93: (0, 464), 94: (0, 505), 95: (0, 252), 96: (0, 228), 10: (0, 63), 97: (0, 185), 98: (0, 422), 99: (0, 210), 101: (0, 256), 102: (0, 260), 103: (0, 457), 104: (0, 207), 105: (0, 237), 106: (0, 484), 107: (0, 208), 109: (0, 243), 108: (0, 408), 110: (0, 227), 100: (0, 49)}, 481: {0: (1, {'@': 188}), 15: (1, {'@': 188}), 16: (1, {'@': 188}), 3: (1, {'@': 188}), 4: (1, {'@': 188}), 6: (1, {'@': 188}), 17: (1, {'@': 188}), 20: (1, {'@': 188}), 13: (1, {'@': 188}), 22: (1, {'@': 188}), 10: (1, {'@': 188}), 12: (1, {'@': 188}), 25: (1, {'@': 188}), 26: (1, {'@': 188})}, 482: {92: (0, 248), 93: (0, 464), 94: (0, 505), 95: (0, 252), 96: (0, 228), 10: (0, 63), 97: (0, 185), 98: (0, 422), 99: (0, 210), 100: (0, 424), 101: (0, 256), 102: (0, 260), 103: (0, 457), 104: (0, 207), 105: (0, 237), 106: (0, 484), 107: (0, 208), 109: (0, 243), 108: (0, 408), 110: (0, 227)}, 483: {27: (1, {'@': 271}), 31: (1, {'@': 271})}, 484: {27: (1, {'@': 247}), 31: (1, {'@': 247}), 58: (1, {'@': 247}), 71: (1, {'@': 247}), 28: (1, {'@': 247}), 47: (1, {'@': 247}), 83: (1, {'@': 247})}, 485: {0: (1, {'@': 176}), 15: (1, {'@': 176}), 16: (1, {'@': 176}), 3: (1, {'@': 176}), 4: (1, {'@': 176}), 6: (1, {'@': 176}), 17: (1, {'@': 176}), 20: (1, {'@': 176}), 13: (1, {'@': 176}), 22: (1, {'@': 176}), 10: (1, {'@': 176}), 12: (1, {'@': 176}), 25: (1, {'@': 176}), 26: (1, {'@': 176})}, 486: {65: (0, 137)}, 487: {0: (1, {'@': 143}), 1: (1, {'@': 143}), 2: (1, {'@': 143}), 3: (1, {'@': 143}), 4: (1, {'@': 143}), 5: (1, {'@': 143}), 6: (1, {'@': 143}), 7: (1, {'@': 143}), 8: (1, {'@': 143}), 9: (1, {'@': 143}), 10: (1, {'@': 143}), 11: (1, {'@': 143}), 12: (1, {'@': 143}), 13: (1, {'@': 143}), 14: (1, {'@': 143}), 15: (1, {'@': 143}), 16: (1, {'@': 143}), 17: (1, {'@': 143}), 18: (1, {'@': 143}), 19: (1, {'@': 143}), 20: (1, {'@': 143}), 21: (1, {'@': 143}), 22: (1, {'@': 143}), 23: (1, {'@': 143}), 24: (1, {'@': 143}), 25: (1, {'@': 143}), 26: (1, {'@': 143}), 27: (1, {'@': 143})}, 488: {29: (1, {'@': 193}), 30: (1, {'@': 193}), 47: (1, {'@': 193}), 27: (1, {'@': 193}), 28: (1, {'@': 193}), 31: (1, {'@': 193}), 48: (1, {'@': 193}), 49: (1, {'@': 193}), 50: (1, {'@': 193}), 51: (1, {'@': 193}), 52: (1, {'@': 193}), 53: (1, {'@': 193}), 54: (1, {'@': 193}), 55: (1, {'@': 193}), 56: (1, {'@': 193}), 57: (1, {'@': 193}), 58: (1, {'@': 193}), 59: (1, {'@': 193}), 60: (1, {'@': 193}), 61: (1, {'@': 193}), 62: (1, {'@': 193}), 63: (1, {'@': 193}), 64: (1, {'@': 193}), 65: (1, {'@': 193}), 66: (1, {'@': 193}), 67: (1, {'@': 193}), 68: (1, {'@': 193}), 69: (1, {'@': 193}), 18: (1, {'@': 193}), 70: (1, {'@': 193})}, 489: {33: (0, 427), 34: (0, 74), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 490: {31: (0, 419), 27: (1, {'@': 210})}, 491: {27: (0, 380)}, 492: {18: (0, 463)}, 493: {29: (0, 406), 30: (0, 399), 27: (0, 433), 31: (0, 325)}, 494: {29: (1, {'@': 224}), 30: (1, {'@': 224}), 47: (1, {'@': 224}), 27: (1, {'@': 224}), 28: (1, {'@': 224}), 31: (1, {'@': 224}), 48: (1, {'@': 224}), 49: (1, {'@': 224}), 50: (1, {'@': 224}), 51: (1, {'@': 224}), 52: (1, {'@': 224}), 53: (1, {'@': 224}), 54: (1, {'@': 224}), 55: (1, {'@': 224}), 56: (1, {'@': 224}), 57: (1, {'@': 224}), 58: (1, {'@': 224}), 59: (1, {'@': 224}), 60: (1, {'@': 224}), 61: (1, {'@': 224}), 62: (1, {'@': 224}), 63: (1, {'@': 224}), 64: (1, {'@': 224}), 65: (1, {'@': 224}), 66: (1, {'@': 224}), 67: (1, {'@': 224}), 68: (1, {'@': 224}), 69: (1, {'@': 224}), 18: (1, {'@': 224}), 70: (1, {'@': 224})}, 495: {0: (1, {'@': 183}), 15: (1, {'@': 183}), 16: (1, {'@': 183}), 3: (1, {'@': 183}), 4: (1, {'@': 183}), 6: (1, {'@': 183}), 17: (1, {'@': 183}), 20: (1, {'@': 183}), 13: (1, {'@': 183}), 22: (1, {'@': 183}), 10: (1, {'@': 183}), 12: (1, {'@': 183}), 25: (1, {'@': 183}), 26: (1, {'@': 183})}, 496: {72: (0, 261)}, 497: {0: (1, {'@': 174}), 15: (1, {'@': 174}), 16: (1, {'@': 174}), 3: (1, {'@': 174}), 4: (1, {'@': 174}), 6: (1, {'@': 174}), 17: (1, {'@': 174}), 20: (1, {'@': 174}), 13: (1, {'@': 174}), 22: (1, {'@': 174}), 10: (1, {'@': 174}), 12: (1, {'@': 174}), 25: (1, {'@': 174}), 26: (1, {'@': 174})}, 498: {29: (0, 406), 30: (0, 399), 27: (0, 432)}, 499: {27: (1, {'@': 267}), 31: (1, {'@': 267}), 28: (1, {'@': 267})}, 500: {28: (0, 166), 31: (0, 501), 143: (0, 486), 65: (0, 474)}, 501: {33: (0, 427), 34: (0, 146), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 158: (0, 173), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 502: {29: (0, 406), 30: (0, 399), 65: (1, {'@': 246})}, 503: {47: (0, 496)}, 504: {29: (1, {'@': 225}), 30: (1, {'@': 225}), 47: (1, {'@': 225}), 27: (1, {'@': 225}), 28: (1, {'@': 225}), 31: (1, {'@': 225}), 48: (1, {'@': 225}), 49: (1, {'@': 225}), 50: (1, {'@': 225}), 51: (1, {'@': 225}), 52: (1, {'@': 225}), 53: (1, {'@': 225}), 54: (1, {'@': 225}), 55: (1, {'@': 225}), 56: (1, {'@': 225}), 57: (1, {'@': 225}), 58: (1, {'@': 225}), 59: (1, {'@': 225}), 60: (1, {'@': 225}), 61: (1, {'@': 225}), 62: (1, {'@': 225}), 63: (1, {'@': 225}), 64: (1, {'@': 225}), 65: (1, {'@': 225}), 66: (1, {'@': 225}), 67: (1, {'@': 225}), 68: (1, {'@': 225}), 69: (1, {'@': 225}), 18: (1, {'@': 225}), 70: (1, {'@': 225})}, 505: {58: (0, 409), 27: (1, {'@': 231}), 31: (1, {'@': 231}), 71: (1, {'@': 231}), 28: (1, {'@': 231}), 47: (1, {'@': 231}), 83: (1, {'@': 231})}, 506: {19: (0, 437), 79: (0, 340), 33: (0, 427), 38: (0, 397), 23: (0, 23), 5: (0, 292), 26: (0, 47), 17: (0, 494), 7: (0, 44), 20: (0, 438), 115: (0, 28), 11: (0, 40), 116: (0, 37), 21: (0, 30), 37: (0, 488), 34: (0, 2), 117: (0, 0), 24: (0, 4), 10: (0, 51), 46: (0, 11), 118: (0, 20), 119: (0, 33), 39: (0, 468), 8: (0, 24), 2: (0, 142), 40: (0, 201), 120: (0, 152), 121: (0, 150), 122: (0, 161), 41: (0, 180), 42: (0, 273), 16: (0, 193), 123: (0, 158), 43: (0, 164), 124: (0, 141), 125: (0, 135), 12: (0, 129), 4: (0, 124), 25: (0, 442), 9: (0, 112), 126: (0, 154), 127: (0, 147), 35: (0, 178), 36: (0, 232), 0: (0, 268), 128: (0, 9), 1: (0, 288), 6: (0, 249), 129: (0, 294), 130: (0, 303), 15: (0, 257), 131: (0, 224), 22: (0, 262), 14: (0, 75), 3: (0, 265), 13: (0, 267), 80: (0, 244), 44: (0, 274), 132: (0, 220), 45: (0, 504), 133: (0, 218), 18: (1, {'@': 98})}, 507: {71: (0, 22)}, 508: {27: (1, {'@': 307}), 31: (1, {'@': 307})}, 509: {47: (0, 477)}, 510: {10: (0, 444)}, 511: {161: (0, 357), 31: (0, 446), 27: (0, 360)}, 512: {27: (1, {'@': 242}), 31: (1, {'@': 242}), 71: (1, {'@': 242}), 28: (1, {'@': 242}), 47: (1, {'@': 242}), 83: (1, {'@': 242})}, 513: {27: (1, {'@': 270}), 31: (1, {'@': 270})}, 514: {33: (0, 427), 34: (0, 132), 35: (0, 178), 26: (0, 47), 36: (0, 232), 17: (0, 494), 20: (0, 438), 37: (0, 488), 10: (0, 51), 38: (0, 402), 0: (0, 268), 39: (0, 468), 6: (0, 249), 40: (0, 201), 41: (0, 180), 42: (0, 273), 15: (0, 257), 16: (0, 193), 22: (0, 262), 43: (0, 223), 3: (0, 265), 13: (0, 267), 12: (0, 203), 4: (0, 188), 44: (0, 274), 25: (0, 442), 45: (0, 504), 46: (0, 416)}, 515: {18: (0, 394)}, 516: {27: (1, {'@': 265}), 31: (1, {'@': 265})}}, 'start_states': {'start': 383}, 'end_states': {'start': 351}}, '__type__': 'ParsingFrontend'}, 'rules': [{'@': 86}, {'@': 87}, {'@': 88}, {'@': 89}, {'@': 90}, {'@': 91}, {'@': 92}, {'@': 93}, {'@': 94}, {'@': 95}, {'@': 96}, {'@': 97}, {'@': 98}, {'@': 99}, {'@': 100}, {'@': 101}, {'@': 102}, {'@': 103}, {'@': 104}, {'@': 105}, {'@': 106}, {'@': 107}, {'@': 108}, {'@': 109}, {'@': 110}, {'@': 111}, {'@': 112}, {'@': 113}, {'@': 114}, {'@': 115}, {'@': 116}, {'@': 117}, {'@': 118}, {'@': 119}, {'@': 120}, {'@': 121}, {'@': 122}, {'@': 123}, {'@': 124}, {'@': 125}, {'@': 126}, {'@': 127}, {'@': 128}, {'@': 129}, {'@': 130}, {'@': 131}, {'@': 132}, {'@': 133}, {'@': 134}, {'@': 135}, {'@': 136}, {'@': 137}, {'@': 138}, {'@': 139}, {'@': 140}, {'@': 141}, {'@': 142}, {'@': 143}, {'@': 144}, {'@': 145}, {'@': 146}, {'@': 147}, {'@': 148}, {'@': 149}, {'@': 150}, {'@': 151}, {'@': 152}, {'@': 153}, {'@': 154}, {'@': 155}, {'@': 156}, {'@': 157}, {'@': 158}, {'@': 159}, {'@': 160}, {'@': 161}, {'@': 162}, {'@': 163}, {'@': 164}, {'@': 165}, {'@': 166}, {'@': 167}, {'@': 168}, {'@': 169}, {'@': 170}, {'@': 171}, {'@': 172}, {'@': 173}, {'@': 174}, {'@': 175}, {'@': 176}, {'@': 177}, {'@': 178}, {'@': 179}, {'@': 180}, {'@': 181}, {'@': 182}, {'@': 183}, {'@': 184}, {'@': 185}, {'@': 186}, {'@': 187}, {'@': 188}, {'@': 189}, {'@': 190}, {'@': 191}, {'@': 192}, {'@': 193}, {'@': 194}, {'@': 195}, {'@': 196}, {'@': 197}, {'@': 198}, {'@': 199}, {'@': 200}, {'@': 201}, {'@': 202}, {'@': 203}, {'@': 204}, {'@': 205}, {'@': 206}, {'@': 207}, {'@': 208}, {'@': 209}, {'@': 210}, {'@': 211}, {'@': 212}, {'@': 213}, {'@': 214}, {'@': 215}, {'@': 216}, {'@': 217}, {'@': 218}, {'@': 219}, {'@': 220}, {'@': 221}, {'@': 222}, {'@': 223}, {'@': 224}, {'@': 225}, {'@': 226}, {'@': 227}, {'@': 228}, {'@': 229}, {'@': 230}, {'@': 231}, {'@': 232}, {'@': 233}, {'@': 234}, {'@': 235}, {'@': 236}, {'@': 237}, {'@': 238}, {'@': 239}, {'@': 240}, {'@': 241}, {'@': 242}, {'@': 243}, {'@': 244}, {'@': 245}, {'@': 246}, {'@': 247}, {'@': 248}, {'@': 249}, {'@': 250}, {'@': 251}, {'@': 252}, {'@': 253}, {'@': 254}, {'@': 255}, {'@': 256}, {'@': 257}, {'@': 258}, {'@': 259}, {'@': 260}, {'@': 261}, {'@': 262}, {'@': 263}, {'@': 264}, {'@': 265}, {'@': 266}, {'@': 267}, {'@': 268}, {'@': 269}, {'@': 270}, {'@': 271}, {'@': 272}, {'@': 273}, {'@': 274}, {'@': 275}, {'@': 276}, {'@': 277}, {'@': 278}, {'@': 279}, {'@': 280}, {'@': 281}, {'@': 282}, {'@': 283}, {'@': 284}, {'@': 285}, {'@': 286}, {'@': 287}, {'@': 288}, {'@': 289}, {'@': 290}, {'@': 291}, {'@': 292}, {'@': 293}, {'@': 294}, {'@': 295}, {'@': 296}, {'@': 297}, {'@': 298}, {'@': 299}, {'@': 300}, {'@': 301}, {'@': 302}, {'@': 303}, {'@': 304}, {'@': 305}, {'@': 306}, {'@': 307}, {'@': 308}], 'options': {'debug': False, 'strict': False, 'keep_all_tokens': False, 'tree_class': None, 'cache': False, 'postlex': None, 'parser': 'lalr', 'lexer': 'contextual', 'transformer': None, 'start': ['start'], 'priority': 'normal', 'ambiguity': 'auto', 'regex': False, 'propagate_positions': False, 'lexer_callbacks': {}, 'maybe_placeholders': False, 'edit_terminals': None, 'g_regex_flags': 0, 'use_bytes': False, 'import_paths': [], 'source_path': None, '_plugins': {}}, '__type__': 'Lark'}
-)
-MEMO = (
-{0: {'name': 'INT', 'pattern': {'value': '(?:[0-9])+', 'flags': [], 'raw': None, '_width': [1, 4294967295], '__type__': 'PatternRE'}, 'priority': 0, '__type__': 'TerminalDef'}, 1: {'name': 'WS', 'pattern': {'value': '(?:[ \t\x0c\r\n])+', 'flags': [], 'raw': None, '_width': [1, 4294967295], '__type__': 'PatternRE'}, 'priority': 0, '__type__': 'TerminalDef'}, 2: {'name': 'IDENT', 'pattern': {'value': '[^\\W\\d]\\w*', 'flags': [], 'raw': '/[^\\W\\d]\\w*/', '_width': [1, 4294967295], '__type__': 'PatternRE'}, 'priority': 0, '__type__': 'TerminalDef'}, 3: {'name': 'ANYWILDCARD', 'pattern': {'value': '[^\\$]+', 'flags': [], 'raw': '/[^\\$]+/', '_width': [1, 4294967295], '__type__': 'PatternRE'}, 'priority': 0, '__type__': 'TerminalDef'}, 4: {'name': 'STRING', 'pattern': {'value': '([ubf]?r?|r[ubf])("(?!"").*?(?<!\\\\)(\\\\\\\\)*?"|\'(?!\'\').*?(?<!\\\\)(\\\\\\\\)*?\')', 'flags': ['i'], 'raw': '/([ubf]?r?|r[ubf])("(?!"").*?(?<!\\\\)(\\\\\\\\)*?"|\'(?!\'\').*?(?<!\\\\)(\\\\\\\\)*?\')/i', '_width': [2, 4294967295], '__type__': 'PatternRE'}, 'priority': 0, '__type__': 'TerminalDef'}, 5: {'name': 'SIGNED_FLOAT', 'pattern': {'value': '(?:(?:\\+|\\-))?(?:(?:[0-9])+(?:e|E)(?:(?:\\+|\\-))?(?:[0-9])+|(?:(?:[0-9])+\\.(?:(?:[0-9])+)?|\\.(?:[0-9])+)(?:(?:e|E)(?:(?:\\+|\\-))?(?:[0-9])+)?)', 'flags': [], 'raw': None, '_width': [2, 4294967295], '__type__': 'PatternRE'}, 'priority': 0, '__type__': 'TerminalDef'}, 6: {'name': 'MODULE', 'pattern': {'value': 'module', 'flags': [], 'raw': '"module"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 7: {'name': 'LBRACE', 'pattern': {'value': '{', 'flags': [], 'raw': '"{"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 8: {'name': 'RBRACE', 'pattern': {'value': '}', 'flags': [], 'raw': '"}"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 9: {'name': 'DEF', 'pattern': {'value': 'def', 'flags': [], 'raw': '"def"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 10: {'name': 'LPAR', 'pattern': {'value': '(', 'flags': [], 'raw': '"("', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 11: {'name': 'RPAR', 'pattern': {'value': ')', 'flags': [], 'raw': '")"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 12: {'name': '__ANON_0', 'pattern': {'value': '->', 'flags': [], 'raw': '"->"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 13: {'name': 'COMMA', 'pattern': {'value': ',', 'flags': [], 'raw': '","', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 14: {'name': 'EQUAL', 'pattern': {'value': '=', 'flags': [], 'raw': '"="', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 15: {'name': 'DECL', 'pattern': {'value': 'decl', 'flags': [], 'raw': '"decl"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 16: {'name': 'SEMICOLON', 'pattern': {'value': ';', 'flags': [], 'raw': '";"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 17: {'name': 'STATIC', 'pattern': {'value': 'static', 'flags': [], 'raw': '"static"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 18: {'name': 'PROTECTED', 'pattern': {'value': 'protected', 'flags': [], 'raw': '"protected"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 19: {'name': 'RETURN', 'pattern': {'value': 'return', 'flags': [], 'raw': '"return"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 20: {'name': 'IN', 'pattern': {'value': 'in', 'flags': [], 'raw': '"in"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 21: {'name': 'LET', 'pattern': {'value': 'let', 'flags': [], 'raw': '"let"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 22: {'name': 'FOR', 'pattern': {'value': 'for', 'flags': [], 'raw': '"for"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 23: {'name': 'ON', 'pattern': {'value': 'on', 'flags': [], 'raw': '"on"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 24: {'name': 'WHILE', 'pattern': {'value': 'while', 'flags': [], 'raw': '"while"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 25: {'name': 'BREAK', 'pattern': {'value': 'break', 'flags': [], 'raw': '"break"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 26: {'name': 'CONTINUE', 'pattern': {'value': 'continue', 'flags': [], 'raw': '"continue"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 27: {'name': 'ELSE', 'pattern': {'value': 'else', 'flags': [], 'raw': '"else"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 28: {'name': 'IF', 'pattern': {'value': 'if', 'flags': [], 'raw': '"if"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 29: {'name': 'ASSERT', 'pattern': {'value': 'assert', 'flags': [], 'raw': '"assert"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 30: {'name': 'BLACKBOX', 'pattern': {'value': 'blackbox', 'flags': [], 'raw': '"blackbox"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 31: {'name': 'DOLLAR', 'pattern': {'value': '$', 'flags': [], 'raw': '"$"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 32: {'name': 'ASM', 'pattern': {'value': 'asm', 'flags': [], 'raw': '"asm"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 33: {'name': 'VOLATILE', 'pattern': {'value': 'volatile', 'flags': [], 'raw': '"volatile"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 34: {'name': '__ANON_1', 'pattern': {'value': '<<<', 'flags': [], 'raw': '"<<<"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 35: {'name': '__ANON_2', 'pattern': {'value': '>>>', 'flags': [], 'raw': '">>>"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 36: {'name': 'PLUS', 'pattern': {'value': '+', 'flags': [], 'raw': '"+"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 37: {'name': 'MINUS', 'pattern': {'value': '-', 'flags': [], 'raw': '"-"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 38: {'name': 'BANG', 'pattern': {'value': '!', 'flags': [], 'raw': '"!"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 39: {'name': 'STAR', 'pattern': {'value': '*', 'flags': [], 'raw': '"*"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 40: {'name': 'SLASH', 'pattern': {'value': '/', 'flags': [], 'raw': '"/"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 41: {'name': 'PERCENT', 'pattern': {'value': '%', 'flags': [], 'raw': '"%"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 42: {'name': '__ANON_3', 'pattern': {'value': '<<', 'flags': [], 'raw': '"<<"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 43: {'name': '__ANON_4', 'pattern': {'value': '>>', 'flags': [], 'raw': '">>"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 44: {'name': 'LESSTHAN', 'pattern': {'value': '<', 'flags': [], 'raw': '"<"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 45: {'name': 'MORETHAN', 'pattern': {'value': '>', 'flags': [], 'raw': '">"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 46: {'name': '__ANON_5', 'pattern': {'value': '<=', 'flags': [], 'raw': '"<="', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 47: {'name': '__ANON_6', 'pattern': {'value': '>=', 'flags': [], 'raw': '">="', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 48: {'name': '__ANON_7', 'pattern': {'value': '==', 'flags': [], 'raw': '"=="', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 49: {'name': '__ANON_8', 'pattern': {'value': '!=', 'flags': [], 'raw': '"!="', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 50: {'name': '__ANON_9', 'pattern': {'value': '&&', 'flags': [], 'raw': '"&&"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 51: {'name': '__ANON_10', 'pattern': {'value': '||', 'flags': [], 'raw': '"||"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 52: {'name': '__ANON_11', 'pattern': {'value': '**', 'flags': [], 'raw': '"**"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 53: {'name': 'CIRCUMFLEX', 'pattern': {'value': '^', 'flags': [], 'raw': '"^"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 54: {'name': 'VBAR', 'pattern': {'value': '|', 'flags': [], 'raw': '"|"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 55: {'name': 'AMPERSAND', 'pattern': {'value': '&', 'flags': [], 'raw': '"&"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 56: {'name': 'LSQB', 'pattern': {'value': '[', 'flags': [], 'raw': '"["', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 57: {'name': 'RSQB', 'pattern': {'value': ']', 'flags': [], 'raw': '"]"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 58: {'name': 'COLON', 'pattern': {'value': ':', 'flags': [], 'raw': '":"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 59: {'name': 'CAST', 'pattern': {'value': 'cast', 'flags': [], 'raw': '"cast"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 60: {'name': 'AS', 'pattern': {'value': 'as', 'flags': [], 'raw': '"as"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 61: {'name': 'DEREF', 'pattern': {'value': 'deref', 'flags': [], 'raw': '"deref"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 62: {'name': 'ADDR', 'pattern': {'value': 'addr', 'flags': [], 'raw': '"addr"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 63: {'name': 'THEN', 'pattern': {'value': 'then', 'flags': [], 'raw': '"then"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 64: {'name': 'TRUE', 'pattern': {'value': 'true', 'flags': [], 'raw': '"true"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 65: {'name': 'FALSE', 'pattern': {'value': 'false', 'flags': [], 'raw': '"false"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 66: {'name': 'NONE', 'pattern': {'value': 'none', 'flags': [], 'raw': '"none"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 67: {'name': 'DOT', 'pattern': {'value': '.', 'flags': [], 'raw': '"."', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 68: {'name': 'TILDE', 'pattern': {'value': '~', 'flags': [], 'raw': '"~"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 69: {'name': 'I8', 'pattern': {'value': 'i8', 'flags': [], 'raw': '"i8"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 70: {'name': 'I16', 'pattern': {'value': 'i16', 'flags': [], 'raw': '"i16"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 71: {'name': 'I32', 'pattern': {'value': 'i32', 'flags': [], 'raw': '"i32"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 72: {'name': 'I64', 'pattern': {'value': 'i64', 'flags': [], 'raw': '"i64"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 73: {'name': 'U8', 'pattern': {'value': 'u8', 'flags': [], 'raw': '"u8"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 74: {'name': 'U16', 'pattern': {'value': 'u16', 'flags': [], 'raw': '"u16"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 75: {'name': 'U32', 'pattern': {'value': 'u32', 'flags': [], 'raw': '"u32"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 76: {'name': 'U64', 'pattern': {'value': 'u64', 'flags': [], 'raw': '"u64"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 77: {'name': 'F16', 'pattern': {'value': 'f16', 'flags': [], 'raw': '"f16"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 78: {'name': 'F32', 'pattern': {'value': 'f32', 'flags': [], 'raw': '"f32"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 79: {'name': 'F64', 'pattern': {'value': 'f64', 'flags': [], 'raw': '"f64"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 80: {'name': 'BOOL', 'pattern': {'value': 'bool', 'flags': [], 'raw': '"bool"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 81: {'name': 'VOID', 'pattern': {'value': 'void', 'flags': [], 'raw': '"void"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 82: {'name': 'HASH', 'pattern': {'value': '#', 'flags': [], 'raw': '"#"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 83: {'name': 'DICT', 'pattern': {'value': 'dict', 'flags': [], 'raw': '"dict"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 84: {'name': 'LIST', 'pattern': {'value': 'list', 'flags': [], 'raw': '"list"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 85: {'name': 'AT', 'pattern': {'value': '@', 'flags': [], 'raw': '"@"', '__type__': 'PatternStr'}, 'priority': 0, '__type__': 'TerminalDef'}, 86: {'origin': {'name': Token('RULE', 'start'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'attribute', '__type__': 'NonTerminal'}, {'name': 'module', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 87: {'origin': {'name': Token('RULE', 'module'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'MODULE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'attribute_name', '__type__': 'NonTerminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': '__module_star_0', '__type__': 'NonTerminal'}, {'name': '__module_star_1', '__type__': 'NonTerminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 88: {'origin': {'name': Token('RULE', 'module'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'MODULE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'attribute_name', '__type__': 'NonTerminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': '__module_star_0', '__type__': 'NonTerminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 89: {'origin': {'name': Token('RULE', 'module'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'MODULE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'attribute_name', '__type__': 'NonTerminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': '__module_star_1', '__type__': 'NonTerminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}], 'order': 2, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 90: {'origin': {'name': Token('RULE', 'module'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'MODULE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'attribute_name', '__type__': 'NonTerminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}], 'order': 3, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 91: {'origin': {'name': Token('RULE', 'function'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'DEF', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'name', '__type__': 'NonTerminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'fn_args', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': '__ANON_0', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'type', '__type__': 'NonTerminal'}, {'name': 'attribute_name', '__type__': 'NonTerminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'stmt_body', '__type__': 'NonTerminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 92: {'origin': {'name': Token('RULE', 'function'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'DEF', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'name', '__type__': 'NonTerminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'fn_args', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': '__ANON_0', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'type', '__type__': 'NonTerminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'stmt_body', '__type__': 'NonTerminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [False, False, False, False, False, False, False, True, False, False, False], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 93: {'origin': {'name': Token('RULE', 'function'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'DEF', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'name', '__type__': 'NonTerminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': '__ANON_0', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'type', '__type__': 'NonTerminal'}, {'name': 'attribute_name', '__type__': 'NonTerminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'stmt_body', '__type__': 'NonTerminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}], 'order': 2, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [False, False, False, True, False, False, False, False, False, False, False], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 94: {'origin': {'name': Token('RULE', 'function'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'DEF', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'name', '__type__': 'NonTerminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': '__ANON_0', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'type', '__type__': 'NonTerminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'stmt_body', '__type__': 'NonTerminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}], 'order': 3, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [False, False, False, True, False, False, False, True, False, False, False], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 95: {'origin': {'name': Token('RULE', 'fn_args'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'def_var', '__type__': 'NonTerminal'}, {'name': '__fn_args_star_2', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 96: {'origin': {'name': Token('RULE', 'fn_args'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'def_var', '__type__': 'NonTerminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 97: {'origin': {'name': Token('RULE', 'stmt_body'), '__type__': 'NonTerminal'}, 'expansion': [{'name': '__stmt_body_star_3', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 98: {'origin': {'name': Token('RULE', 'stmt_body'), '__type__': 'NonTerminal'}, 'expansion': [], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 99: {'origin': {'name': Token('RULE', 'stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'declare_stmt', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 100: {'origin': {'name': Token('RULE', 'stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'buffer_store_stmt', '__type__': 'NonTerminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 101: {'origin': {'name': Token('RULE', 'stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'assign_stmt', '__type__': 'NonTerminal'}], 'order': 2, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 102: {'origin': {'name': Token('RULE', 'stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'return_stmt', '__type__': 'NonTerminal'}], 'order': 3, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 103: {'origin': {'name': Token('RULE', 'stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'let_stmt', '__type__': 'NonTerminal'}], 'order': 4, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 104: {'origin': {'name': Token('RULE', 'stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'for_stmt', '__type__': 'NonTerminal'}], 'order': 5, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 105: {'origin': {'name': Token('RULE', 'stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'while_stmt', '__type__': 'NonTerminal'}], 'order': 6, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 106: {'origin': {'name': Token('RULE', 'stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'break_stmt', '__type__': 'NonTerminal'}], 'order': 7, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 107: {'origin': {'name': Token('RULE', 'stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'continue_stmt', '__type__': 'NonTerminal'}], 'order': 8, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 108: {'origin': {'name': Token('RULE', 'stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'if_stmt', '__type__': 'NonTerminal'}], 'order': 9, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 109: {'origin': {'name': Token('RULE', 'stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'assert_stmt', '__type__': 'NonTerminal'}], 'order': 10, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 110: {'origin': {'name': Token('RULE', 'stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'black_box_stmt', '__type__': 'NonTerminal'}], 'order': 11, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 111: {'origin': {'name': Token('RULE', 'stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'asm_stmt', '__type__': 'NonTerminal'}], 'order': 12, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 112: {'origin': {'name': Token('RULE', 'stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'launch_kernel_stmt', '__type__': 'NonTerminal'}], 'order': 13, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 113: {'origin': {'name': Token('RULE', 'stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'evaluate_stmt', '__type__': 'NonTerminal'}], 'order': 14, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 114: {'origin': {'name': Token('RULE', 'declare_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'static_decl', '__type__': 'NonTerminal'}, {'name': 'DECL', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'def_var', '__type__': 'NonTerminal'}, {'name': 'EQUAL', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'attribute_name', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 115: {'origin': {'name': Token('RULE', 'declare_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'static_decl', '__type__': 'NonTerminal'}, {'name': 'DECL', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'def_var', '__type__': 'NonTerminal'}, {'name': 'EQUAL', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [False, False, False, False, False, False, True], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 116: {'origin': {'name': Token('RULE', 'declare_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'static_decl', '__type__': 'NonTerminal'}, {'name': 'DECL', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'def_var', '__type__': 'NonTerminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'attribute_name', '__type__': 'NonTerminal'}], 'order': 2, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [False, False, False, True, False, False], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 117: {'origin': {'name': Token('RULE', 'declare_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'static_decl', '__type__': 'NonTerminal'}, {'name': 'DECL', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'def_var', '__type__': 'NonTerminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 3, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [False, False, False, True, False, True], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 118: {'origin': {'name': Token('RULE', 'declare_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'DECL', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'def_var', '__type__': 'NonTerminal'}, {'name': 'EQUAL', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'attribute_name', '__type__': 'NonTerminal'}], 'order': 4, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [True, False, False, False, False, False, False], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 119: {'origin': {'name': Token('RULE', 'declare_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'DECL', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'def_var', '__type__': 'NonTerminal'}, {'name': 'EQUAL', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 5, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [True, False, False, False, False, False, True], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 120: {'origin': {'name': Token('RULE', 'declare_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'DECL', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'def_var', '__type__': 'NonTerminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'attribute_name', '__type__': 'NonTerminal'}], 'order': 6, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [True, False, False, True, False, False], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 121: {'origin': {'name': Token('RULE', 'declare_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'DECL', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'def_var', '__type__': 'NonTerminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 7, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [True, False, False, True, False, True], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 122: {'origin': {'name': Token('RULE', 'static_decl'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'STATIC', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 123: {'origin': {'name': Token('RULE', 'buffer_store_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'protected', '__type__': 'NonTerminal'}, {'name': 'get_item_expr', '__type__': 'NonTerminal'}, {'name': 'EQUAL', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 124: {'origin': {'name': Token('RULE', 'buffer_store_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'get_item_expr', '__type__': 'NonTerminal'}, {'name': 'EQUAL', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [True, False, False, False, False], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 125: {'origin': {'name': Token('RULE', 'protected'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'PROTECTED', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 126: {'origin': {'name': Token('RULE', 'assign_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'var', '__type__': 'NonTerminal'}, {'name': 'EQUAL', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 127: {'origin': {'name': Token('RULE', 'return_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'RETURN', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 128: {'origin': {'name': Token('RULE', 'return_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'RETURN', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [False, True, False], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 129: {'origin': {'name': Token('RULE', 'let_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'LET', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': '__let_stmt_plus_4', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'IN', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'stmt_body', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 130: {'origin': {'name': Token('RULE', 'let_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'LET', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': '__let_stmt_plus_4', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [False, False, False, False, True, False], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 131: {'origin': {'name': Token('RULE', 'for_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'FOR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'for_ind_var', '__type__': 'NonTerminal'}, {'name': 'IN', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'call_expr', '__type__': 'NonTerminal'}, {'name': 'attribute_name', '__type__': 'NonTerminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'stmt_body', '__type__': 'NonTerminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 132: {'origin': {'name': Token('RULE', 'for_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'FOR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'for_ind_var', '__type__': 'NonTerminal'}, {'name': 'IN', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'call_expr', '__type__': 'NonTerminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'stmt_body', '__type__': 'NonTerminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [False, False, False, False, True, False, False, False], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 133: {'origin': {'name': Token('RULE', 'for_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'FOR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'for_ind_var', '__type__': 'NonTerminal'}, {'name': 'IN', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'call_expr', '__type__': 'NonTerminal'}, {'name': 'task_mapping', '__type__': 'NonTerminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'stmt_body', '__type__': 'NonTerminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}], 'order': 2, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 134: {'origin': {'name': Token('RULE', 'for_ind_var'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'def_var', '__type__': 'NonTerminal'}, {'name': '__fn_args_star_2', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 135: {'origin': {'name': Token('RULE', 'for_ind_var'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'def_var', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 136: {'origin': {'name': Token('RULE', 'task_mapping'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'ON', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 137: {'origin': {'name': Token('RULE', 'while_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'WHILE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'stmt_body', '__type__': 'NonTerminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 138: {'origin': {'name': Token('RULE', 'break_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'BREAK', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 139: {'origin': {'name': Token('RULE', 'continue_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'CONTINUE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 140: {'origin': {'name': Token('RULE', 'if_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'IF', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'stmt_body', '__type__': 'NonTerminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'ELSE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'stmt_body', '__type__': 'NonTerminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 141: {'origin': {'name': Token('RULE', 'if_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'IF', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'stmt_body', '__type__': 'NonTerminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [False, False, False, False, False, True], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 142: {'origin': {'name': Token('RULE', 'assert_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'ASSERT', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'STRING', 'filter_out': False, '__type__': 'Terminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 143: {'origin': {'name': Token('RULE', 'assert_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'ASSERT', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [False, False, False, True, False, False], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 144: {'origin': {'name': Token('RULE', 'black_box_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'BLACKBOX', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': '__black_box_stmt_star_5', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'DOLLAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'ANYWILDCARD', 'filter_out': False, '__type__': 'Terminal'}, {'name': 'DOLLAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 145: {'origin': {'name': Token('RULE', 'black_box_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'BLACKBOX', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'DOLLAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'ANYWILDCARD', 'filter_out': False, '__type__': 'Terminal'}, {'name': 'DOLLAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 146: {'origin': {'name': Token('RULE', 'black_box_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'BLACKBOX', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'DOLLAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'ANYWILDCARD', 'filter_out': False, '__type__': 'Terminal'}, {'name': 'DOLLAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 2, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [False, False, True, False, False, False, False, False, False, False], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 147: {'origin': {'name': Token('RULE', 'asm_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'volatile', '__type__': 'NonTerminal'}, {'name': 'ASM', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'STRING', 'filter_out': False, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'asm_group', '__type__': 'NonTerminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'asm_group', '__type__': 'NonTerminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 148: {'origin': {'name': Token('RULE', 'asm_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'volatile', '__type__': 'NonTerminal'}, {'name': 'ASM', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'STRING', 'filter_out': False, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'asm_group', '__type__': 'NonTerminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [False, False, False, False, False, False, False, False, True, False, False, False], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 149: {'origin': {'name': Token('RULE', 'asm_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'volatile', '__type__': 'NonTerminal'}, {'name': 'ASM', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'STRING', 'filter_out': False, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'asm_group', '__type__': 'NonTerminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 2, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [False, False, False, False, False, True, False, False, False, False, False, False], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 150: {'origin': {'name': Token('RULE', 'asm_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'volatile', '__type__': 'NonTerminal'}, {'name': 'ASM', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'STRING', 'filter_out': False, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 3, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [False, False, False, False, False, True, False, False, True, False, False, False], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 151: {'origin': {'name': Token('RULE', 'asm_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'ASM', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'STRING', 'filter_out': False, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'asm_group', '__type__': 'NonTerminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'asm_group', '__type__': 'NonTerminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 4, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [True, False, False, False, False, False, False, False, False, False, False, False], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 152: {'origin': {'name': Token('RULE', 'asm_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'ASM', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'STRING', 'filter_out': False, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'asm_group', '__type__': 'NonTerminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 5, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [True, False, False, False, False, False, False, False, True, False, False, False], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 153: {'origin': {'name': Token('RULE', 'asm_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'ASM', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'STRING', 'filter_out': False, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'asm_group', '__type__': 'NonTerminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 6, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [True, False, False, False, False, True, False, False, False, False, False, False], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 154: {'origin': {'name': Token('RULE', 'asm_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'ASM', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'STRING', 'filter_out': False, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 7, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [True, False, False, False, False, True, False, False, True, False, False, False], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 155: {'origin': {'name': Token('RULE', 'volatile'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'VOLATILE', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 156: {'origin': {'name': Token('RULE', 'asm_label'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'STRING', 'filter_out': False, '__type__': 'Terminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 157: {'origin': {'name': Token('RULE', 'asm_group'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'asm_label', '__type__': 'NonTerminal'}, {'name': '__asm_group_star_6', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 158: {'origin': {'name': Token('RULE', 'asm_group'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'asm_label', '__type__': 'NonTerminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 159: {'origin': {'name': Token('RULE', 'launch_kernel_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'fn_name', '__type__': 'NonTerminal'}, {'name': '__ANON_1', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'dim3', '__type__': 'NonTerminal'}, {'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'dim3', '__type__': 'NonTerminal'}, {'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'INT', 'filter_out': False, '__type__': 'Terminal'}, {'name': '__ANON_2', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': '__black_box_stmt_star_5', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 160: {'origin': {'name': Token('RULE', 'launch_kernel_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'fn_name', '__type__': 'NonTerminal'}, {'name': '__ANON_1', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'dim3', '__type__': 'NonTerminal'}, {'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'dim3', '__type__': 'NonTerminal'}, {'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'INT', 'filter_out': False, '__type__': 'Terminal'}, {'name': '__ANON_2', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 161: {'origin': {'name': Token('RULE', 'launch_kernel_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'fn_name', '__type__': 'NonTerminal'}, {'name': '__ANON_1', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'dim3', '__type__': 'NonTerminal'}, {'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'dim3', '__type__': 'NonTerminal'}, {'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': '__ANON_2', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': '__black_box_stmt_star_5', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 2, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [False, False, False, False, False, False, True, False, False, False, False, False, False], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 162: {'origin': {'name': Token('RULE', 'launch_kernel_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'fn_name', '__type__': 'NonTerminal'}, {'name': '__ANON_1', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'dim3', '__type__': 'NonTerminal'}, {'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'dim3', '__type__': 'NonTerminal'}, {'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': '__ANON_2', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 3, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [False, False, False, False, False, False, True, False, False, False, False, False], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 163: {'origin': {'name': Token('RULE', 'dim3'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 164: {'origin': {'name': Token('RULE', 'evaluate_stmt'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 165: {'origin': {'name': Token('RULE', 'expr'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': '_unary_op', '__type__': 'NonTerminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': 'unary_op', 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 166: {'origin': {'name': Token('RULE', 'expr'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': '_binary_op', '__type__': 'NonTerminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}], 'order': 1, 'alias': 'binary_op', 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 167: {'origin': {'name': Token('RULE', 'expr'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'fn_call_expr', '__type__': 'NonTerminal'}], 'order': 2, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 168: {'origin': {'name': Token('RULE', '_unary_op'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'PLUS', 'filter_out': False, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 169: {'origin': {'name': Token('RULE', '_unary_op'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'MINUS', 'filter_out': False, '__type__': 'Terminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 170: {'origin': {'name': Token('RULE', '_unary_op'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'BANG', 'filter_out': False, '__type__': 'Terminal'}], 'order': 2, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 171: {'origin': {'name': Token('RULE', '_binary_op'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'PLUS', 'filter_out': False, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 172: {'origin': {'name': Token('RULE', '_binary_op'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'MINUS', 'filter_out': False, '__type__': 'Terminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 173: {'origin': {'name': Token('RULE', '_binary_op'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'STAR', 'filter_out': False, '__type__': 'Terminal'}], 'order': 2, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 174: {'origin': {'name': Token('RULE', '_binary_op'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'SLASH', 'filter_out': False, '__type__': 'Terminal'}], 'order': 3, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 175: {'origin': {'name': Token('RULE', '_binary_op'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'PERCENT', 'filter_out': False, '__type__': 'Terminal'}], 'order': 4, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 176: {'origin': {'name': Token('RULE', '_binary_op'), '__type__': 'NonTerminal'}, 'expansion': [{'name': '__ANON_3', 'filter_out': False, '__type__': 'Terminal'}], 'order': 5, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 177: {'origin': {'name': Token('RULE', '_binary_op'), '__type__': 'NonTerminal'}, 'expansion': [{'name': '__ANON_4', 'filter_out': False, '__type__': 'Terminal'}], 'order': 6, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 178: {'origin': {'name': Token('RULE', '_binary_op'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'LESSTHAN', 'filter_out': False, '__type__': 'Terminal'}], 'order': 7, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 179: {'origin': {'name': Token('RULE', '_binary_op'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'MORETHAN', 'filter_out': False, '__type__': 'Terminal'}], 'order': 8, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 180: {'origin': {'name': Token('RULE', '_binary_op'), '__type__': 'NonTerminal'}, 'expansion': [{'name': '__ANON_5', 'filter_out': False, '__type__': 'Terminal'}], 'order': 9, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 181: {'origin': {'name': Token('RULE', '_binary_op'), '__type__': 'NonTerminal'}, 'expansion': [{'name': '__ANON_6', 'filter_out': False, '__type__': 'Terminal'}], 'order': 10, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 182: {'origin': {'name': Token('RULE', '_binary_op'), '__type__': 'NonTerminal'}, 'expansion': [{'name': '__ANON_7', 'filter_out': False, '__type__': 'Terminal'}], 'order': 11, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 183: {'origin': {'name': Token('RULE', '_binary_op'), '__type__': 'NonTerminal'}, 'expansion': [{'name': '__ANON_8', 'filter_out': False, '__type__': 'Terminal'}], 'order': 12, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 184: {'origin': {'name': Token('RULE', '_binary_op'), '__type__': 'NonTerminal'}, 'expansion': [{'name': '__ANON_9', 'filter_out': False, '__type__': 'Terminal'}], 'order': 13, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 185: {'origin': {'name': Token('RULE', '_binary_op'), '__type__': 'NonTerminal'}, 'expansion': [{'name': '__ANON_10', 'filter_out': False, '__type__': 'Terminal'}], 'order': 14, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 186: {'origin': {'name': Token('RULE', '_binary_op'), '__type__': 'NonTerminal'}, 'expansion': [{'name': '__ANON_11', 'filter_out': False, '__type__': 'Terminal'}], 'order': 15, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 187: {'origin': {'name': Token('RULE', '_binary_op'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'CIRCUMFLEX', 'filter_out': False, '__type__': 'Terminal'}], 'order': 16, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 188: {'origin': {'name': Token('RULE', '_binary_op'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'VBAR', 'filter_out': False, '__type__': 'Terminal'}], 'order': 17, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 189: {'origin': {'name': Token('RULE', '_binary_op'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'AMPERSAND', 'filter_out': False, '__type__': 'Terminal'}], 'order': 18, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 190: {'origin': {'name': Token('RULE', 'fn_call_expr'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'get_item_expr', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 191: {'origin': {'name': Token('RULE', 'fn_call_expr'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'cast_expr', '__type__': 'NonTerminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 192: {'origin': {'name': Token('RULE', 'fn_call_expr'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'deref_expr', '__type__': 'NonTerminal'}], 'order': 2, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 193: {'origin': {'name': Token('RULE', 'fn_call_expr'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'addr_expr', '__type__': 'NonTerminal'}], 'order': 3, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 194: {'origin': {'name': Token('RULE', 'fn_call_expr'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'call_expr', '__type__': 'NonTerminal'}], 'order': 4, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 195: {'origin': {'name': Token('RULE', 'fn_call_expr'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'let_expr', '__type__': 'NonTerminal'}], 'order': 5, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 196: {'origin': {'name': Token('RULE', 'fn_call_expr'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'if_then_else_expr', '__type__': 'NonTerminal'}], 'order': 6, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 197: {'origin': {'name': Token('RULE', 'fn_call_expr'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'atom', '__type__': 'NonTerminal'}], 'order': 7, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 198: {'origin': {'name': Token('RULE', 'get_item_expr'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'LSQB', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': '__get_item_expr_star_7', '__type__': 'NonTerminal'}, {'name': 'RSQB', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 199: {'origin': {'name': Token('RULE', 'get_item_expr'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'LSQB', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'RSQB', 'filter_out': True, '__type__': 'Terminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 200: {'origin': {'name': Token('RULE', 'get_item_expr'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'LSQB', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'slice', '__type__': 'NonTerminal'}, {'name': '__get_item_expr_star_7', '__type__': 'NonTerminal'}, {'name': 'RSQB', 'filter_out': True, '__type__': 'Terminal'}], 'order': 2, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 201: {'origin': {'name': Token('RULE', 'get_item_expr'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'LSQB', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'slice', '__type__': 'NonTerminal'}, {'name': 'RSQB', 'filter_out': True, '__type__': 'Terminal'}], 'order': 3, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 202: {'origin': {'name': Token('RULE', 'slice'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'COLON', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 203: {'origin': {'name': Token('RULE', 'slice'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'COLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [False, False, True], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 204: {'origin': {'name': Token('RULE', 'slice'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'COLON', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}], 'order': 2, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [True, False, False], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 205: {'origin': {'name': Token('RULE', 'slice'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'COLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 3, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [True, False, True], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 206: {'origin': {'name': Token('RULE', 'call_expr'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'fn_name', '__type__': 'NonTerminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'call_args', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 207: {'origin': {'name': Token('RULE', 'call_expr'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'fn_name', '__type__': 'NonTerminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [False, False, True, False], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 208: {'origin': {'name': Token('RULE', 'call_args'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'expr', '__type__': 'NonTerminal'}, {'name': '__black_box_stmt_star_5', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 209: {'origin': {'name': Token('RULE', 'call_args'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'expr', '__type__': 'NonTerminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 210: {'origin': {'name': Token('RULE', 'call_args'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'keyword_arg', '__type__': 'NonTerminal'}, {'name': '__call_args_star_8', '__type__': 'NonTerminal'}], 'order': 2, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 211: {'origin': {'name': Token('RULE', 'call_args'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'keyword_arg', '__type__': 'NonTerminal'}], 'order': 3, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 212: {'origin': {'name': Token('RULE', 'call_args'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'expr', '__type__': 'NonTerminal'}, {'name': '__black_box_stmt_star_5', '__type__': 'NonTerminal'}, {'name': '__call_args_star_8', '__type__': 'NonTerminal'}], 'order': 4, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 213: {'origin': {'name': Token('RULE', 'call_args'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'expr', '__type__': 'NonTerminal'}, {'name': '__call_args_star_8', '__type__': 'NonTerminal'}], 'order': 5, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 214: {'origin': {'name': Token('RULE', 'keyword_arg'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'name', '__type__': 'NonTerminal'}, {'name': 'EQUAL', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 215: {'origin': {'name': Token('RULE', 'cast_expr'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'CAST', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'AS', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'type', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 216: {'origin': {'name': Token('RULE', 'deref_expr'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'DEREF', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 217: {'origin': {'name': Token('RULE', 'addr_expr'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'ADDR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 218: {'origin': {'name': Token('RULE', 'let_expr'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'LET', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'def_var', '__type__': 'NonTerminal'}, {'name': 'EQUAL', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'IN', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 219: {'origin': {'name': Token('RULE', 'if_then_else_expr'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'IF', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'THEN', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'ELSE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LBRACE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'RBRACE', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 220: {'origin': {'name': Token('RULE', 'atom'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 221: {'origin': {'name': Token('RULE', 'atom'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'get_attr', '__type__': 'NonTerminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 222: {'origin': {'name': Token('RULE', 'atom'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'TRUE', 'filter_out': True, '__type__': 'Terminal'}], 'order': 2, 'alias': 'true', 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 223: {'origin': {'name': Token('RULE', 'atom'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'FALSE', 'filter_out': True, '__type__': 'Terminal'}], 'order': 3, 'alias': 'false', 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 224: {'origin': {'name': Token('RULE', 'atom'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'NONE', 'filter_out': True, '__type__': 'Terminal'}], 'order': 4, 'alias': 'none', 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 225: {'origin': {'name': Token('RULE', 'atom'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'symbol_var', '__type__': 'NonTerminal'}], 'order': 5, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 226: {'origin': {'name': Token('RULE', 'atom'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'var', '__type__': 'NonTerminal'}], 'order': 6, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 227: {'origin': {'name': Token('RULE', 'atom'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'STRING', 'filter_out': False, '__type__': 'Terminal'}], 'order': 7, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 228: {'origin': {'name': Token('RULE', 'atom'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'INT', 'filter_out': False, '__type__': 'Terminal'}], 'order': 8, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 229: {'origin': {'name': Token('RULE', 'atom'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'SIGNED_FLOAT', 'filter_out': False, '__type__': 'Terminal'}], 'order': 9, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 230: {'origin': {'name': Token('RULE', 'get_attr'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'DOT', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'name', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 231: {'origin': {'name': Token('RULE', 'type'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'data_type', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 232: {'origin': {'name': Token('RULE', 'type'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'tensor_type', '__type__': 'NonTerminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 233: {'origin': {'name': Token('RULE', 'type'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'ptr_type', '__type__': 'NonTerminal'}], 'order': 2, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 234: {'origin': {'name': Token('RULE', 'type'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'fn_type', '__type__': 'NonTerminal'}], 'order': 3, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 235: {'origin': {'name': Token('RULE', 'fn_type'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'type', '__type__': 'NonTerminal'}, {'name': '__fn_type_star_9', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': '__ANON_0', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'type', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 236: {'origin': {'name': Token('RULE', 'fn_type'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'type', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': '__ANON_0', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'type', '__type__': 'NonTerminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 237: {'origin': {'name': Token('RULE', 'fn_type'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': '__ANON_0', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'type', '__type__': 'NonTerminal'}], 'order': 2, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [False, True, False, False, False], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 238: {'origin': {'name': Token('RULE', 'ptr_type'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'TILDE', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'type', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 239: {'origin': {'name': Token('RULE', 'tensor_type'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'data_type', '__type__': 'NonTerminal'}, {'name': 'LESSTHAN', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'tensor_tt_shape', '__type__': 'NonTerminal'}, {'name': '__tensor_type_star_10', '__type__': 'NonTerminal'}, {'name': 'tensor_layout', '__type__': 'NonTerminal'}, {'name': 'MORETHAN', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 240: {'origin': {'name': Token('RULE', 'tensor_type'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'data_type', '__type__': 'NonTerminal'}, {'name': 'LESSTHAN', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'tensor_tt_shape', '__type__': 'NonTerminal'}, {'name': '__tensor_type_star_10', '__type__': 'NonTerminal'}, {'name': 'MORETHAN', 'filter_out': True, '__type__': 'Terminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [False, False, False, False, True, False], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 241: {'origin': {'name': Token('RULE', 'tensor_type'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'data_type', '__type__': 'NonTerminal'}, {'name': 'LESSTHAN', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'tensor_tt_shape', '__type__': 'NonTerminal'}, {'name': 'tensor_layout', '__type__': 'NonTerminal'}, {'name': 'MORETHAN', 'filter_out': True, '__type__': 'Terminal'}], 'order': 2, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 242: {'origin': {'name': Token('RULE', 'tensor_type'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'data_type', '__type__': 'NonTerminal'}, {'name': 'LESSTHAN', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'tensor_tt_shape', '__type__': 'NonTerminal'}, {'name': 'MORETHAN', 'filter_out': True, '__type__': 'Terminal'}], 'order': 3, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [False, False, False, True, False], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 243: {'origin': {'name': Token('RULE', 'tensor_type'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'data_type', '__type__': 'NonTerminal'}, {'name': 'LESSTHAN', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'tensor_layout', '__type__': 'NonTerminal'}, {'name': 'MORETHAN', 'filter_out': True, '__type__': 'Terminal'}], 'order': 4, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [False, False, True, False, False], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 244: {'origin': {'name': Token('RULE', 'tensor_type'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'data_type', '__type__': 'NonTerminal'}, {'name': 'LESSTHAN', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'MORETHAN', 'filter_out': True, '__type__': 'Terminal'}], 'order': 5, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [False, False, True, True, False], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 245: {'origin': {'name': Token('RULE', 'tensor_tt_shape'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'expr', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 246: {'origin': {'name': Token('RULE', 'tensor_layout'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 247: {'origin': {'name': Token('RULE', 'data_type'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'I8', 'filter_out': False, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 248: {'origin': {'name': Token('RULE', 'data_type'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'I16', 'filter_out': False, '__type__': 'Terminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 249: {'origin': {'name': Token('RULE', 'data_type'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'I32', 'filter_out': False, '__type__': 'Terminal'}], 'order': 2, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 250: {'origin': {'name': Token('RULE', 'data_type'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'I64', 'filter_out': False, '__type__': 'Terminal'}], 'order': 3, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 251: {'origin': {'name': Token('RULE', 'data_type'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'U8', 'filter_out': False, '__type__': 'Terminal'}], 'order': 4, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 252: {'origin': {'name': Token('RULE', 'data_type'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'U16', 'filter_out': False, '__type__': 'Terminal'}], 'order': 5, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 253: {'origin': {'name': Token('RULE', 'data_type'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'U32', 'filter_out': False, '__type__': 'Terminal'}], 'order': 6, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 254: {'origin': {'name': Token('RULE', 'data_type'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'U64', 'filter_out': False, '__type__': 'Terminal'}], 'order': 7, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 255: {'origin': {'name': Token('RULE', 'data_type'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'F16', 'filter_out': False, '__type__': 'Terminal'}], 'order': 8, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 256: {'origin': {'name': Token('RULE', 'data_type'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'F32', 'filter_out': False, '__type__': 'Terminal'}], 'order': 9, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 257: {'origin': {'name': Token('RULE', 'data_type'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'F64', 'filter_out': False, '__type__': 'Terminal'}], 'order': 10, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 258: {'origin': {'name': Token('RULE', 'data_type'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'BOOL', 'filter_out': False, '__type__': 'Terminal'}], 'order': 11, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 259: {'origin': {'name': Token('RULE', 'data_type'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'VOID', 'filter_out': False, '__type__': 'Terminal'}], 'order': 12, 'alias': None, 'options': {'keep_all_tokens': True, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 260: {'origin': {'name': Token('RULE', 'attribute_name'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'HASH', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'IDENT', 'filter_out': False, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 261: {'origin': {'name': Token('RULE', 'attribute'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'attribute_name', '__type__': 'NonTerminal'}, {'name': 'EQUAL', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'dict', '__type__': 'NonTerminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 262: {'origin': {'name': Token('RULE', 'value'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'dict', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 263: {'origin': {'name': Token('RULE', 'value'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'list', '__type__': 'NonTerminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 264: {'origin': {'name': Token('RULE', 'value'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'expr', '__type__': 'NonTerminal'}], 'order': 2, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 265: {'origin': {'name': Token('RULE', 'value'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'type', '__type__': 'NonTerminal'}], 'order': 3, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': True, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 266: {'origin': {'name': Token('RULE', 'pair'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'STRING', 'filter_out': False, '__type__': 'Terminal'}, {'name': 'COLON', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'value', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 267: {'origin': {'name': Token('RULE', 'dict'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'DICT', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'pair', '__type__': 'NonTerminal'}, {'name': '__dict_star_11', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 268: {'origin': {'name': Token('RULE', 'dict'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'DICT', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'pair', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 269: {'origin': {'name': Token('RULE', 'dict'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'DICT', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}], 'order': 2, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 270: {'origin': {'name': Token('RULE', 'list'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'LIST', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'value', '__type__': 'NonTerminal'}, {'name': '__list_star_12', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 271: {'origin': {'name': Token('RULE', 'list'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'LIST', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'value', '__type__': 'NonTerminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 272: {'origin': {'name': Token('RULE', 'list'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'LIST', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'LPAR', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'RPAR', 'filter_out': True, '__type__': 'Terminal'}], 'order': 2, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 273: {'origin': {'name': Token('RULE', 'symbol_var'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'AT', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'IDENT', 'filter_out': False, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 274: {'origin': {'name': Token('RULE', 'var'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'IDENT', 'filter_out': False, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 275: {'origin': {'name': Token('RULE', 'def_var'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'name', '__type__': 'NonTerminal'}, {'name': 'COLON', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'type', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 276: {'origin': {'name': Token('RULE', 'name'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'IDENT', 'filter_out': False, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 277: {'origin': {'name': Token('RULE', 'fn_name'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'IDENT', 'filter_out': False, '__type__': 'Terminal'}, {'name': 'attribute_name', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 278: {'origin': {'name': Token('RULE', 'fn_name'), '__type__': 'NonTerminal'}, 'expansion': [{'name': 'IDENT', 'filter_out': False, '__type__': 'Terminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': [False, True], '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 279: {'origin': {'name': '__module_star_0', '__type__': 'NonTerminal'}, 'expansion': [{'name': 'attribute', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 280: {'origin': {'name': '__module_star_0', '__type__': 'NonTerminal'}, 'expansion': [{'name': '__module_star_0', '__type__': 'NonTerminal'}, {'name': 'attribute', '__type__': 'NonTerminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 281: {'origin': {'name': '__module_star_1', '__type__': 'NonTerminal'}, 'expansion': [{'name': 'function', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 282: {'origin': {'name': '__module_star_1', '__type__': 'NonTerminal'}, 'expansion': [{'name': 'declare_stmt', '__type__': 'NonTerminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 283: {'origin': {'name': '__module_star_1', '__type__': 'NonTerminal'}, 'expansion': [{'name': '__module_star_1', '__type__': 'NonTerminal'}, {'name': 'function', '__type__': 'NonTerminal'}], 'order': 2, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 284: {'origin': {'name': '__module_star_1', '__type__': 'NonTerminal'}, 'expansion': [{'name': '__module_star_1', '__type__': 'NonTerminal'}, {'name': 'declare_stmt', '__type__': 'NonTerminal'}], 'order': 3, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 285: {'origin': {'name': '__fn_args_star_2', '__type__': 'NonTerminal'}, 'expansion': [{'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'def_var', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 286: {'origin': {'name': '__fn_args_star_2', '__type__': 'NonTerminal'}, 'expansion': [{'name': '__fn_args_star_2', '__type__': 'NonTerminal'}, {'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'def_var', '__type__': 'NonTerminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 287: {'origin': {'name': '__stmt_body_star_3', '__type__': 'NonTerminal'}, 'expansion': [{'name': 'stmt', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 288: {'origin': {'name': '__stmt_body_star_3', '__type__': 'NonTerminal'}, 'expansion': [{'name': '__stmt_body_star_3', '__type__': 'NonTerminal'}, {'name': 'stmt', '__type__': 'NonTerminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 289: {'origin': {'name': '__let_stmt_plus_4', '__type__': 'NonTerminal'}, 'expansion': [{'name': 'def_var', '__type__': 'NonTerminal'}, {'name': 'EQUAL', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 290: {'origin': {'name': '__let_stmt_plus_4', '__type__': 'NonTerminal'}, 'expansion': [{'name': '__let_stmt_plus_4', '__type__': 'NonTerminal'}, {'name': 'def_var', '__type__': 'NonTerminal'}, {'name': 'EQUAL', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}, {'name': 'SEMICOLON', 'filter_out': True, '__type__': 'Terminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 291: {'origin': {'name': '__black_box_stmt_star_5', '__type__': 'NonTerminal'}, 'expansion': [{'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 292: {'origin': {'name': '__black_box_stmt_star_5', '__type__': 'NonTerminal'}, 'expansion': [{'name': '__black_box_stmt_star_5', '__type__': 'NonTerminal'}, {'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 293: {'origin': {'name': '__asm_group_star_6', '__type__': 'NonTerminal'}, 'expansion': [{'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'asm_label', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 294: {'origin': {'name': '__asm_group_star_6', '__type__': 'NonTerminal'}, 'expansion': [{'name': '__asm_group_star_6', '__type__': 'NonTerminal'}, {'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'asm_label', '__type__': 'NonTerminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 295: {'origin': {'name': '__get_item_expr_star_7', '__type__': 'NonTerminal'}, 'expansion': [{'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 296: {'origin': {'name': '__get_item_expr_star_7', '__type__': 'NonTerminal'}, 'expansion': [{'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'slice', '__type__': 'NonTerminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 297: {'origin': {'name': '__get_item_expr_star_7', '__type__': 'NonTerminal'}, 'expansion': [{'name': '__get_item_expr_star_7', '__type__': 'NonTerminal'}, {'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'expr', '__type__': 'NonTerminal'}], 'order': 2, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 298: {'origin': {'name': '__get_item_expr_star_7', '__type__': 'NonTerminal'}, 'expansion': [{'name': '__get_item_expr_star_7', '__type__': 'NonTerminal'}, {'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'slice', '__type__': 'NonTerminal'}], 'order': 3, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 299: {'origin': {'name': '__call_args_star_8', '__type__': 'NonTerminal'}, 'expansion': [{'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'keyword_arg', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 300: {'origin': {'name': '__call_args_star_8', '__type__': 'NonTerminal'}, 'expansion': [{'name': '__call_args_star_8', '__type__': 'NonTerminal'}, {'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'keyword_arg', '__type__': 'NonTerminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 301: {'origin': {'name': '__fn_type_star_9', '__type__': 'NonTerminal'}, 'expansion': [{'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'type', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 302: {'origin': {'name': '__fn_type_star_9', '__type__': 'NonTerminal'}, 'expansion': [{'name': '__fn_type_star_9', '__type__': 'NonTerminal'}, {'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'type', '__type__': 'NonTerminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 303: {'origin': {'name': '__tensor_type_star_10', '__type__': 'NonTerminal'}, 'expansion': [{'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'tensor_tt_shape', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 304: {'origin': {'name': '__tensor_type_star_10', '__type__': 'NonTerminal'}, 'expansion': [{'name': '__tensor_type_star_10', '__type__': 'NonTerminal'}, {'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'tensor_tt_shape', '__type__': 'NonTerminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 305: {'origin': {'name': '__dict_star_11', '__type__': 'NonTerminal'}, 'expansion': [{'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'pair', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 306: {'origin': {'name': '__dict_star_11', '__type__': 'NonTerminal'}, 'expansion': [{'name': '__dict_star_11', '__type__': 'NonTerminal'}, {'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'pair', '__type__': 'NonTerminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 307: {'origin': {'name': '__list_star_12', '__type__': 'NonTerminal'}, 'expansion': [{'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'value', '__type__': 'NonTerminal'}], 'order': 0, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}, 308: {'origin': {'name': '__list_star_12', '__type__': 'NonTerminal'}, 'expansion': [{'name': '__list_star_12', '__type__': 'NonTerminal'}, {'name': 'COMMA', 'filter_out': True, '__type__': 'Terminal'}, {'name': 'value', '__type__': 'NonTerminal'}], 'order': 1, 'alias': None, 'options': {'keep_all_tokens': False, 'expand1': False, 'priority': None, 'template_source': None, 'empty_indices': (), '__type__': 'RuleOptions'}, '__type__': 'Rule'}}
-)
-Shift = 0
-Reduce = 1
-def Lark_StandAlone(**kwargs):
-  return Lark._load_from_dict(DATA, MEMO, **kwargs)
