--- python/hidet/ir/test.txt
@@ -1,11 +1,0 @@
-module Test {
-    # attr1 = {"strided_layout" : [0, 1]};
-    # test_attr = {"fn_kind": "cuda_kernel"};
-    def test(a: i32, b: ~u32, c: u32<32, 64; #attr1>) -> u32 #test_attr {
-        decl a : u32 = let (k: i32 = a + deref(b)) in (k * c[1, 2] + 3 << 32) + 3 + 4 * (32 % 3 / 6 ^ 4 ** 34);
-        for i in @spatial(shape = [1, 2], ranks = [2, 3]) {
-            a = a + 1;
-            a = a ^ threadIdx.x;
-        }
-    }
-}
