+++ python/hidet/ir/tools/__init__.py
@@ -16,3 +16,4 @@ from .free_var_collector import collect_free_vars
 from .printer import IRPrinter, astext
 from .simplifier import simplify, simplify_to_int
 from .hasher import ExprHash
+# from .ir_dumper import astext2, parse
